% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{adjustPvaluesSidak}
\alias{adjustPvaluesSidak}
\title{Multiple testing}
\usage{
adjustPvaluesSidak(pvalues)
}
\arguments{
\item{pvalues}{vector of p values (can contain NA's)}
}
\value{
numeric
}
\description{
Return the p values adjusted according to the Sidak correction.
}
\author{
Timothee Flutre
}
