% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{adjustThreshGalwey}
\alias{adjustThreshGalwey}
\title{Multiple testing}
\usage{
adjustThreshGalwey(effNbIndepTests, thresh = 0.05)
}
\arguments{
\item{effNbIndepTests}{effective number of independent tests}

\item{thresh}{nominal significance threshold on the familywise error rate (FWER)}
}
\value{
numeric
}
\description{
Return the significance threshold adjusted according to the \href{https://dx.doi.org/10.1002/gepi.20408}{Galwey correction (2009)}.
}
\seealso{
\code{\link{effNbIndepTests}}
}
\author{
Timothee Flutre
}
