% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{aireml}
\alias{aireml}
\title{Estimate variance components via REML using AI}
\usage{
aireml(A, y, X, Z, initE, initU, verbose = 0)
}
\arguments{
\item{A}{dditive relationship matrix}

\item{y}{vector of responses}

\item{X}{incidence matrix for fixed effects}

\item{Z}{incidence matrix for random effects}

\item{initE}{initial value for the residual variance}

\item{initU}{initial value for the additive genetic variance}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
vector
}
\description{
Given the MME components, estimate variance components by the AI algorithm.
References are:
\itemize{
\item Tsuruta, S. 2006. Estimation of Variance Components in Animal Breeding. The University of Georgia.
\item Mrode, R.A. 2005. Linear Models for the Prediction of Animal Breeding Values. CAB International, Oxon, UK.
}
}
\note{
Last modified by Morota on April 8, 2010.
}
\author{
Goto Morota
}
