% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{allPairAligns}
\alias{allPairAligns}
\title{All pairwise alignments}
\usage{
allPairAligns(x, type = "global", ...)
}
\arguments{
\item{x}{vector of sequences (e.g. "AGGT"), possibly with names}

\item{type}{type of alignment (default="global", i.e. Needleman-Wunsch)}

\item{...}{arguments to be passed to Biostrings::pairwiseAlignment()}
}
\value{
list of instances of class PairwiseAlignments
}
\description{
Align each sequence against each other (and itself)
}
\details{
Requires the Biostrings package.
}
\author{
Timothee Flutre
}

