% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{barplotGeneticMap}
\alias{barplotGeneticMap}
\title{Stacked barplot of markers}
\usage{
barplotGeneticMap(counts, las = 1, border = NA,
  col = grDevices::rainbow(nrow(counts)), leg.txt = rownames(counts),
  args.leg = list(x = "topright", bty = "n", border = NA, fill =
  rev(grDevices::rainbow(nrow(counts)))), ...)
}
\arguments{
\item{counts}{matrix with the number of markers, with chromosomes in rows and linkage groups in columns, or vice-versa; should have row and column names}

\item{las}{see \code{\link[graphics]{par}}}

\item{border}{see \code{\link[graphics]{barplot}}}

\item{col}{see \code{\link[graphics]{barplot}}}

\item{leg.txt}{see \code{\link[graphics]{barplot}}; if NULL, the stack labels will appear in the middle of each bar}

\item{args.leg}{see \code{\link[graphics]{barplot}}}

\item{...}{arguments to be passed to \code{\link[graphics]{barplot}}}
}
\value{
see \code{\link[graphics]{barplot}}
}
\description{
Make a barplot of markers per linkage group (or chromosome), stacked per chromosome (or linkage group).
}
\author{
Timothee Flutre
}
