% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{barplotReadCounts}
\alias{barplotReadCounts}
\title{Stacked barplot of read counts}
\usage{
barplotReadCounts(counts, xlab = paste0("Individuals (", ncol(counts), ")"),
  ylab = "Number of read pairs", main = "Read pairs per individual",
  lines.h = c(10^5, 10^6, 5 * 10^6), mar = c(4, 4, 4, 1), xlab.line = 3)
}
\arguments{
\item{counts}{matrix}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{main}{an overall title for the plot}

\item{lines.h}{the y-value(s) for horizontal line(s)}

\item{mar}{see \code{par}}

\item{xlab.line}{see \code{line} from \code{mtext}}
}
\value{
nothing
}
\description{
Make a stacked barplot from the output of \code{\link{formatReadCountsPerLane}}.
}
\author{
Timothee Flutre
}

