% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{betterSummary}
\alias{betterSummary}
\title{Summary}
\usage{
betterSummary(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
vector
}
\description{
Format the output of \code{\link{summary}} in a single line, as well as the total number of observations and the number of missing data.
}
\author{
Timothee Flutre
}

