% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{binaryClassif}
\alias{binaryClassif}
\title{Hypothesis testing}
\usage{
binaryClassif(known.nulls, called.nulls)
}
\arguments{
\item{known.nulls}{vector of booleans (TRUE if the null is true)}

\item{called.nulls}{vector of booleans (TRUE if the null is accepted); should be in the same order as the othr vector!}
}
\value{
vector with names
}
\description{
Return the number of true positives, false positives, true negatives,
false negatives, true positive proportion (sensitivity), false positive
proportion, accuracy, true negative proportion (specificity), false
discovery proportion, false negative proportion and positive predictive
value (precision).
More details on \href{http://en.wikipedia.org/wiki/Sensitivity_and_specificity}{Wikipedia}.
}
\author{
Timothee Flutre
}
