% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{blast2granges}
\alias{blast2granges}
\title{BLAST}
\usage{
blast2granges(coords)
}
\arguments{
\item{coords}{data.frame with 14 columns (see \code{\link{loadBlast}})}
}
\value{
GRanges
}
\description{
Convert a data.frame containing alignments coordinates from BLAST into a GRanges object.
}
\examples{
\dontrun{## BLAST should be run beforehand, see the example of `loadBlast`:
coords <- loadBlast("megablast_chroms_seqs.txt.gz", asGRanges=FALSE)
coords.gr <- blast2granges(coords)
}
}
\seealso{
\code{\link{loadBlast}}
}
\author{
Timothee Flutre
}
