% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{boxplotCandidateQtl}
\alias{boxplotCandidateQtl}
\title{Boxplot of QTL}
\usage{
boxplotCandidateQtl(y, X, snp, xlab = "Genotype", ylab = "Phenotype",
  verbose = 1, ...)
}
\arguments{
\item{y}{vector of phenotypes with individual names}

\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{snp}{character with SNP name corresponding to the candidate QTL to plot}

\item{xlab}{label ox the x-axis}

\item{ylab}{label of the y-axis}

\item{verbose}{verbosity level (0/1)}

\item{...}{other arguments to \code{\link[graphics]{boxplot}}}
}
\value{
invisible list with \code{y} and \code{x} used to make the boxplot (same order, with no missing data)
}
\description{
Make a boxplot of a candidate QTL.
}
\author{
Timothee Flutre
}

