% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcAvgPwDiffBtwHaplos}
\alias{calcAvgPwDiffBtwHaplos}
\title{Pi}
\usage{
calcAvgPwDiffBtwHaplos(haplos.chr)
}
\arguments{
\item{haplos.chr}{matrix with haplotypes in rows and sites in columns}
}
\value{
numeric(1)
}
\description{
Calculate the average number of differences between pairs of sequences, named pi in the population genetics literature (Tajima, 1983; Wakeley, 2008).
}
\author{
Timothee Flutre
}

