% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcFreqMissGenos}
\alias{calcFreqMissGenos}
\title{Missing genotypes}
\usage{
calcFreqMissGenos(X)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}
}
\value{
vector
}
\description{
Calculate the frequency of missing genotypes for each SNP.
}
\author{
Timothee Flutre
}

