% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcFreqMissSnpGenosPerGeno}
\alias{calcFreqMissSnpGenosPerGeno}
\title{Missing genotypes}
\usage{
calcFreqMissSnpGenosPerGeno(X)
}
\arguments{
\item{X}{matrix of marker genotypes, with genotypes in rows and markers in columns; missing values should be encoded as NA}
}
\value{
vector
}
\description{
Calculate the frequency of missing marker genotypes for each genotype.
}
\author{
Timothee Flutre
}
