% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{calcFreqNaVcf}
\alias{calcFreqNaVcf}
\title{Missing genotypes in VCF}
\usage{
calcFreqNaVcf(vcf, with.coords = FALSE)
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg \href{http://bioconductor.org/packages/VariantAnnotation/}{VariantAnnotation})}

\item{with.coords}{if TRUE, a GRanges object is returned}
}
\value{
numeric vector or GRanges
}
\description{
Calculate the frequency of missing genotypes for each marker.
}
\author{
Timothee Flutre
}
