% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcL10ApproximateBayesFactorWen}
\alias{calcL10ApproximateBayesFactorWen}
\title{Approximate Bayes factor}
\usage{
calcL10ApproximateBayesFactorWen(
  Y,
  Xg,
  Xc,
  Wg = NULL,
  phi = NULL,
  oma = NULL,
  alpha = 0,
  H = 0,
  nu = 0,
  ES = TRUE
)
}
\arguments{
\item{Y}{n x r phenotype matrix}

\item{Xg}{n x p genotype matrix}

\item{Xc}{n x q design matrix of covariates}

\item{Wg}{r x r prior var-covar matrix on beta_g (if NULL, requires phi and oma)}

\item{phi}{used to make Wg}

\item{oma}{used to make Wg}

\item{alpha}{mixing proportion}

\item{H}{r x r matrix parameter of Wishart prior}

\item{nu}{positive scalar parameter of Wishart prior}

\item{ES}{if TRUE, use the scale-invariant prior formulation}
}
\value{
numeric
}
\description{
Calculate the log10(ABF) under the SSMR model (equation 11 with s=1) of \href{http://dx.doi.org/10.1111/biom.12112}{Wen (Biometrics, 2013)}.
}
\seealso{
\code{\link{calcL10ApproximateBayesFactorWenStephens}}
}
\author{
Xiaoquan Wen [aut], Timothee Flutre [ctb,cre]
}
