% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcL10ApproximateBayesFactorWenStephens}
\alias{calcL10ApproximateBayesFactorWenStephens}
\title{Approximate Bayes factor}
\usage{
calcL10ApproximateBayesFactorWenStephens(sstats, phi2, oma2)
}
\arguments{
\item{sstats}{matrix of summary statistics with one row per subgroup and three columns, "bhat", "sebhat" and "t"}

\item{phi2}{prior variance of the \eqn{b_s} given \eqn{\bar{b}}; controls the prior expected degree of heterogeneity among subgroups}

\item{oma2}{prior variance of \eqn{\bar{b}}; controls the prior expected size of the average effect across subgroups}
}
\value{
numeric
}
\description{
Calculate the log10(ABF) of Wen & Stephens in Annals of Applied Statistics (2014, \url{http://dx.doi.org/10.1214/13-AOAS695}) according to the "exchangeable standardized effects" model.
}
\author{
Xiaoquan Wen [aut], Timothee Flutre [ctb,cre]
}

