% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{caretFitBglr}
\alias{caretFitBglr}
\title{Fit with BGLR for caret}
\usage{
caretFitBglr(
  x,
  y,
  wts,
  param,
  lev,
  last,
  weights,
  classProbs,
  ETA,
  saveAt,
  nIter,
  burnIn,
  thin,
  keep.samples = FALSE,
  ...
)
}
\arguments{
\item{x}{current predictors used to fit the model}

\item{y}{current outcome used to fit the model}

\item{wts}{optional instance weights}

\item{param}{current tuning parameter values}

\item{lev}{class levels of the outcome (or NULL in regression)}

\item{last}{logical for whether the current fit is the final fit}

\item{weights}{?}

\item{classProbs}{logical for whether class probabilities should be computed}

\item{ETA}{two-level list used to specify the regression function for \code{BGLR::BGLR}}

\item{saveAt}{string that may include a path and a pre-fix that will be added to the name of the files that are saved as \code{BGLR::BGLR} runs}

\item{nIter}{number of iterations}

\item{burnIn}{number of burn-in}

\item{thin}{thinning}

\item{keep.samples}{logical for whether the samples will be returned (as a \code{coda::mcmc.list})}

\item{...}{arguments passed on to \code{BGLR::BGLR}}
}
\value{
output of \code{BGLR::BGLR} and, optionally, the samples
}
\description{
The argument named ETA is compulsory for \code{BGLR::BGLR}.
Only its first component will be taken into account.
The design matrix of its first component will be set to be \code{x}.
The model of its first component is used in output files.
}
\author{
Timothee Flutre
}
