% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{caretFitQtl}
\alias{caretFitQtl}
\title{Fit with qtl for caret}
\usage{
caretFitQtl(
  x,
  y,
  wts,
  param,
  lev,
  last,
  weights,
  classProbs,
  genmap,
  cross.geno.prob = NULL,
  phase,
  tuneThreshold = TRUE,
  alpha = c(0.05, NA),
  nperm = 10,
  pop.type = "CP",
  QTLmethod = "SIM",
  p2d = "",
  plot = FALSE,
  QTL_position = NULL,
  nb.cores = parallel::detectCores() - 2,
  verbose = 0
)
}
\arguments{
\item{x}{current predictors used to fit the model, i.e genotypic matrix in Joinmap format with genotypes in rows and markers in columns}

\item{y}{current outcome used to fit the model, i.e trait to study in numeric format with genotype names}

\item{wts}{optional instance weights}

\item{param}{data frame with current tuning parameter values}

\item{lev}{class levels of the outcome (or NULL in regression)}

\item{last}{logical for whether the current fit is the final fit}

\item{weights}{?}

\item{classProbs}{logical for whether class probabilities should be computed}

\item{genmap}{genetic map in data frame with columns linkage.group, genetic.distance and locus}

\item{cross.geno.prob}{list containing genotypic class probabilities calculated from calc.genoprob.
If not passed (NULL), prediction is done with allelic effects at the closest marker.}

\item{phase}{vector of character (length nb of markers) in the format \{--\}}

\item{tuneThreshold}{logical, if TRUE (default), the threshold is chosen for minimizing prediction error with a criterion,
else threshold is chosen by permutations for minimizing False Positive error in QTL detection.}

\item{alpha}{numeric vector with first element corresponding to alpha rate error for chosing LOD significance threshold after permutations,
and the 2nd element is a threshold applied after linear regression with selected markers, if NA, no selection is applied.}

\item{nperm}{number of permutations to be done is tuneThreshold=FALSE}

\item{pop.type}{population type in the JoinMap format, only "CP" is handled here}

\item{QTLmethod}{character (SIM/MIM)}

\item{p2d}{path to directory with results of \code{scantwo} (passed to MIMQTL)}

\item{plot}{logical, default is FALSE. If method is MIM, 2 arguments are needed.}

\item{QTL_position}{data frame with columns linkage.group, genetic.distance and locus to plot abline at simulated QTL position}

\item{nb.cores}{number of cores (passed to MIMQTL)}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
output of \code{\link{SIMQTL}} or \code{\link{MIMQTL}}
}
\description{
Fits the simple or multiple interval mapping model for usage with caret.
First, the genotypes in the JoinMap format are saved into a temporary file.
Second, a \code{cross} object is made.
Third, SIMQTL or MIMQTL is run (calc.genoprob, scanone or scantwo).
Fourth, allelic effects are estimated.
}
\author{
Timothee Flutre
}
