% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{caretFitRrblup}
\alias{caretFitRrblup}
\title{Fit with rrBLUP for caret}
\usage{
caretFitRrblup(x, y, wts, param, lev, last, weights, classProbs, ...)
}
\arguments{
\item{x}{current predictors used to fit the model}

\item{y}{current outcome used to fit the model}

\item{wts}{optional instance weights}

\item{param}{current tuning parameter values}

\item{lev}{class levels of the outcome (or NULL in regression)}

\item{last}{logical for whether the current fit is the final fit}

\item{weights}{?}

\item{classProbs}{logical for whether class probabilities should be computed}

\item{...}{arguments passed on to \code{rrBLUP::mixed.solve}}
}
\value{
output of \code{rrBLUP::mixed.solve}
}
\description{
Fit with rrBLUP for caret
}
\author{
Timothee Flutre
}
