% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{caretPredictBglr}
\alias{caretPredictBglr}
\title{Predict with BGLR for caret}
\usage{
caretPredictBglr(modelFit, newdata, submodels = NULL)
}
\arguments{
\item{modelFit}{model produced by \code{\link{caretFitBglr}}}

\item{newdata}{predictor values of the instances being predicted (e.g. out-of-bag samples)}

\item{submodels}{optional list of tuning parameters only used with the "loop" element}
}
\value{
vector
}
\description{
Predict with BGLR for caret
}
\author{
Timothee Flutre
}
