% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{caretPredictQtl}
\alias{caretPredictQtl}
\title{Predict with qtl for caret}
\usage{
caretPredictQtl(modelFit, newdata, submodels = NULL)
}
\arguments{
\item{modelFit}{model produced by \code{\link{caretFitQtl}}; should be a list with a component named \code{"allelic.effects"} which contains a vector in the same order as the columns of newdata and, as values, the estimated allele effects (0 if not significant, non-zero otherwise)}

\item{newdata}{predictor values of the instances being predicted (e.g. out-of-bag samples); should be in the JoinMap format}

\item{submodels}{optional list of tuning parameters only used with the "loop" element}
}
\value{
vector
}
\description{
Predict with qtl for caret
}
\author{
Charlotte Brault [aut], Timothee Flutre [ctb]
}
