% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{clusteredness}
\alias{clusteredness}
\title{Clusteredness}
\usage{
clusteredness(q.ik)
}
\arguments{
\item{q.ik}{matrix of assignment probabilities with individuals in rows and clusters in columns}
}
\value{
numeric
}
\description{
Computes the statistic which "measures the extent to which a randomly chosen individual is inferred to have ancestry in only one cluster (clusteredness = 1), with the other extreme being equal membership in all clusters (clusteredness = 0)", from \href{http://dx.plos.org/10.1371/journal.pgen.0010070}{Rosenberg et al (2005)}.
}
\author{
Timothee Flutre
}
