% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{coancestry2addrel}
\alias{coancestry2addrel}
\title{Return the additive relationship matrix, A, from the output of the coancestry() function in the "related" package.}
\usage{
coancestry2addrel(x, estim.coancestry, estim.inbreeding = NULL,
  debug = FALSE)
}
\arguments{
\item{x}{list returned by coancestry()}

\item{estim.coancestry}{name of the coancestry estimator (e.g. "dyadml")}

\item{estim.inbreeding}{name of the inbreeding estimator (e.g. "LR")}

\item{debug}{boolean (TRUE to check the output matrix is indeed symmetric)}
}
\value{
matrix
}
\description{
The "related" package can be found here: http://frasierlab.wordpress.com/software/. The A matrix is also known as the numerator relationship matrix. It is calculated as explained in chapter 2 from Mrode (2005).
}
\author{
Timothée Flutre
}

