% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{confidenceGenoOneVar}
\alias{confidenceGenoOneVar}
\title{Confidence in one variant's genotypes}
\usage{
confidenceGenoOneVar(x, plot.it = FALSE)
}
\arguments{
\item{x}{CollapsedVCF corresponding to a SNV (see pkg VariantAnnotation)}

\item{plot.it}{plot GQ=f(DP) if TRUE}
}
\value{
invisible data.frame
}
\description{
Provide measure of confidence in the genotypes at a given variant.
}
\author{
Timothee Flutre
}

