% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertFactorColumnsToCharacter}
\alias{convertFactorColumnsToCharacter}
\title{Data.frame}
\usage{
convertFactorColumnsToCharacter(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Convert the factor columns into character.
}
\author{
Timothee Flutre
}

