% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{convertImputedTo012}
\alias{convertImputedTo012}
\title{Convert imputed genotypes to 0/1/2}
\usage{
convertImputedTo012(X)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in [0,2], with genotypes in rows and SNPs in columns; missing values should be encoded as NA}
}
\value{
matrix
}
\description{
Converts imputed genotypes to 0/1/2.
}
\author{
Timothee Flutre
}
