% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cor2cov}
\alias{cor2cov}
\title{Scales a correlation matrix into the corresponding covariance matrix efficiently.}
\usage{
cor2cov(x, sd)
}
\arguments{
\item{x}{correlation matrix}

\item{sd}{standard deviations}
}
\value{
matrix
}
\description{
Use sweep.
See \url{https://en.wikipedia.org/wiki/Covariance_matrix#Correlation_matrix}.
}
\author{
Timothee Flutre
}
