% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{correspondenceJoinMap2qtl}
\alias{correspondenceJoinMap2qtl}
\title{JoinMap/MapQTL to R/qtl}
\usage{
correspondenceJoinMap2qtl()
}
\value{
data frame
}
\description{
Return the correspondence in terms of genotype coding between the format used by \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} and the one used by \href{https://cran.r-project.org/package=qtl}{qtl}.
}
\author{
Timothee Flutre
}
