% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{corrMatAR1}
\alias{corrMatAR1}
\title{AR(1)}
\usage{
corrMatAR1(n, rho)
}
\arguments{
\item{n}{dimension of the matrix (number of rows and columns)}

\item{rho}{correlation between successive variables}
}
\value{
matrix
}
\description{
Return a first-order auto-regressive correlation matrix, as noted in equation 4 of \href{http://dx.doi.org/10.1139/x02-111}{Dutkowski et al (2002)}.
}
\author{
Timothee Flutre
}

