% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{countGenotypicClasses}
\alias{countGenotypicClasses}
\title{Genotypic classes}
\usage{
countGenotypicClasses(X)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; missing values should be encoded as NA}
}
\value{
matrix with 4 columns and as many rows as there are SNPs
}
\description{
Count the genotypic classes per SNP (not imputed, i.e. 0, 1, 2 and NA).
}
\author{
Timothee Flutre
}

