% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{countRequestedBreedTypes}
\alias{countRequestedBreedTypes}
\title{Count types}
\usage{
countRequestedBreedTypes(df)
}
\arguments{
\item{df}{data.frame from \code{\link{readCheckBreedPlantFile}} or \code{\link{readCheckBreedDataFile}}}
}
\value{
named vector
}
\description{
Count the types of breeding requests.
}
\author{
Timothee Flutre
}

