% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{coverage.regions}
\alias{coverage.regions}
\title{Plot the covered fraction of regions as a function of depth}
\usage{
coverage.regions(path = NULL, pattern = NULL, covrg = NULL,
  plot.it = TRUE, xlim = NULL, ylim = c(0, 1), points.type = "l",
  plot.legend = TRUE, verbose = 1)
}
\arguments{
\item{path}{string}

\item{pattern}{character string containing a regular expression passed to gsub to be removed to simplify the file basenames used in the legend (e.g. "bedtools-coverage|\\.txt\\.gz")}

\item{covrg}{list of data frames}

\item{plot.it}{boolean (default=TRUE)}

\item{xlim}{vector of limits for the x-axis}

\item{ylim}{vector of limits for the y-axis}

\item{points.type}{character indicating the type of plotting ("p"/default="l"/"b")}

\item{plot.legend}{boolean (default=TRUE)}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
invisible list of covrg and cumcovrg
}
\description{
Need to first run bedtools coverage as in http://www.gettinggeneticsdone.com/2014/03/visualize-coverage-exome-targeted-ngs-bedtools.html by Stephen Turner
}
\author{
Timothee Flutre
}

