% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{coverageBams}
\alias{coverageBams}
\title{Coverage}
\usage{
coverageBams(bamFiles, seq.ids = NULL, min.mapq = 5, verbose = 1,
  nb.cores = 1)
}
\arguments{
\item{bamFiles}{vector of paths to BAM file(s), each of them having an index file with the same name but finishing by ".bai"; if there are several BAM files, all of them should contain reads aligned on the same set of sequences (i.e. the same reference genome)}

\item{seq.ids}{sequence identifier(s) to focus on, e.g. "chr2" or c("chr2","chr5"); if NULL, all of them}

\item{min.mapq}{only count reads with mapping quality greater than the given value}

\item{verbose}{verbosity level (0/1/2)}

\item{nb.cores}{the number of cores to use to parallelize over bamFiles, i.e. at most how many child processes will be run simultaneously (not on Windows)}
}
\value{
list with one component per BAM file, each being a matrix with columns "nb.bases", "mapped.bases", "count.sum", "breadth", "depth"
}
\description{
Calculate the depth (average number of times each position in the genome has been sequenced) and breadth (fraction of the genome that is covered by reads) of coverage from a set of BAM file(s), using \code{samtools depth} (see \href{http://dx.doi.org/10.1093/bib/bbu029}{Molnar & Ilie (2015)}).
}
\author{
Timothee Flutre (with the help of Martin Morgan)
}
\seealso{
\code{\link{freadBedtoolsCoverageHist}}
}

