% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{covMatAR1}
\alias{covMatAR1}
\title{AR(1)}
\usage{
covMatAR1(n, rho, sigma2)
}
\arguments{
\item{n}{dimension of the matrix (number of rows and columns)}

\item{rho}{correlation between successive variables}

\item{sigma2}{variance of the errors (also called "innovations" in the time-series literature)}
}
\value{
matrix
}
\description{
Return a first-order auto-regressive covariance matrix.
}
\author{
Timothee Flutre
}
