% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{createA}
\alias{createA}
\title{Compute the inverse of the additive relationship matrix using Henderson algorithm}
\usage{
createA(s, d)
}
\arguments{
\item{s}{vector of sires}

\item{d}{vector of dams}
}
\value{
matrix
}
\description{
Given the vectors of sire and dam, directly return the additive relationship matrix 'A' (tabular method).
Reference:
\itemize{
\item Henderson, C. R. 1976. Simple Method for Computing the Inverse of a Numerator Relationship Matrix Used in Prediction of Breeding Values. Biometrics 32:69-83.
}
}
\note{
Unknown parents should be coded as zero. Last modified by Morota on April 1, 2010.
}
\author{
Gota Morota
}
