% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{defLinkgroupsWithCarthagene}
\alias{defLinkgroupsWithCarthagene}
\title{Define linkage groups with CarthaGene}
\usage{
defLinkgroupsWithCarthagene(file = NULL, task.id = NULL, cg = NULL,
  dist.thresh = 0.3, lod.thresh = 3, mrk2chr = NULL, close.cg = FALSE,
  verbose = 1)
}
\arguments{
\item{file}{path to the file containing the dataset (used if cg is NULL)}

\item{task.id}{identifier of the task (used if cg is NULL)}

\item{cg}{list returned by \code{\link{openCarthagene}} (used if file and task.id are NULL)}

\item{dist.thresh}{distance threshold}

\item{lod.thresh}{LOD threshold}

\item{mrk2chr}{named vector which values are markers and names are chromosomes}

\item{close.cg}{if TRUE, the connection to CarthaGene will be closed and the temporary file removed}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame
}
\description{
Defines linkage groups with CarthaGene ("group" command), so that two markers whose 2-point Haldane distance is below the given distance threshold, and 2-point LOD is above the LOD threshold, will be put in the same linkage group.
Choose these thresholds by trial and error until it makes sense given the ploidy of the species of interest.
}
\seealso{
\code{\link{openCarthagene}}, \code{\link{runCarthagene}}, \code{\link{parseCgMrkinfo}}, \code{\link{parseCgGroup}}, \code{\link{closeCarthagene}}
}
\author{
Timothee Flutre
}
