% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{depthsPerSample}
\alias{depthsPerSample}
\title{Depth per sample across regions}
\usage{
depthsPerSample(dat, min.reg.len = 30, max.reg.len = 500,
  min.reg.dep = 10, max.reg.dep = 200, min.reg.frac = 0.25)
}
\arguments{
\item{dat}{data.table (see freadBedtoolsCoverageHist)}

\item{min.reg.len}{minimum length of a region to be considered}

\item{max.reg.len}{maximum length of a region to be considered}

\item{min.reg.dep}{minimum depth of a region when reporting the number of interesting regions}

\item{max.reg.dep}{maximum depth of a region when reporting the number of interesting regions}

\item{min.reg.frac}{minimum fraction of a region when reporting the number of interesting regions}
}
\value{
data.table
}
\description{
Summarize the depth per sample across regions of a given length range.
}
\author{
Timothee Flutre
}

