% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{desc.plate}
\alias{desc.plate}
\title{Describe plate}
\usage{
desc.plate(plate, plate.name, verbose = 1)
}
\arguments{
\item{plate}{matrix}

\item{plate.name}{string}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
invisible named vector
}
\description{
Describe the given plate.
}
\author{
Timothée Flutre
}

