% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{descPlate}
\alias{descPlate}
\title{Describe plate}
\usage{
descPlate(plate, plate.name, verbose = 1)
}
\arguments{
\item{plate}{matrix}

\item{plate.name}{string}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
invisible named vector
}
\description{
Describe the given plate.
}
\author{
Timothee Flutre
}

