% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{dimVcf}
\alias{dimVcf}
\title{VCF dimensions}
\usage{
dimVcf(vcf.file, genome = "", yieldSize = 10000)
}
\arguments{
\item{vcf.file}{path to the VCF file (if the bgzip index doesn't exist in the same directory, it will be created)}

\item{genome}{genome identifier (e.g. "VITVI_12x2")}

\item{yieldSize}{number of records to yield each time the file is read from}
}
\value{
vector
}
\description{
Return the dimensions of a VCF file, i.e. the number of sites (rows) and samples (columns)
}
\author{
Timothee Flutre
}
