% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{discardMarkersMissGenos}
\alias{discardMarkersMissGenos}
\title{Missing genotypes}
\usage{
discardMarkersMissGenos(X, verbose = 1)
}
\arguments{
\item{X}{matrix of markers' genotypes numerically encoded (if SNPs, in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}), with individuals in rows and markers in columns; NA if missing}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix similar to X but possibly with less columns
}
\description{
Discard all markers with at least one missing genotype.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{imputeGenosWithMean}}
}

