% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{discardSnpsMissGenos}
\alias{discardSnpsMissGenos}
\title{Missing genotypes}
\usage{
discardSnpsMissGenos(X, verbose = 1)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns; NA if missing}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix similar to X but possibly with less columns
}
\description{
Discard the SNPs with any missing genotype.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{imputeGenosWithMean}}
}

