% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{distConsecutiveSnps}
\alias{distConsecutiveSnps}
\title{Distance between consecutive SNPs}
\usage{
distConsecutiveSnps(snp.coords, only.chr = NULL, nb.cores = 1)
}
\arguments{
\item{snp.coords}{data.frame with SNP identifiers as row names, and with two columns "chr" and "coord" or "pos"}

\item{only.chr}{identifier of a given chromosome}

\item{nb.cores}{the number of cores to use}
}
\value{
list with one component per chromosome
}
\description{
For each pair of consecutive SNPs, return the number of "blocks" (i.e. nucleotides) between both SNPs (to be coherent with \code{\link{distSnpPairs}}).
}
\examples{
\dontrun{## make fake data
snp.coords <- data.frame(chr=c("chr1","chr1","chr1","chr2"),
                         pos=c(150, 131, 171, 17))
rownames(snp.coords) <- paste0("snp", 1:nrow(snp.coords))

distConsecutiveSnps(snp.coords)
}
}
\seealso{
\code{\link{estimLd}}
}
\author{
Timothee Flutre
}
