% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{distSnpPairs}
\alias{distSnpPairs}
\title{Distance between SNP pairs}
\usage{
distSnpPairs(snp.pairs, snp.coords, nb.cores = 1, verbose = 0)
}
\arguments{
\item{snp.pairs}{data.frame with two columns "loc1" and "loc2"}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"}

\item{nb.cores}{the number of cores to use}

\item{verbose}{verbosity level (default=0/1)}
}
\value{
vector
}
\description{
For each SNP pair, return the number of "blocks" (i.e. nucleotides) between both SNPs via the \code{\link[GenomicRanges]{distance}} function.
}
\author{
Timothee Flutre
}

