% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{doesBreederExist}
\alias{doesBreederExist}
\title{Check}
\usage{
doesBreederExist(breeder, root.dir, lang = "fr")
}
\arguments{
\item{breeder}{name of the breeder (e.g. "breeder3" or "selectionneur2", depending on the language)}

\item{root.dir}{path to the root directory}

\item{lang}{language to be used (default=fr/en)}
}
\value{
logical
}
\description{
Check if the given breeder is part of the game.
}
\author{
Timothee Flutre
}

