% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{dose2alleles}
\alias{dose2alleles}
\title{Convert genotypes}
\usage{
dose2alleles(X = NULL, tX = NULL, alleles, na.string)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{tX}{matrix with SNPs in rows and individuals in columns}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (first is "minor", second is "major")}

\item{na.string}{a character used to replace NA values (for instance "??")}
}
\value{
data.frame with SNPs in rows and individuals in columns
}
\description{
Convert SNP genotypes from minor allele doses (say, 0 and 1) into alleles (say, "AA" and "AT" with "T" being the minor allele). Not particularly efficient, but at least it exists.
}
\author{
Timothee Flutre
}

