% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{dose2bimbam}
\alias{dose2bimbam}
\title{Convert genotypes}
\usage{
dose2bimbam(X = NULL, tX = NULL, alleles, file = NULL, format = "mean")
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{tX}{matrix with SNPs in rows and individuals in columns}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (named "minor" and "major")}

\item{file}{write the genotype data to this file if non NULL (for instance 'genotypes_bimbam.txt' or 'genotypes_bimbam.txt.gz', but don't use \code{\link[base]{gzfile}})}

\item{format}{BimBam's format in which the data should be saved (mean/basic)}
}
\value{
invisible data.frame
}
\description{
Convert SNP genotypes to the file formats used by BimBam, as specified in its \href{http://www.haplotype.org/download/bimbam-manual.pdf}{manual}.
}
\author{
Timothee Flutre
}

