% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{dsVcf2dose}
\alias{dsVcf2dose}
\title{Convert DS to dosage}
\usage{
dsVcf2dose(vcf)
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg \href{http://bioconductor.org/packages/VariantAnnotation/}{VariantAnnotation})}
}
\value{
matrix (in numeric \code{\link{mode}}) with variants in rows and samples in columns
}
\description{
Non-bi-allelic variants are discarded.
From Martin Morgan (see http://grokbase.com/t/r/bioconductor/135b460s2b/bioc-how-to-convert-genotype-snp-matrix-to-nucleotide-genotypes).
}
\seealso{
\code{\link{vcf2dosage}}, \code{\link{gtVcf2dose}}
}
\author{
Timothee Flutre
}
