% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{effNbIndepTests}
\alias{effNbIndepTests}
\title{Multiple testing}
\usage{
effNbIndepTests(evCorX, corX, X, method = "galwey")
}
\arguments{
\item{evCorX}{eigenvalues of the correlation matrix between SNP genotypes}

\item{corX}{correlation matrix between SNP genotypes}

\item{X}{SNP genotypes in allele dose (i.e., 0, 1, 2) with individuals in rows and SNPs in columns}

\item{method}{for now, only "galwey" is implemented}
}
\value{
numeric
}
\description{
Return the effective number of independent tests according to the method of \href{https://dx.doi.org/10.1002/gepi.20408}{Galwey (2009)}.
}
\seealso{
\code{\link{adjustThreshGalwey}}
}
\author{
Timothee Flutre
}
