% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{effNbIndepTestsPerChr}
\alias{effNbIndepTestsPerChr}
\title{Multiple testing}
\usage{
effNbIndepTestsPerChr(lcorX, method = "galwey", pkg = "base", k, nb.cores = 1)
}
\arguments{
\item{lcorX}{list of correlation matrices between SNP genotypes, one per chromosome (see \code{estimLdPerChr})}

\item{method}{for now, only "galwey" is implemented}

\item{pkg}{package to compute the eigenvalues of \code{corX} if \code{evCorX} is not supplied (base/RSpectra)}

\item{k}{number of eigenvalues requested (used by the RSpectra package)}

\item{nb.cores}{number of cores to to run each chromosome in parallel}
}
\value{
numeric
}
\description{
Return the effective number of independent tests according to the method of \href{https://dx.doi.org/10.1002/gepi.20408}{Galwey (2009)} for a given chromosome after estimating the LD along it.
}
\seealso{
\code{\link{effNbIndepTests}}, \code{\link{adjustThreshGalwey}}
}
\author{
Maxence Remerand, Timothee Flutre
}
