% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{empty.wells}
\alias{empty.wells}
\title{Identify empty wells, if any, in a plate.}
\usage{
empty.wells(plate.w)
}
\arguments{
\item{plate.w}{matrix of a plate in the "wide" format}
}
\value{
2 column data.frame (row;col) corresponding to empty wells
}
\description{
Identify empty wells, if any, in a plate.
}
\author{
Timothee Flutre
}

