% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{emreml}
\alias{emreml}
\title{Estimate variance components via REML using EM}
\usage{
emreml(Ainv, y, X, Z, initE, initU, verbose = 0)
}
\arguments{
\item{Ainv}{inverse of additive relationship matrix}

\item{y}{vector of responses}

\item{X}{incidence matrix for fixed effects}

\item{Z}{incidence matrix for random effects}

\item{initE}{initial value for the residual variance}

\item{initU}{initial value for the additive genetic variance}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
vector
}
\description{
Given the MME components, estimate variance components by the EM algorithm.
References are:
\itemize{
\item Suzuki, M. 2007. Applied Animal Breeding & Genetics. Course Notes. Obihiro University of Agriculture and Veterinary Medicine.
\item Mrode, R.A. 2005. Linear Models for the Prediction of Animal Breeding Values. CAB International, Oxon, UK.
}
}
\note{
Unknown parents should be coded as zero. Last modified by Morota on April 8, 2010.
}
\seealso{
\code{link{makeMmeElements}}, \code{\link{makeMmeRhs}}, \code{\link{makeMmeLhs}}
}
\author{
Goto Morota [aut], Timothee Flutre [ctb]
}
