% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estim.genrel}
\alias{estim.genrel}
\title{Estimate genetic relationships between individuals from their SNP genotypes.}
\usage{
estim.genrel(X, mafs = NULL, thresh = 0.01, relationships = "additive",
  method = "astle-balding", verbose = 1)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses ({0,1,2}), with SNPs in
columns and individuals in rows}

\item{mafs}{vector with minor allele frequencies (calculated with `maf.from.dose` if NULL)}

\item{thresh}{threshold on allele frequencies below which SNPs are ignored (default=0.01, NULL to skip this step)}

\item{relationships}{genetic relationship to estimate (default=additive/dominance)}

\item{method}{if additive relationships, can be "astle-balding" (default), "animal-model", "center" or "center-std"}

\item{verbose}{verbosity level (default=1)}
}
\value{
matrix
}
\description{
SNPs with missing data are ignored.
}
\author{
Timothee Flutre
}

