% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estim.kinship}
\alias{estim.kinship}
\title{Estimate kinship matrix from SNPs.}
\usage{
estim.kinship(X, mafs = NULL, thresh = 0.01, method = "astle-balding",
  verbose = 1)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses, with SNPs in
columns and individuals in rows}

\item{mafs}{vector with minor allele frequencies (calculated with `maf.from.dose` if NULL)}

\item{thresh}{threshold on allele frequencies below which SNPs are ignored (default=0.01, NULL to skip this step)}

\item{method}{default is "astle-balding"; "animal-model"; "center", "center-std"}

\item{verbose}{verbosity level (default=1)}
}
\value{
matrix
}
\description{
SNPs with missing data are ignored.
}
\author{
Timothée Flutre
}

