% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{estimatePi0WithEbf}
\alias{estimatePi0WithEbf}
\title{FDR}
\usage{
estimatePi0WithEbf(log10.bfs, verbose = 1)
}
\arguments{
\item{log10.bfs}{vector containing the log10(BF) of each test (NA will be discarded)}

\item{verbose}{verbosity level (0/1)}
}
\value{
numeric
}
\description{
Estimate pi0 (proba for a null hypothesis to be true) via the EBF procedure (Wen, arXiv:1311.3981).
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{estimatePi0WithQbf}}, \code{\link{controlBayesFdr}}
}

