% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{estimatePi0WithQbf}
\alias{estimatePi0WithQbf}
\title{FDR}
\usage{
estimatePi0WithQbf(log10.bfs, gamma = 0.5, verbose = 1)
}
\arguments{
\item{log10.bfs}{matrix with tests in rows and two columns, the true log10(BF) and the gamma-quantile log10(BF) under the null}

\item{gamma}{level of the quantile (e.g. 0.5 for the median)}

\item{verbose}{verbosity level (0/1)}
}
\value{
numeric
}
\description{
Estimate pi0 (proba for a null hypothesis to be true) via the QBF procedure (Wen, arXiv:1311.3981).
}
\seealso{
\code{\link{estimatePi0WithEbf}}, \code{\link{controlBayesFdr}}
}
\author{
Timothee Flutre
}
