% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimH2means}
\alias{estimH2means}
\title{Broad-sense heritability}
\usage{
estimH2means(dat, colname.resp, vc)
}
\arguments{
\item{dat}{data frame of input data after missing data have been excluded, with columns named \code{colname.resp}, "geno" and "year"}

\item{colname.resp}{name of the column containing the response}

\item{vc}{data frame of variance components with columns "grp" and "vcov" (i.e. formatted as \code{as.data.frame(VarCorr())} from the "lme4" package)}
}
\value{
list with the mean number of years, the mean number of replicates per year, the broad-sense heritability and a function to compute summary statistics whch can be used for estimating confidence intervals by bootstrap
}
\description{
Estimate broad-sense heritability of phenotypic means via the classical formula for balanced data sets as in the introduction of \href{http://www.genetics.org/cgi/doi/10.1534/genetics.107.074229}{Piepho and Mohring (2007)}.
For unbalanced data sets, the mean number of years and replicates per year are used.
}
\author{
Timothee Flutre
}
