% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimLd}
\alias{estimLd}
\title{Pairwise linkage disequilibrium}
\usage{
estimLd(X, K = NULL, pops = NULL, snp.coords, only.chr = NULL,
  only.pop = NULL, use.ldcorsv = FALSE, verbose = 1)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; missing values should be encoded as NA}

\item{K}{matrix of "kinship" (additive genetic relationships)}

\item{pops}{vector of characters indicating the population of each genotype}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "pos"}

\item{only.chr}{identifier of a given chromosome}

\item{only.pop}{identifier of a given population}

\item{use.ldcorsv}{required if K and/or pops are not NULL; otherwise use the square of \code{\link{cor}}}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame
}
\description{
Estimates linkage disequilibrium between pairs of SNPs when the observations are the genotypes of genotypes, not their gametes (i.e. the gametic phases are unknown).
When ignoring kinship and population structure, the estimator of Rogers and Huff (Genetics, 2009) can be used.
When kinship and/or population structure are controlled for, the estimator of Mangin et al (Heredity, 2012) is used via their LDcorSV package.
}
\author{
Timothee Flutre
}

