% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimLdPerChr}
\alias{estimLdPerChr}
\title{Pairwise linkage disequilibrium}
\usage{
estimLdPerChr(
  X,
  snp.coords,
  K = NULL,
  pops = NULL,
  only.pop = NULL,
  use.ldcorsv = FALSE,
  use.snpStats = FALSE,
  as.cor = FALSE,
  as.symmat = FALSE,
  nb.cores = 1,
  verbose = 1
)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in \{0,1,2\}, with genotypes in rows and SNPs in columns; missing values should be encoded as NA}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "pos"}

\item{K}{matrix of "kinship" (additive genetic relationships)}

\item{pops}{vector of characters indicating the population of each genotype}

\item{only.pop}{identifier of a given population}

\item{use.ldcorsv}{required if K and/or pops are not NULL; otherwise use the square of \code{\link{cor}}}

\item{use.snpStats}{if TRUE, the \code{ld} function of the snpStats package is used (see \href{https://dx.doi.org/10.1159/000101422}{Clayton and Leung, 2007})}

\item{as.cor}{if TRUE, the square root of the LD estimates is returned}

\item{as.symmat}{if TRUE, LD values are returned as a symmetric matrix (not with \code{use.ldcorsv} set to TRUE)}

\item{nb.cores}{number of cores to estimate LD for each chromosome in parallel}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame with at least three columns, "loc1", "loc2" and the LD values
}
\description{
Estimates linkage disequilibrium between pairs of SNPs belonging to the same chromosome when the observations are the genotypes of genotypes, not their gametes (i.e. the gametic phases are unknown).
When ignoring kinship and population structure, the estimator of Rogers and Huff (Genetics, 2009) can be used.
When kinship and/or population structure are controlled for, the estimator of Mangin et al (Heredity, 2012) is used via their LDcorSV package.
}
\seealso{
\code{\link{estimLd}}, \code{\link{plotLd}}
}
\author{
Timothee Flutre
}
