% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{estMrkOrderGenDistsWithCarthagene}
\alias{estMrkOrderGenDistsWithCarthagene}
\title{Estimate marker order and genetic distances with CarthaGene}
\usage{
estMrkOrderGenDistsWithCarthagene(cg, linkgroups, lg.id, keep.thresh = 3,
  add.thresh = 3, init.order = NULL, postprocessing = 1, flips.size = 4,
  flips.lod.thresh = 3, flips.iter = 1, squeeze.thresh = 20,
  mrk2phy = NULL, verbose = 1)
}
\arguments{
\item{cg}{list returned by \code{\link{openCarthagene}}}

\item{linkgroups}{data frame returned by \code{\link{defLinkgroupsWithCarthagene}}, towhich a column of logicals named "todrop" may have been added, indicating which marker(s) to drop via "mrkdel" before executing "buildfw"}

\item{lg.id}{identifier of the linkage group of interest}

\item{keep.thresh}{the minimum difference in log-likelihood between the best insertion point and the second best insertion point required for the map to be considered in the future}

\item{add.thresh}{the minimum difference in log-likelihood between the best insertion point and the second best insertion point required for a locus to be insertable; this threshold is also used to filter out the differences in log-likelihood reported by the post-processing option (only differences lower than the threshold will be reported)}

\item{init.order}{if not NULL, vector containing at least three marker identifiers in a given order to start from}

\item{postprocessing}{if 0, only framework mapping is performed; if 1, post-processing is also applied (see "add.thresh"); if 2, post-processing is applied only to markers in "mrk.list"}

\item{flips.size}{size of the sliding window for the "flips" command (cannot exceed 9)}

\item{flips.lod.thresh}{maximum difference of log-likelihood with the best map so that "flips" reports the permutation}

\item{flips.iter}{if 1, "flips" will be iterated as long as a new, improved map has been found; specify O, otherwise}

\item{squeeze.thresh}{distance threshold in cMorgan (Haldane distance) for genetics data or cRay for Radiation Hybrids data used by "squeeze" to expunge non-reliable locus from the map}

\item{mrk2phy}{named vector which names are the locus and values are the physical distances on the chromosome corresponding to the linkage group of interest}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame returned by \code{\link{parseCgMaprintd}} containing the best map for the given linkage group according to CarthaGene, and possibly a column "physical.distance"
}
\description{
Estimates marker order and genetic distances for a given linkage group with CarthaGene (commands "buildfw", "flips", polish" and "squeeze").
}
\seealso{
\code{\link{openCarthagene}}, \code{\link{runCarthagene}}, \code{\link{defLinkgroupsWithCarthagene}}, \code{\link{parseCgMaprintd}}, \code{\link{closeCarthagene}}
}
\author{
Timothee Flutre
}
