% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{fastphase}
\alias{fastphase}
\title{Genotype imputation}
\usage{
fastphase(X, snp.coords, alleles, out.dir = getwd(), task.id = "fastphase",
  nb.starts = 20, nb.iters = 25, nb.samp.haplos = 50,
  estim.haplos = FALSE, nb.clusters = 10, seed = 1859, clean = FALSE,
  verbose = 1)
}
\arguments{
\item{X}{NxP matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (first is "minor", second is "major")}

\item{out.dir}{directory in which the output files will be saved}

\item{task.id}{identifier of the task (used in temporary and output file names)}

\item{nb.starts}{number of random starts of the EM algorithm}

\item{nb.iters}{number of iterations of the EM algorithm}

\item{nb.samp.haplos}{number of haplotypes sampled from the posterior distribution obtained from a particular random start of the EM algorithm}

\item{estim.haplos}{estimate haplotypes by minimizing individual error}

\item{nb.clusters}{number of clusters}

\item{seed}{seed for random number generation}

\item{clean}{remove files}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with matrix of genotypes before imputation, and matrix of haplotypes (2 per individual, in rows) after imputation
}
\description{
Impute SNP genotypes via fastPHASE (Scheet and Stephens, 2006).
}
\author{
Timothee Flutre
}

