% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{fecundation}
\alias{fecundation}
\title{Fecundation}
\usage{
fecundation(gam1, gam2, child.name)
}
\arguments{
\item{gam1}{list of 1-row matrices (one per chromosome)}

\item{gam2}{list of 1-row matrices (one per chromosome)}

\item{child.name}{identifier of the child}
}
\value{
list of 2-row matrices (one per chromosome)
}
\description{
Perform fecundation by uniting the two gametes.
}
\author{
Timothee Flutre
}

