% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{filterSegreg}
\alias{filterSegreg}
\title{Filter for segregation}
\usage{
filterSegreg(x, verbose = 1)
}
\arguments{
\item{x}{data.frame output from \code{\link{genoClasses2JoinMap}}}

\item{verbose}{verbosity level (0/1)}
}
\value{
data.frame
}
\description{
Filter genotypes based on the chi2 statistic to test for segregation distortion.
}
\author{
Timothee Flutre
}
\seealso{
\code{link{updateJoinMap}}
}

