% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{filterSnpCalls}
\alias{filterSnpCalls}
\title{Filter SNP calls}
\usage{
filterSnpCalls(vcf.file, dict.file, genome, seq, min.dp, max.dp, min.af = 0.1,
  max.af = 0.9, verbose = 1)
}
\arguments{
\item{vcf.file}{path to the VCF file (bgzip index should exist in same directory; should only contain SNPs and be already filtered for QD, FS, MQ, etc)}

\item{dict.file}{path to the SAM dict file (see \url{https://broadinstitute.github.io/picard/command-line-overview.html#CreateSequenceDictionary})}

\item{genome}{genome identifier (e.g. "VITVI_12x2")}

\item{seq}{sequence identifier to work on (e.g. "chr2")}

\item{min.dp}{minimum depth ("INFO/DP" field in VCF)}

\item{max.dp}{maximum depth}

\item{min.af}{minimum alternate allele frequency ("INFO/AF" field)}

\item{max.af}{maximum alternate allele frequency}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
CollapsedVCF from the VariantAnnotation package
}
\description{
Return subset of calls from VCF file corresponding to bi-allelic SNPs with proper depth and allele frequency.
}
\author{
Timothee Flutre
}

