% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{fitPhyDistVsLd}
\alias{fitPhyDistVsLd}
\title{Pairwise linkage disequilibrium}
\usage{
fitPhyDistVsLd(x, y, method = "loess", span = 0.2, newdata)
}
\arguments{
\item{x}{numeric vector of predictors (LD values or physical distances)}

\item{y}{numeric vector of responses (LD values or physical distances)}

\item{method}{loess/lm/loglm/asyreg/biexp}

\item{span}{the parameter alpha which controls the degree of loess smoothing}

\item{newdata}{an optional data frame in which to look for variables with which to predict}
}
\value{
list with data, model fit and prediction(s)
}
\description{
Fit a model of physical distance versus LD (or reciprocally).
}
\seealso{
\code{\link{summarizeLd}}
}
\author{
Timothee Flutre
}
