% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{formatReadCountsPerLane}
\alias{formatReadCountsPerLane}
\title{Reformat read counts per lane}
\usage{
formatReadCountsPerLane(x)
}
\arguments{
\item{x}{data frame with columns "lane", "ind" and "assigned"}
}
\value{
matrix with lanes in rows and individuals in columns
}
\description{
Reformat as a matrix of counts a data frame formatted, for instance, as the output of \code{\link{loadReadCountsPerIndAndLane}}
}
\author{
Timothee Flutre
}
