% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{fread.bedtools.coverage.hist}
\alias{fread.bedtools.coverage.hist}
\title{Read bedtools-coverage-hist as data.table}
\usage{
fread.bedtools.coverage.hist(files, verbose = 1)
}
\arguments{
\item{files}{character vector of relative or absolute filepaths (e.g. from Sys.glob).}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
list of data.tables
}
\description{
Read output files from bedtools coverage with -hist into a list of data.tables. All lines starting by "all" must have been discarded beforehand.
}
\author{
Timothee Flutre
}

