% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{freadBedtoolsCoverageHist}
\alias{freadBedtoolsCoverageHist}
\title{Read bedtools-coverage-hist as data.table}
\usage{
freadBedtoolsCoverageHist(files, verbose = 1)
}
\arguments{
\item{files}{character vector of relative or absolute filepaths (e.g. from Sys.glob).}

\item{verbose}{verbosity level (0/1)}
}
\value{
list of data.tables
}
\description{
Read output files from bedtools coverage with -hist into a list of data.tables. All lines starting by "all" must have been discarded beforehand.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{depthsPerSample}}, \code{\link{depthsPerRegion}}
}

