% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{gc.content}
\alias{gc.content}
\title{Calculate the GC content of a set of sequences.}
\usage{
gc.content(x)
}
\arguments{
\item{x}{vector of sequences (e.g. "AGGT"), possibly with names}
}
\value{
vector
}
\description{
Requires the Biostrings package.
}
\author{
Timothee Flutre
}

