% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{gcContent}
\alias{gcContent}
\title{Calculate the GC content of a set of sequences}
\usage{
gcContent(x)
}
\arguments{
\item{x}{vector of sequences (e.g. "AGGT"), possibly with names}
}
\value{
vector
}
\description{
Requires the Biostrings package.
}
\author{
Timothee Flutre
}

