% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEMMA.R
\name{gemmaUlmmPerChr}
\alias{gemmaUlmmPerChr}
\title{GEMMA uLMM per chromosome}
\usage{
gemmaUlmmPerChr(y, X, snp.coords, recode.genos = TRUE, alleles = NULL,
  maf = 0.01, chr.ids = NULL, W, weights = NULL,
  exe = Sys.which("gemma"), out.dir = getwd(), task.id = "gemma",
  clean = "none", verbose = 1)
}
\arguments{
\item{y}{vector of phenotypes with genotype names}

\item{X}{matrix of bi-allelic SNP genotypes encoded, for each SNP, in number of copies of its second allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; the "second" allele is arbitrary, it corresponds to the second column of \code{alleles}, which can be the minor or the major allele}

\item{snp.coords}{data.frame with SNPs as row names and two columns named "coord" and "chr"}

\item{recode.genos}{if TRUE, SNP genotypes in X will be recoded so that the minor allele is counted}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (exactly 2 columns are required); the second column should correspond to the allele which number of copies is counted at each SNP in \code{X}; if NULL, fake alleles will be generated}

\item{maf}{SNPs with minor allele frequency strictly below this threshold will be discarded}

\item{chr.ids}{set of chromosome identifiers to analyze (optional, all by default)}

\item{W}{matrix of covariates with genotypes in rows (names as row names), a first column of 1 and a second column of covariates values}

\item{weights}{vector of positive weights with genotype names}

\item{exe}{path to the executable of GEMMA}

\item{out.dir}{directory in which the data files will be found}

\item{task.id}{identifier of the task (used in output file names)}

\item{clean}{remove files: none, some (temporary only), all (temporary and results)}

\item{verbose}{verbosity level (0/1)}
}
\value{
a data.frame with GEMMA's output for all chromosomes
}
\description{
See Zhou & Stephens (Nature Genetics, 2012).
}
\seealso{
\code{link{gemma}}, \code{\link{plotHistPval}}, \code{\link{qqplotPval}}
}
\author{
Timothee Flutre [aut,cre], Dalel Ahmed [ctb]
}
