% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEMMA.R
\name{genoDoses2bimbam}
\alias{genoDoses2bimbam}
\title{Convert genotypes}
\usage{
genoDoses2bimbam(X = NULL, tX = NULL, alleles, file = NULL,
  format = "mean")
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; for each SNP, the allele which copy number is counted, should correspond to the second column of \code{alleles}, whether this column corresponds to minor or major alleles (for GEMMA, it should be the minor allele)}

\item{tX}{matrix with SNPs in rows and genotypes in columns}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns; the second column should correspond to the allele which number of copies is counted at each SNP in \code{X}}

\item{file}{write the genotype data to this file if not NULL (for instance 'genotypes_bimbam.txt' or 'genotypes_bimbam.txt.gz', but don't use \code{\link[base]{gzfile}})}

\item{format}{BimBam's format in which the data should be saved (mean/basic)}
}
\value{
invisible data.frame
}
\description{
Convert SNP genotypes to the file formats used by BimBam, as specified in its \href{http://www.haplotype.org/download/bimbam-manual.pdf}{manual}.
}
\author{
Timothee Flutre
}
