% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{genoDoses2genoClasses}
\alias{genoDoses2genoClasses}
\title{Convert genotypes}
\usage{
genoDoses2genoClasses(X = NULL, tX = NULL, alleles, na.string = "--",
  verbose = 1)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the second allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns; the "second" allele is arbitrary, it corresponds to the second column of \code{alleles}, which can be the minor or the major allele}

\item{tX}{matrix with SNPs in rows and individuals in columns}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (exactly 2 columns are required)}

\item{na.string}{character used to replace NA values}

\item{verbose}{verbosity level (0/1)}
}
\value{
data.frame with SNPs in rows and individuals in columns
}
\description{
Convert SNP genotypes from "allele doses" into "genotypic classes".
Let us denote by "A" (or A1) the first allele, and "B" (or A2) the second allele.
The specific nucleotides (A, C, G or T) corresponding to "A" and "B" are specified via the \code{allele} argument to the function.
Moreover, whether A is the major (most frequent) allele or not, doesn't matter for the function to work.
At a given SNP, genotypes being homozygous for the first allele are coded as "0" in allele dose, and will be converted into "AA" in genotypic class.
Those being heterozygous are "1" and will be "AB".
Finally, those being homozygous for the second allele are "2" and will be "BB".
The code is not particularly efficient, but at least it exists.
}
\author{
Timothee Flutre
}
\seealso{
\code{link{genoClasses2genoDoses}}
}

