% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{genoDoses2Vcf}
\alias{genoDoses2Vcf}
\title{Convert genotypes}
\usage{
genoDoses2Vcf(X, snp.coords, alleles, verbose = 1)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded, for each SNP, in number of copies of its second allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; the "second" allele corresponds to the second column of \code{alleles} and will be interpreted as 'alt'}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "pos"}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (exactly 2 columns are required); the first column will be interpreted as 'ref' and the second column, which should correspond to the allele which number of copies is counted at each SNP in \code{X}, will be interpretd as 'alt'}

\item{verbose}{verbosity level (0/1)}
}
\value{
CollapsedVCF (see pkg \href{http://bioconductor.org/packages/VariantAnnotation/}{VariantAnnotation})
}
\description{
Convert SNP genotypes from "allele doses" into the \href{http://samtools.github.io/hts-specs/}{VCF} format.
}
\author{
Timothee Flutre
}

