% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{genomicControl}
\alias{genomicControl}
\title{Genomic control}
\usage{
genomicControl(pvalues)
}
\arguments{
\item{pvalues}{vector of raw p values (missing values will be omitted)}
}
\value{
numeric
}
\description{
Estimates lambda (\href{http://dx.doi.org/10.1006/tpbi.2001.1542}{Devlin et al, 2001}).
}
\examples{
\dontrun{## simulate SNP genotypes
set.seed(1859)
library(scrm)
nb.genos <- 200
Ne <- 10^4
chrom.len <- 10^5
mu <- 10^(-8)
c <- 10^(-8)
genomes <- simulCoalescent(nb.inds=nb.genos,
                           pop.mut.rate=4 * Ne * mu * chrom.len,
                           pop.recomb.rate=4 * Ne * c * chrom.len,
                           chrom.len=chrom.len)
X <- genomes$genos

## simulate SNP effects
Q <- 1
modelA <- simulBvsr(Q=Q, X=X, pi=0.1, pve=0.7, sigma.a2=1)

## filter out SNPs with low MAF
mafs <- estimSnpMaf(X)
maf_threshold <- 0.05
subset_snps <- names(mafs[which(mafs >= maf_threshold)])
length(subset_snps)

## perform SNP-by-SNP GWAS *without* kinship matrix
library(MM4LMM)
out_mmest <- MMEst(Y=modelA$Y[,1], X=X[,subset_snps],
                   VarList=list(Error=diag(nrow(modelA$Y))),
                                Verbose=FALSE)
out_anovatest <- AnovaTest(out_mmest, Type="TypeI")
out_gwas <- t(sapply(out_anovatest, function(x){x["Xeffect",]}))
head(out_gwas)

## look at p values
qqplotPval(out_gwas[,"pval"])
genomicControl(out_gwas[,"pval"])

## estimate genetic relatedness
A_vr <- estimGenRel(X, relationships="additive", method="vanraden1")

## perform SNP-by-SNP GWAS *with* kinship matrix
library(MM4LMM)
out_mmest <- MMEst(Y=modelA$Y[,1], X=X[,subset_snps],
                   VarList=list(Additive=A_vr,
                                Error=diag(nrow(modelA$Y))),
                                Verbose=FALSE)
out_anovatest <- AnovaTest(out_mmest, Type="TypeI")
out_gwas <- t(sapply(out_anovatest, function(x){x["Xeffect",]}))

## look at p values
qqplotPval(out_gwas[,"pval"])
genomicControl(out_gwas[,"pval"])
}
}
\author{
Timothee Flutre
}
