% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{genotypes.dose2bimbam}
\alias{genotypes.dose2bimbam}
\title{Convert genotype data to the "mean genotype" file format from BimBam}
\usage{
genotypes.dose2bimbam(X = NULL, tX = NULL, alleles, file = NULL)
}
\arguments{
\item{X}{matrix with individuals in rows and SNPs in columns}

\item{tX}{matrix with SNPs in rows and individuals in columns}

\item{alleles}{data.frame with SNPs in rows (names as row names) and
alleles in columns (first is "minor", second is "major")}

\item{file}{prints the genotype data to this file if non NULL (for instance
'genotypes_bimbam.txt' or gzfile('genotypes_bimbam.txt.gz'))}
}
\value{
data.frame
}
\description{
The format is specified in BimBam's manual http://www.haplotype.org/download/bimbam-manual.pdf#page=6
}
\author{
Timothée Flutre
}

