% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{getBreedingDirs}
\alias{getBreedingDirs}
\title{Set up}
\usage{
getBreedingDirs(root.dir)
}
\arguments{
\item{root.dir}{path to the root directory}
}
\value{
list
}
\description{
Retrieve the paths to the directories used for the breeding game.
}
\author{
Timothee Flutre
}

