% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{getBreedingGameConstants}
\alias{getBreedingGameConstants}
\title{Get the breeding game constants}
\usage{
getBreedingGameConstants(dbname, table = "constants")
}
\arguments{
\item{dbname}{name of the SQLite database (full path)}

\item{table}{name of the table}
}
\value{
list
}
\description{
Retrieve the constants used to parametrized the breeding game from the SQLite database.
}
\author{
Timothee Flutre
}
