% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{getBreedingGameSetup}
\alias{getBreedingGameSetup}
\title{Set up breeding game}
\usage{
getBreedingGameSetup(root.dir)
}
\arguments{
\item{root.dir}{path to the root directory}
}
\value{
list
}
\description{
Retrieve the paths to the directories used for the breeding game.
}
\author{
Timothee Flutre
}
