% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getEllipsisCoords}
\alias{getEllipsisCoords}
\title{Coords of vcov as ellipsis}
\usage{
getEllipsisCoords(mat, center = c(x = 0, y = 0))
}
\arguments{
\item{mat}{input square matrix}

\item{center}{vector with the x and y coordinates of the center (population means)}
}
\value{
two-column matrix with x and y coordinates; see intermediary values in the attributes
}
\description{
Returns the coordinates of the ellipsis corresponding to a variance-covariance matrix.
}
\seealso{
\code{plotEllipsis}
}
\author{
Timothee Flutre
}
