% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{getJoinMapSegregs}
\alias{getJoinMapSegregs}
\title{JoinMap segregation types}
\usage{
getJoinMapSegregs(genos, na.string = "--")
}
\arguments{
\item{genos}{data frame containing genotypes encoded in the JoinMap format,  similar to the output from \code{\link{genoClasses2JoinMap}}}

\item{na.string}{character indicating a missing genotype}
}
\value{
vector that can be used in \code{\link{writeSegregJoinMap}}
}
\description{
Get segregation types for each marker from a data frame of marker genotypes in the JoinMap format.
}
\author{
Charlotte Brault, Timothee Flutre
}
