% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getMeanVarBetaDist}
\alias{getMeanVarBetaDist}
\title{Beta distribution}
\usage{
getMeanVarBetaDist(alpha, beta)
}
\arguments{
\item{alpha}{first parameter}

\item{beta}{second parameter}
}
\value{
vector
}
\description{
Return the values of the mean and variance for the \href{https://en.wikipedia.org/wiki/Beta_distribution}{Beta distribution} given its parameters.
}
\author{
Timothee Flutre
}
