% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getNbPCsMinimAvgSqPartCor}
\alias{getNbPCsMinimAvgSqPartCor}
\title{Choose the number of PCs}
\usage{
getNbPCsMinimAvgSqPartCor(X)
}
\arguments{
\item{X}{matrix (e.g. genotypes in \{0,1,2\} with SNPs in rows and individuals in columns)}
}
\value{
integer
}
\description{
Return the number of PCs that minimizes the average squared partial correlation, from Shriner (Heredity, 2011).
}
\author{
Daniel Shriner [aut], Timothee Flutre [cre]
}
