% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getParamsBetaDist}
\alias{getParamsBetaDist}
\title{Beta distribution}
\usage{
getParamsBetaDist(m, v)
}
\arguments{
\item{m}{mean of the Beta distribution}

\item{v}{variance of the Beta distribution}
}
\value{
vector
}
\description{
Return the value of both parameters for the \href{https://en.wikipedia.org/wiki/Beta_distribution}{Beta distribution} given its mean and variance.
}
\examples{
\dontrun{set.seed(1859)
(beta.params <- getParamsBetaDist(m=0.5, v=0.1))
x <- rbeta(n=10^3, shape1=beta.params[1], shape2=beta.params[2])
mean(x)
var(x)
}
}
\author{
Timothee Flutre
}

