% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{getSamplesFromVcfFile}
\alias{getSamplesFromVcfFile}
\title{Samples from VCF file}
\usage{
getSamplesFromVcfFile(vcf.file)
}
\arguments{
\item{vcf.file}{path to the VCF file}
}
\value{
vector of samples
}
\description{
Return the samples from a VCF file using \code{\link{pipe}} and \code{cat} or \code{zcat} (depending on the file extension).
}
\author{
Timothee Flutre
}
