% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{getSparseInv}
\alias{getSparseInv}
\title{Entries of a sparse inverse}
\usage{
getSparseInv(D, U, verbose = FALSE)
}
\arguments{
\item{D}{diagonal matrix}

\item{U}{upper triangular matrix}

\item{verbose}{verbosity level}
}
\value{
matrix
}
\description{
Efficiently computes the entries of the sparse inverse (C) of a square, symmetric matrix (W) from its square-root-free Cholesky decomposition (W = L D U with U=L') as implemented in \code{\link{rootfreeChol}}.
This is notably used in quantitative genetics as described initially by \href{http://doi.org/10.3168/jds.S0022-0302(93)77478-0}{Misztal and Perez-Enciso (1993)}.
See the section 15.7 "Computing entries of the inverse of a sparse matrix" from Duff et al (2017).
}
\examples{
\dontrun{## example matrix from Misztal and Perez-Enciso (1993):
W <- matrix(NA, nrow=5, ncol=5)
W[1,] <- c(5, 0, 0, 3, 0)
W[2,-1] <- c(3, 0, 0, 1)
W[3,-c(1:2)] <- c(4, 1, 1)
W[4,-c(1:3)] <- c(4, 0)
W[5,-c(1:4)] <- 2
W
W[lower.tri(W)] <- t(W)[lower.tri(W)]
W

## square root free Cholesky decomposition:
decomp <- rootfreeChol(W)

## C, inverse of W:
D <- diag(diag(decomp))
U <- decomp
diag(U) <- 1
U<- t(U)
getSparseInv(D, U)
}
}
\seealso{
\code{\link{rootfreeChol}}
}
\author{
Timothee Flutre
}
