% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{getVcfOnlySingleElem}
\alias{getVcfOnlySingleElem}
\title{Keep single-element VCF records}
\usage{
getVcfOnlySingleElem(vcf)
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg VariantAnnotation)}
}
\value{
CollapsedVCF
}
\description{
Returns a vcf object in which all multi-element 'ref' or 'alt' records and indels were discarded.
}
\author{
Gautier Sarah, Timothee Flutre
}

