% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{gtVcf2dose}
\alias{gtVcf2dose}
\title{Convert GT to dosage}
\usage{
gtVcf2dose(vcf)
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg VariantAnnotation)}
}
\value{
matrix with variants in rows and samples in columns
}
\description{
From Martin Morgan (see http://grokbase.com/t/r/bioconductor/135b460s2b/bioc-how-to-convert-genotype-snp-matrix-to-nucleotide-genotypes)
}
\author{
Timothee Flutre
}

