% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{gtVcf2genoClasses}
\alias{gtVcf2genoClasses}
\title{Convert GT to genotypic classes}
\usage{
gtVcf2genoClasses(vcf, na.string = "NN")
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg VariantAnnotation)}

\item{na.string}{a symbol to indicate missing genotypes (e.g. "NN" or NA)}
}
\value{
matrix with variants in rows and samples in columns
}
\description{
Non-bi-allelic variants are discarded.
From Martin Morgan (see http://grokbase.com/t/r/bioconductor/135b460s2b/bioc-how-to-convert-genotype-snp-matrix-to-nucleotide-genotypes).
}
\author{
Gautier Sarah
}

