% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{h2v1WX2019}
\alias{h2v1WX2019}
\title{GWAS power analysis}
\usage{
h2v1WX2019(lambda = 1, n0 = 2000, power = 0.85, m = 1e+05, alpha = 0.05)
}
\arguments{
\item{lambda}{lambda}

\item{n0}{effective population size}

\item{power}{power}

\item{m}{number of markers}

\item{alpha}{alpha}
}
\value{
numeric
}
\description{
Computes a QTL using heritability effective population size.
Useful for the power analysis of a GWAS from \href{http://www.nature.com/articles/s41437-019-0205-3}{Wang and Xu (2019)}.
Code from \href{https://onlinelibrary.wiley.com/doi/10.1002/csc2.20692}{White et al (2020)}.
}
\author{
Jon White
}
