% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hinton}
\alias{hinton}
\title{Hinton diagram}
\usage{
hinton(m, main = "", max.sqrt.m = NULL)
}
\arguments{
\item{m}{matrix}

\item{main}{main title}

\item{max.sqrt.m}{to play with the scaling}
}
\description{
Plot a Hinton diagram
}
\details{
Modified from http://www.cs.princeton.edu/~mimno/factor-analysis.R
}
\author{
Timothee Flutre
}

