% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{imageWithScale}
\alias{imageWithScale}
\title{Plot a matrix as a heatmap in its natural orientation, with a colored
scale on the right side, and optionally using its dimension names for
rows and columns}
\usage{
imageWithScale(z, main = NULL, idx.rownames = NULL, idx.colnames = NULL,
  breaks = NULL)
}
\arguments{
\item{z}{matrix to be plotted}

\item{main}{title to appear above the heatmap}

\item{idx.rownames}{vector giving the indices of the row names of z to be added on the left side of the plot}

\item{idx.colnames}{vector giving the indices of the column names of z to be added on top of the plot}

\item{breaks}{vector (default=seq(min(z), max(z), length.out=100))}
}
\description{
To print all row names, choose idx.rownames=1:nrow(z). To print a subset
of 10 row names, choose idx.rownames=floor(seq(1, nrow(z), length.out=10)).
Similarly for column names.
}
\author{
Timothee Flutre
}

