% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{indexGenoDoses}
\alias{indexGenoDoses}
\title{Index the dose of each SNP genotype}
\usage{
indexGenoDoses(X, boundaries = seq(from = 0, to = 2, length.out = 4))
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded, for each SNP, in number of copies of its second allele, i.e. as allele doses in \{0,1,2\}, with genotypes in rows and SNPs in columns; the "second" allele is arbitrary, which can be the minor or the major allele}

\item{boundaries}{vector with interval boundaries of allele dose}
}
\value{
matrix of logicals with two columns, is.0 and is.1
}
\description{
Return if a given SNP enotype is 0 or 1, even if it was imputed.
}
\author{
Timothee Flutre
}
