% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{indexVcfFile}
\alias{indexVcfFile}
\title{Index a VCF file}
\usage{
indexVcfFile(vcf.file, software = "Rsamtools", nb.cores = 1)
}
\arguments{
\item{vcf.file}{path to a VCF file (must be sorted!)}

\item{software}{name of the software with which to index the file (Rsamtools/bcftools)}

\item{nb.cores}{number of threads for "bcftools"}
}
\value{
invisible path to the index file
}
\description{
Index a VCF file if it exists and if its index file doesn't already exist or is older.
}
\seealso{
\code{\link{sortVcfFile}}
}
\author{
Timothee Flutre
}
