% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{infoVariantCalls}
\alias{infoVariantCalls}
\title{Information on variant-level calls}
\usage{
infoVariantCalls(x, type = "SNP", thresh.qual = 20, thresh.qd = 2,
  thresh.fs = 60, thresh.mq = 40, thresh.mqrs = -12.5, thresh.rprs = -8,
  thresh.hs = 13, thresh.dp = NULL, thresh.crs = NULL,
  thresh.pnb = NULL, thresh.bqrs = NULL, thresh.gc = NULL,
  thresh.hw = 7, thresh.hr = 6, thresh.ic = NULL, thresh.lrs = NULL,
  thresh.sor = NULL)
}
\arguments{
\item{x}{data.frame, e.g. from "bcftools query --format '\%CHROM\\t\%POS\\t\%TYPE\\t...'"}

\item{type}{variant type}

\item{thresh.qual}{exclude variant if QUAL < threshold}

\item{thresh.qd}{exclude variant if QD < threshold}

\item{thresh.fs}{exclude variant if FS > threshold}

\item{thresh.mq}{exclude variant if MQ < threshold}

\item{thresh.mqrs}{exclude variant if MQRankSum < threshold (must be negative)}

\item{thresh.rprs}{exclude variant if ReadPosRankSum < threshold (must be negative)}

\item{thresh.hs}{exclude variant if HaplotypeScore > threshold (must be positive, \url{http://dx.doi.org/10.1038/sdata.2015.11})}

\item{thresh.dp}{if NULL, exclude variant if DP > 3 * mean(DP) (\url{https://www.biostars.org/p/110670/#110671})}

\item{thresh.crs}{exclude variant if ClippingRankSum < threshold (must be negative)}

\item{thresh.pnb}{exclude variant if PercentNBase < threshold (must be in [0,100])}

\item{thresh.bqrs}{exclude variant if BaseQRankSum < threshold (must be negative)}

\item{thresh.gc}{exclude variant if GC > threshold (must be in [0,1])}

\item{thresh.hw}{exclude variant if HW > threshold (must be positive)}

\item{thresh.hr}{exclude variant if HRun > threshold (must be positive, \url{http://dx.doi.org/10.1038/sdata.2015.11})}

\item{thresh.ic}{exclude variant if InbreedingCoeff > threshold}

\item{thresh.lrs}{exclude variant if LikelihoodRankSum < threshold (must be negative)}

\item{thresh.sor}{exclude variant if SOR > threshold (must be positive)}
}
\value{
list
}
\description{
Return some information to help in hard-filtering variant-level calls. See GATK's Best Practices tutorial (\url{https://www.broadinstitute.org/gatk/guide/topic?name=tutorials#tutorials2806}).
}
\author{
Timothee Flutre
}

