% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{initPlates}
\alias{initPlates}
\title{Initialize plate(s)}
\usage{
initPlates(n, nrow, ncol, names)
}
\arguments{
\item{n}{number of plates}

\item{nrow}{vector of number of rows for each plate}

\item{ncol}{vector of number of columns for each plate}

\item{names}{vector of names for each plate}
}
\value{
list of matrices, one per plate, in the "wide" format
}
\description{
Initialize plates as matrices with missing data.
}
\author{
Timothee Flutre
}
