% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inlineFctForm}
\alias{inlineFctForm}
\title{Inline function in formula}
\usage{
inlineFctForm(form, only.resp = TRUE)
}
\arguments{
\item{form}{character corresponding to a formula, e.g. \code{"y ~ 1 + x"}}

\item{only.resp}{if TRUE, the presence of inline function(s) is detected only in the response}
}
\value{
list with one component per term of the formula (first the response and then the predictors, if any), each component containing a vector of characters with the untransformed variable first and then the inline function(s) from the outermost to the innermost, or NA if there is no inline function corresponding to this term
}
\description{
Detect the presence of inline function(s) in a formula and parse them.
}
\author{
Timothee Flutre
}
