% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{isSingular}
\alias{isSingular}
\title{Singular matrix}
\usage{
isSingular(x)
}
\arguments{
\item{x}{matrix}
}
\value{
logical
}
\description{
Assess if a matrix is singular, i.e. not full rank, by comparing its condition number with the relative machine precision.
As \href{http://www.stat.wisc.edu/~st849-1/Rnotes/ModelMatrices.html#sec-2_2}{explained by Douglas Bates}, "a matrix is regarded as being numerically singular when its reciprocal condition number is less than the relative machine precision".
}
\author{
Timothee Flutre
}
