% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{jagsAM}
\alias{jagsAM}
\title{Animal model}
\usage{
jagsAM(dat, relmat, inits = NULL, priors = list(fix = list(dist = "c", par =
  5), vc = list(dist = "hc", par = 5)), nb.chains = 1, burnin = 10^2,
  nb.iters = 10^3, thin = 10, rm.jags.file = TRUE, verbose = 0)
}
\arguments{
\item{dat}{data.frame containing the data corresponding to relmat; should have a column grep-able for "response" as well as a column "geno.add" used with matrix A; if a column "geno.dom" exists, it will be used with matrix D; any other column will be interpreted as corresponding to "fixed effects"}

\item{relmat}{list containing the matrices of genetic relationships (see \code{\link{estimGenRel}}); additive relationships (matrix A) are compulsory, with name "geno.add"; dominant relationships (matrix D) are optional, with name "geno.dom"; can be in the "matrix" class (base) or the "dsCMatrix" class (Matrix package)}

\item{inits}{list of initial values (possible to use 1 sub-list per chain, see \code{\link[rjags]{jags.model}}); if NULL, JAGS will choose typical values (usually mean, median, or mode of the prior)}

\item{priors}{list of specifying priors; each component is itself a list for which "dist" specifies the distribution and "par" the parameter; a component named "fix" is used for fixed effects,  so that ("c",5) means c[q] ~ Cauchy(0,5) and ("n",10^6) means c[q] ~ Normal(0,10^6), but note that the global intercept always is Normal(mean(y),10^6); a component named "vc" is used for variance components, so that ("hc",5) means stdev ~ half-Cauchy(0,5) and ("ig",0.001) means var ~ InvGamma(0.001,0.001)}

\item{nb.chains}{the number of parallel chains for the model (see \code{\link[rjags]{jags.model}})}

\item{burnin}{number of initial iterations to discard}

\item{nb.iters}{number of iterations to monitor (see \code{\link[rjags]{coda.samples}})}

\item{thin}{thinning interval for monitors (see \code{\link[rjags]{coda.samples}})}

\item{rm.jags.file}{remove the file specifying the model written in the JAGS-dialect of the BUGS language}

\item{verbose}{verbosity level (0/1)}
}
\value{
\code{\link[coda]{mcmc.list}} object
}
\description{
Given I genotypes, Q covariates and N=I*Q phenotypes for the trait, fit an "animal model" with the rjags package via the following likelihood: y = W c + Z g_A + Z g_D + epsilon, where y is Nx1; W is NxQ; Z is NxI; g_A ~ Normal_I(0, sigma_A^2 A) with A the known matrix of additive genetic relationships; g_D ~ Normal_I(0, sigma_D^2 D) with D the known matrix of dominant genetic relationships; epsilon ~ Normal_N(0, sigma^2 Id_N); Cov(g_A,g_D)=0; Cov(g_A,e)=0; Cov(g_D,e)=0.
}
\examples{
\dontrun{## simulate genotypes
set.seed(1859)
I <- 100 # genotypes
P <- 2000 # SNPs
X <- matrix(sample(0:2, size=I*P, replace=TRUE), nrow=I, ncol=P,
            dimnames=list(paste0("geno", 1:I), paste0("snp", 1:P)))

## simulate phenotypes with only additive part of genotypic values
A <- estimGenRel(X, relationships="additive", method="vanraden1", verbose=0)
A <- as.matrix(Matrix::nearPD(A)$mat) # not always necessary
modelA <- simulAnimalModelMultivar(T=1, Q=3, A=A, sigma.A2=15, V.E=5)

## infer with rjags
modelA$dat$geno.add <- modelA$dat$geno; modelA$dat$geno <- NULL
fitA <- jagsAM(dat=modelA$dat, relmat=list(geno.add=A))
plotMcmcChain(fitA[[1]], "sigma.A2", 1:4, modelA$sigma.A2)
cbind(truth=c(c(modelA$C), modelA$sigma.A2, modelA$V.E),
      summaryMcmcChain(fitA[[1]], c("c[1]", "c[2]", "c[3]", "sigma.A2", "V.E")))
}
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{lmerAM}}, \code{\link{inlaAM}}
}

