% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{jagsAMmv}
\alias{jagsAMmv}
\title{Animal model (multivariate)}
\usage{
jagsAMmv(
  data,
  relmat,
  nb.chains = 1,
  nb.adapt = 10^3,
  burnin = 10^2,
  nb.iters = 10^3,
  thin = 10,
  progress.bar = NULL,
  rm.jags.file = TRUE,
  verbose = 0
)
}
\arguments{
\item{data}{data.frame containing the data corresponding to relmat; should have columns grep-able for "response" as well as a column "geno.add" used with matrix A; if a column "geno.dom" exists, it will be used with matrix D; any other column will be interpreted as corresponding to "fixed effects"}

\item{relmat}{list containing the matrices of genetic relationships (see \code{\link{estimGenRel}}); additive relationships (matrix A) are compulsory, with name "geno.add"; dominant relationships (matrix D) are optional, with name "geno.dom"; can be in the "matrix" class (base) or the "dsCMatrix" class (Matrix package)}

\item{nb.chains}{number of independent chains to run (see \code{\link[rjags]{jags.model}})}

\item{nb.adapt}{number of iterations for adaptation (see \code{\link[rjags]{jags.model}})}

\item{burnin}{number of initial iterations to discard (see the update function of the rjags package)}

\item{nb.iters}{number of iterations to monitor (see \code{\link[rjags]{coda.samples}})}

\item{thin}{thinning interval for monitored iterations (see \code{\link[rjags]{coda.samples}})}

\item{progress.bar}{type of progress bar (text/gui/none or NULL)}

\item{rm.jags.file}{remove the file specifying the model written in the JAGS-dialect of the BUGS language}

\item{verbose}{verbosity level (0/1)}
}
\description{
Given T traits, I genotypes, Q covariates and N=I*Q phenotypes per trait, fit an "animal model" with the rstan package via the following likelihood: \eqn{Y = W C + Z G_A + Z G_D + E}, where Y is NxT; W is NxQ; Z is NxI; G_A ~ Normal_IxT(0, A, V_G_A) with A the known matrix of additive genetic relationships; G_D ~ Normal_IxT(0, D, V_G_D) with D the known matrix of dominant genetic relationships; E ~ Normal_NxT(0, Id_N, V_E); Missing phenotypes are jointly imputed with the other unknown variables, and the errors can follow a Student's t distribution to handle outliers.
}
\seealso{
\code{\link{jagsAM}}
}
\author{
Timothee Flutre, inspired by code from Najla Saad Elhezzani (arXiv:1507.08638)
}
