% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{joinMap2backcross}
\alias{joinMap2backcross}
\title{Genotype coding}
\usage{
joinMap2backcross(x, alias.hom = "A", alias.het = "H", alias.dup = "Z",
  alias.miss = "-", parent.names = c("parent1", "parent2"), verbose = 1)
}
\arguments{
\item{x}{data frame in the JoinMap format, for instance from \code{\link{genoClasses2JoinMap}}; any column named "phase" or "clas" will be discarded from the start; a column named "seg" should be present (markers with missing segregation will be discarded); a column named "locus" should be present unless row names are given; all other columns are supposed to correspond to genotypes}

\item{alias.hom}{alias to specify homozygotes ("A" for CarthaGene, 1 for R/qtl); should be different from alias.dup}

\item{alias.het}{alias to specify heterozygotes ("H" for CarthaGene, 2 for R/qtl); should be different from alias.dup}

\item{alias.dup}{alias used when duplicating the locus; should be different from both alias.hom and alias.het, but should be of the same mode (i.e. character or numeric)}

\item{alias.miss}{alias to specify missing values}

\item{parent.names}{vector containing the names of the parents}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
list of two matrices, one per parent
}
\description{
Convert unphased genotypes encoded in the \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} format into the backcross configuration according to the \href{http://www7.inra.fr/mia/T/CarthaGene/}{CarthaGene} or R/qtl formats.
Genotypes are assumed to come from a bi-parental cross with outbred parents (population type "CP" in JoinMap) and the goal is to build two parental genetic maps via a pseudo-test-cross strategy (Grattapaglia and Sederoff, 1994).
This function hence encodes genotypes as two backcrosses, one per parent.
All locus will be duplicated and the suffix "_m" (for "mirror") added to their identifiers.
}
\seealso{
\code{\link{genoClasses2JoinMap}}, \code{\link{openCarthagene}}, \code{\link{setJoinMapPhasesFromParentalLinkGroups}}
}
\author{
Timothee Flutre [aut], Agnes Doligez [ctb]
}
