% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{joinMap2CarthaGene}
\alias{joinMap2CarthaGene}
\title{Genotype coding}
\usage{
joinMap2CarthaGene(x, verbose = 1)
}
\arguments{
\item{x}{data frame in the JoinMap format, for instance from \code{\link{genoClasses2JoinMap}}; columns named "phase" and "clas" will be discarded from the start; a column named "seg" should be present; a column named "locus" should be present unless row names are given; all other columns are supposed to correspond to genotypes}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
list of two matrices, one per parent
}
\description{
Convert genotypes encoded in the \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} format into the \href{http://www7.inra.fr/mia/T/CarthaGene/}{CarthaGene} format.
Genotypes are assumed to come from a bi-parental family with heterozygous parents (population type CP in JoinMap) and the goal is to build two parental genetic maps via a pseudo-test-cross strategy (Grattapaglia and Sederoff, 1994).
This function hence encodes genotypes as two backcrosses, one per parent, with "A" for "homozygous" and "H" for "heterozygous".
All locus will be duplicated and the suffix "_m" (for "mirror") added to their identifiers.
}
\seealso{
\code{\link{genoClasses2JoinMap}}
}
\author{
Timothee Flutre [aut], Agnes Doligez [ctb]
}
