% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{lengthenPlate}
\alias{lengthenPlate}
\title{Lengthen plate}
\usage{
lengthenPlate(plate.w)
}
\arguments{
\item{plate.w}{matrix of a plate in the "wide" format}
}
\value{
data.frame of a plate in the "long" format (1 well per row)
}
\description{
Lengthen a "wide" plate into 3 columns for easier processing.
}
\author{
Timothee Flutre
}

