% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{lmerAM}
\alias{lmerAM}
\title{Fit a basic "animal model" with lme4}
\usage{
lmerAM(formula, data, relmat, REML = TRUE, verbose = 1)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame containing the data corresponding to formula and relmat}

\item{relmat}{list containing the matrix of additive genetic relationships (should use the same name as the colname in data)}

\item{REML}{default is TRUE, but use FALSE to compare models with different fixed effects}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
merMod object
}
\description{
y = W alpha + Z u + epsilon
where y is N x 1; W is N x Q; Z is N x I
u ~ Norm_I(0, G=sigma_u^2 A); epsilon_n ~ N(0, sigma^2 I); Cov(u,e)=0
See http://stackoverflow.com/q/19327088/597069.
Still problems: see https://github.com/lme4/lme4/issues/340.
}
\author{
Timothee Flutre
}

