% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{loadBlast}
\alias{loadBlast}
\title{BLAST}
\usage{
loadBlast(file.coords, reformat.strand = TRUE, apply.sort = TRUE,
  verbose = 1)
}
\arguments{
\item{file.coords}{path to the file with alignment coordinates obtained with \code{-outfmt "6 std qlen sstrand"}}

\item{reformat.strand}{if TRUE, "plus" is replaced by "+", and "minus" by "-"}

\item{apply.sort}{if TRUE, return the alignments sorted in increasing order for qseqid, then increasing for evalue, then decreasing for bitscore, then decreasing for length, then decreasing for pident, then increasing for sseqid}

\item{verbose}{verbosity level (0/1)}
}
\value{
data.frame with 14 columns
}
\description{
Load alignment coordinates from the \href{http://blast.ncbi.nlm.nih.gov/Blast.cgi}{NCBI BLAST}.
}
\examples{
\dontrun{## BLAST should be run beforehand, for instance:
## zcat chroms.fa.gz | makeblastdb -dbtype nucl -in - -title chroms -out chroms -logfile makeblastdb_chroms.log
## echo "zcat seqs.fa.gz | blastn -query - -task megablast -db chroms -out /dev/stdout -outfmt \\"6 std qlen sstrand\\" | gzip > megablast_chroms_seqs.txt.gz"  | qsub ...
coords <- loadBlast("megablast_chroms_seqs.txt.gz")
}
}
\author{
Timothee Flutre
}

