% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{loadMummer}
\alias{loadMummer}
\title{MUMmer}
\usage{
loadMummer(file.coords, algo = "nucmer", keep.nlong.qry = NULL,
  ref.lim = NULL, verbose = 1)
}
\arguments{
\item{file.coords}{path to the file with the coordinates obtained via \code{show-coords -c -l -r -T ... | gzip > ...} (must be gzipped)}

\item{algo}{nucmer or promer (the latter isn't available yet)}

\item{keep.nlong.qry}{keep the N queries having the longest cumulated alignments on the reference (keep all of them by default)}

\item{ref.lim}{limits on reference coordinates outside which alignments are discarded (keep all of them by default)}

\item{verbose}{verbosity level (0/1)}
}
\value{
data.frame with the coordinates
}
\description{
Load alignment coordinates from \href{http://mummer.sourceforge.net/}{MUMmer}.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{plotMummer}}
}

