% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{loadReadCountsPerIndAndLane}
\alias{loadReadCountsPerIndAndLane}
\title{Load read counts}
\usage{
loadReadCountsPerIndAndLane(lanes.dir)
}
\arguments{
\item{lanes.dir}{vector of paths to "lane" directories, each containing a "demultiplex" directory in which \href{https://github.com/timflutre/quantgen/blob/master/demultiplex.py}{demultiplex.py} was executed}
}
\value{
data.frame with (individual,lane) in rows and counts in columns
}
\description{
Load read counts per individual and lane
}
\author{
Timothee Flutre
}
