% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{log10WeightedSum}
\alias{log10WeightedSum}
\title{Log of weighted sum}
\usage{
log10WeightedSum(x, weights = NULL)
}
\arguments{
\item{x}{vector of log10 values}

\item{weights}{weights (use equal weights if not specified)}
}
\value{
numeric
}
\description{
Performs the stable computation of \eqn{log_{10}(\sum_i w_i 10^x_i)}.
}
\author{
Timothee Flutre
}
