% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(error)
}
\arguments{
\item{error}{vector \eqn{\hat{\theta}_i - \theta_i}}
}
\value{
numeric
}
\description{
Returns the mean absolute error (MAE).
}
\author{
Timothee Flutre
}
