% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{maf.from.dose}
\alias{maf.from.dose}
\title{Estimate minor allele frequencies of SNPs.}
\usage{
maf.from.dose(X)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses, with SNPs in
columns and individuals in rows}
}
\value{
vector
}
\description{
Missing values should be encoded as NA.
}
\author{
Timothee Flutre
}

