% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{maf2genoFreq}
\alias{maf2genoFreq}
\title{Genotype frequencies}
\usage{
maf2genoFreq(maf)
}
\arguments{
\item{maf}{frequency of the minor allele, a}
}
\value{
vector of genotype frequencies
}
\description{
Calculate the genotype frequencies at a bi-allelic SNP, from its minor allele frequency, assuming Hardy-Weinberg equilibrium.
}
\examples{
\dontrun{set.seed(1859)
genos <- sample(x=c(0,1,2), size=n, replace=TRUE, prob=maf2genoFreq(maf))
table(genos)
}
}
\author{
Timothee Flutre
}
