% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{make.formula}
\alias{make.formula}
\title{make.formula}
\usage{
make.formula(cross, big_list, range_qtl, nb_run, pLOD = FALSE)
}
\arguments{
\item{cross}{object}

\item{big_list}{list}

\item{range_qtl}{vector}

\item{nb_run}{integer}

\item{pLOD}{logical}
}
\value{
list
}
\description{
Useful function for MIMQTL.
}
\seealso{
\code{\link{MIMQTL}}
}
