% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeCross}
\alias{makeCross}
\title{Cross}
\usage{
makeCross(haplos.par1, loc.crossovers.par1, haplos.par2 = NULL,
  loc.crossovers.par2 = NULL, child.name = NULL, verbose = 0)
}
\arguments{
\item{haplos.par1}{list of matrices (one per chromosome) for the first parent}

\item{loc.crossovers.par1}{list of vectors (one per chromosome) for the first parent}

\item{haplos.par2}{list of matrices (one per chromosome) for the second parent}

\item{loc.crossovers.par2}{list of vectors (one per chromosome) for the second parent}

\item{child.name}{identifier of the child (if NULL, <parent1>"-x-"<parent2> or <parent1>"-hd")}

\item{verbose}{verbosity level (default=0/1)}
}
\value{
list of matrices (one per chromosome) for the child
}
\description{
Make a cross. If two different individuals are given, then a fecundation is made with one gamete from each individual. If the same individual is given twice, an autofecondation is made with two different gametes from this individual. If a single individual is given, a haplodiploidization is made with a single gamete from this individual.
}
\author{
Timothee Flutre
}

