% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{makeDfInitPhenos}
\alias{makeDfInitPhenos}
\title{Simul breeding game}
\usage{
makeDfInitPhenos(nb.lines.per.year = 150, nb.years = 10,
  nb.plots.per.line.per.year = 2, first.year = 2005, line.ids)
}
\arguments{
\item{nb.lines.per.year}{number of lines phenotyped each year}

\item{nb.years}{number of years of phenotyping}

\item{nb.plots.per.line.per.year}{number of plots per line per year}

\item{first.year}{numeric of the year of the first phenotyping}

\item{line.ids}{vector of line identifiers}
}
\value{
data.frame
}
\description{
Make the structure of the data.frame that will be given to the players at the beginning of the game.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{makeDfPhenos}}
}

