% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{makeDfPhenos}
\alias{makeDfPhenos}
\title{Simul breeding game}
\usage{
makeDfPhenos(ind.ids, nb.plots.per.ind, year, pathogen)
}
\arguments{
\item{ind.ids}{vector of genotype identifiers}

\item{nb.plots.per.ind}{vector with the number of plots at which each genotype should be phenotype}

\item{year}{numeric of the year at which phenotyping occurs}

\item{pathogen}{if TRUE, the pathogen will be present the given year}
}
\value{
data.frame
}
\description{
Make the structure of the data.frame that will be given to the players when they request phenotyping during the game.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{makeDfInitPhenos}}
}

