% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeDoubledHaploids}
\alias{makeDoubledHaploids}
\title{Doubled haploids}
\usage{
makeDoubledHaploids(haplos, ind.names = NULL, loc.crossovers = NULL,
  child.names = NULL, nb.cores = 1, verbose = 1)
}
\arguments{
\item{haplos}{list of matrices (one per chromosome)}

\item{ind.names}{identifiers of the individuals to handle (if NULL, all individuals)}

\item{loc.crossovers}{list of lists (one per chromosome, then one per individual); if NULL, draw many crossing-overs localizations at once (as Poisson with parameter 2, assuming all chromosomes roughly have the same length)}

\item{child.names}{identifiers of the children (if NULL, <ind>"_x_"<ind>)}

\item{nb.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously (not on Windows)}

\item{verbose}{verbosity level (default=0/1)}
}
\value{
list of matrices
}
\description{
Make a doubled haploid for each individual under consideration.
}
\author{
Timothee Flutre
}

