% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeDoubledHaploidsSingleInd}
\alias{makeDoubledHaploidsSingleInd}
\title{Doubled haploids}
\usage{
makeDoubledHaploidsSingleInd(haplos.ind, loc.crossovers, child.name = NULL)
}
\arguments{
\item{haplos.ind}{list of matrices (one per chromosome)}

\item{loc.crossovers}{list of vectors (one per chromosome)}

\item{child.name}{identifier of the child (if NULL, <parent1>"_x_"<parent2>)}
}
\value{
list of matrices
}
\description{
Make a doubled haploid from all chromosomes of the individual under consideration.
}
\author{
Timothee Flutre
}

