% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{makeExampleDataFile}
\alias{makeExampleDataFile}
\title{Example for breeding game}
\usage{
makeExampleDataFile(out.dir, lang = "fr")
}
\arguments{
\item{out.dir}{path to the directory in which the file will be saved}

\item{lang}{language to be used (fr/en)}
}
\value{
invisible data.frame
}
\description{
Make a file with examples of data to request (phenotypes, genotypes).
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{readCheckBreedDataFile}}
}

