% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeGameteSingleInd}
\alias{makeGameteSingleInd}
\title{Gamete}
\usage{
makeGameteSingleInd(haplos.par, loc.crossovers)
}
\arguments{
\item{haplos.par}{list of matrices containing both haplotypes of a parent for each chromosome}

\item{loc.crossovers}{list of vectors with positions of the crossing overs (the coordinate of the first nucleotide is assumed to be 1, and crossing-overs are assumed to occur right after the given localization)}
}
\value{
list of vectors (one per chromosome)
}
\description{
Make a gamete for all chromosomes by recombining two parental haplotypes.
}
\author{
Timothee Flutre
}

