% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeGridWenStephens}
\alias{makeGridWenStephens}
\title{Grid for Bayes Factors}
\usage{
makeGridWenStephens(grid.type = "general", no.het = FALSE)
}
\arguments{
\item{grid.type}{"general" indicates the meta-analysis grid (large), otherwise the configuration grid (small) is returned}

\item{no.het}{if TRUE, the grid is built without heterogeneity}
}
\value{
matrix with two columns named "phi2" and "oma2"
}
\description{
Makes the grid of prior variances used to compute Bayes factors as in Wen and Stephens (Annals of Applied Statistics, 2014).
}
\seealso{
\code{\link{calcL10ApproximateBayesFactorWenStephens}}
}
\author{
Timothee Flutre
}
