% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeMmeElements}
\alias{makeMmeElements}
\title{Make elements of MME}
\usage{
makeMmeElements(y, X, Z)
}
\arguments{
\item{y}{vector of phenotypes (or matrix with one column)}

\item{X}{incidence matrix of fixed effects}

\item{Z}{incidence matrix of random variables}
}
\value{
list of matrices
}
\description{
Make elements of Henderson's MME.
}
\seealso{
\code{link{solveMme}}
}
\author{
Timothee Flutre
}
