% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeMmeLhs}
\alias{makeMmeLhs}
\title{Make MME's left-hand side}
\usage{
makeMmeLhs(tX.X, tZ.X, tX.Z, tZ.Z, lambda, Ainv)
}
\arguments{
\item{tX.X}{X being the incidence matrix of fixed effects}

\item{tZ.X}{Z being the incidence matrix of random variables}

\item{tX.Z}{see above}

\item{tZ.Z}{see above}

\item{lambda}{ratio of the variance component of the errors over the variance component of the additive genotypic values (also called breeding values)}

\item{Ainv}{inverse of A, the matrix of additive genetic relationships}
}
\value{
matrix
}
\description{
Make Henderson's MME's left-hand side.
}
\seealso{
\code{link{solveMme}}
}
\author{
Timothee Flutre
}
