% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeMmeRhs}
\alias{makeMmeRhs}
\title{Make MME's right-hand side}
\usage{
makeMmeRhs(tX.y, tZ.y)
}
\arguments{
\item{tX.y}{X being the incidence matrix of fixed effects and y the vector of phenotypes}

\item{tZ.y}{Z being the incidence matrix of random variables}
}
\value{
matrix
}
\description{
Make Henderson's MME's right-hand side.
}
\seealso{
\code{link{solveMme}}
}
\author{
Timothee Flutre
}
