% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{matrixTrace}
\alias{matrixTrace}
\title{Trace of a matrix}
\usage{
matrixTrace(x = NULL, l = NULL)
}
\arguments{
\item{x}{matrix to be decomposed; ignored if \code{l} is provided}

\item{l}{vector of eigenvalues}
}
\value{
numeric
}
\description{
Return the trace of a matrix by summing its eigenvalues.
}
\author{
Timothee Flutre
}
