% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{matWide2Long}
\alias{matWide2Long}
\title{Reformat}
\usage{
matWide2Long(mat, stringsAsFactors = FALSE)
}
\arguments{
\item{mat}{matrix}

\item{stringsAsFactors}{if TRUE, character vectors will be converted to factors}
}
\value{
data frame
}
\description{
Reformats the diagonal and upper triangular parts of a matrix from the wide format into the long format.
}
\author{
Zheyuan Li [aut] (https://stackoverflow.com/a/41745373/597069), Timothee Flutre [ctb]
}
