% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{mergeOverlaps}
\alias{mergeOverlaps}
\title{Merge overlapping genomic intervals}
\usage{
mergeOverlaps(x, thresh = 0)
}
\arguments{
\item{x}{input \code{GRanges}}

\item{thresh}{threshold of relative overlap above which \code{GRanges} are merged}
}
\value{
\code{GRanges}
}
\description{
Merges overlapping genomic intervals (as proposed \href{https://support.bioconductor.org/p/68021/}{here}).
}
\author{
Herve Pages [aut], Timothee Flutre [ctb]
}
