% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{mme}
\alias{mme}
\title{Animal model}
\usage{
mme(y, W, Z, sigma.A2, Ainv, V.E)
}
\arguments{
\item{y}{vector of phenotypes}

\item{W}{incidence matrix of fixed effects}

\item{Z}{incidence matrix of random effects, the breeding values}

\item{sigma.A2}{variance component of the breeding values}

\item{Ainv}{inverse of A, the matrix of additive genetic relationships}

\item{V.E}{variance component of the errors}
}
\value{
vector of length QI containing the BLUEs of c and the BLUPs of g_A
}
\description{
Given T=1 trait, I genotypes, Q covariates and N=I*Q phenotypes per trait, compute the BLUEs and BLUPs of the following "animal model" via Henderson's mixed model equations: y = W c + Z g_A + e, where y is N x 1; W is N x Q; Z is N x I; g_A ~ Normal_Ix1(0, sigma_A^2 A); e ~ Normal_Nx1(0, V.E Id_N).
}
\examples{
## simulate genotypes
set.seed(1859)
X <- simulGenosDose(nb.genos=200, nb.snps=2000)

## estimate the additive genetic relationships
A <- estimGenRel(X, relationships="additive", method="vanraden1", verbose=0)
A <- as.matrix(Matrix::nearPD(A)$mat) # not always necessary

## simulate phenotypes
model <- simulAnimalModel(T=1, Q=3, A=A, V.G.A=15, V.E=5)

## calculate BLUEs and BLUPs
fit <- mme(y=model$Y[,1,drop=FALSE], W=model$W, Z=model$Z,
           sigma.A2=model$V.G.A, Ainv=solve(A), V.E=model$V.E)
cbind(model$C, fit[1:3])
cor(model$G.A, fit[4:length(fit)])
}
\author{
Timothee Flutre
}

