% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{mp.inv}
\alias{mp.inv}
\title{Pseudo-inverse}
\usage{
mp.inv(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix
}
\description{
Return the Moore-Penrose pseudo-inverse of a matrix
}
\author{
Timothee Flutre
}

