% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{msd}
\alias{msd}
\title{Mean signed difference}
\usage{
msd(error)
}
\arguments{
\item{error}{vector \eqn{\hat{\theta}_i - \theta_i}}
}
\value{
numeric
}
\description{
Returns the mean signed difference (MSD).
}
\author{
Timothee Flutre
}
