% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{n0WX2019}
\alias{n0WX2019}
\title{GWAS power analysis}
\usage{
n0WX2019(lambda = 1, eigvalues)
}
\arguments{
\item{lambda}{lambda}

\item{eigvalues}{eigenvalues of the kinship matrix}
}
\value{
numeric
}
\description{
Computes the effective population size.
Useful for the power analysis of a GWAS from \href{http://www.nature.com/articles/s41437-019-0205-3}{Wang and Xu (2019)}.
Code from \href{https://onlinelibrary.wiley.com/doi/10.1002/csc2.20692}{White et al (2020)}.
}
\author{
Jon White
}
