% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{orthoRotate2D}
\alias{orthoRotate2D}
\title{Rotation}
\usage{
orthoRotate2D(x, angle = 90)
}
\arguments{
\item{x}{2-column matrix to be rotated}

\item{angle}{angle of rotation (in degrees)}
}
\value{
matrix
}
\description{
Rotates the input matrix by a certain angle (useful for the output of PCA).
}
\seealso{
\code{\link{pca}}, \code{\link{plotPca}}
}
\author{
Timothee Flutre
}
