% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{overlayBarplots}
\alias{overlayBarplots}
\title{Overlay two barplots}
\usage{
overlayBarplots(
  x1,
  x2,
  nbBins,
  leg1,
  leg2,
  cols = c(rgb(0, 0, 1, 1/4), rgb(1, 0, 0, 1/4)),
  main = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{x1}{numeric vector for the first histogram}

\item{x2}{numeric vector for the second histogram}

\item{nbBins}{number of bins; will be the same for both histograms}

\item{leg1}{legend for the first histogram}

\item{leg2}{legend for the second histogram}

\item{cols}{vector of colors, one per histogram}

\item{main}{main title}

\item{xlab}{label of the x-axis}

\item{ylab}{label of the y-axis}
}
\value{
nothing
}
\description{
Plots two histograms as densities by overlaying two barplots.
}
\author{
Timothee Flutre
}
