% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{parseCgGroup}
\alias{parseCgGroup}
\title{Parse CarthaGene's output}
\usage{
parseCgGroup(out.group, mrk.info = NULL)
}
\arguments{
\item{out.group}{vector of character output from \code{\link{runCarthagene}} corresponding to "group"}

\item{mrk.info}{data frame output from \code{\link{parseCgMrkinfo}} (optional; useful to have the marker names instead of only the marker identifiers)}
}
\value{
list with one data frame per linkage group
}
\description{
Parse the output of CarthaGene's command "group".
}
\author{
Timothee Flutre
}
