% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{parseCgHeaprint}
\alias{parseCgHeaprint}
\title{Parse CarthaGene's output}
\usage{
parseCgHeaprint(out.heaprint)
}
\arguments{
\item{out.heaprint}{vector of character output from \code{\link{runCarthagene}} corresponding to "heaprint"}
}
\value{
data frame with one row per map in the heap
}
\description{
Parse the output of CarthaGene's command "heaprint".
}
\author{
Timothee Flutre
}
