% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{parseCgMaprintd}
\alias{parseCgMaprintd}
\title{Parse CarthaGene's output}
\usage{
parseCgMaprintd(out.maprintd)
}
\arguments{
\item{out.maprintd}{vector of character output from \code{\link{runCarthagene}} corresponding to "maprintd" or "bestmaprintd"}
}
\value{
data frame corresponding to the given linkage group
}
\description{
Parse the output of CarthaGene's command "maprintd" and "bestmaprintd", and add the cumulative Kosambi distance.
}
\author{
Timothee Flutre
}
