% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{parseCgMrkinfo}
\alias{parseCgMrkinfo}
\title{Parse CarthaGene's output}
\usage{
parseCgMrkinfo(out.mrkinfo, verbose = 1)
}
\arguments{
\item{out.mrkinfo}{vector of character output from \code{\link{runCarthagene}} corresponding to "mrkinfo"}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame with one row per marker and columns "id", "name", etc
}
\description{
Parse the output of CarthaGene's command "mrkinfo".
}
\author{
Timothee Flutre
}
