% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{parseCgPwMatrix}
\alias{parseCgPwMatrix}
\title{Parse CarthaGene's output}
\usage{
parseCgPwMatrix(out, cmd = "mrkfr2p", mrk.info = NULL, verbose = 1)
}
\arguments{
\item{out}{vector of character output from \code{\link{runCarthagene}} corresponding to "mrkfr2p" or "mrklod2p" or "mrkdist2p"}

\item{cmd}{"mrkfr2p" or "mrklod2p" or "mrkdist2p"}

\item{mrk.info}{data frame output from \code{\link{parseCgMrkinfo}} (optional; useful to have the marker names as row and column names in the returned matrix)}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix
}
\description{
Parse the output of CarthaGene's commands "mrkfr2p" or "mrklod2p" or "mrkdist2p".
}
\author{
Timothee Flutre
}
