% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{permuteAllelesInGenosDose}
\alias{permuteAllelesInGenosDose}
\title{Genotypes}
\usage{
permuteAllelesInGenosDose(X, snps.toperm, verbose = 0)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded, for each SNP, in number of copies of its second allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; the "second" allele is arbitrary, it can be the minor or the major allele}

\item{snps.toperm}{vector of SNP identifiers corresponding to the SNPs to which allele permutation will be performed (column names of \code{X})}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix of genotypes
}
\description{
Permute alleles in genotypes once alleles have been permuted in the corresponding haplotypes.
}
\seealso{
\code{\link{permuteAllelesInHaplosNum}}
}
\author{
Timothee Flutre
}
