% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{permuteAllelesInHaplosNum}
\alias{permuteAllelesInHaplosNum}
\title{Haplotypes}
\usage{
permuteAllelesInHaplosNum(haplos, snps.toperm, verbose = 0)
}
\arguments{
\item{haplos}{list of haplotypes returned by \code{scrm}, each component of which corresponds to a matrix with haplotypes in rows and SNP in columns}

\item{snps.toperm}{vector of SNP identifiers corresponding to the SNPs to which allele permutation will be performed (column names of \code{haplos})}

\item{verbose}{verbosity level (0/1)}
}
\value{
list of haplotypes
}
\description{
Permute alleles in haplotypes.
In the \href{https://cran.r-project.org/package=scrm}{scrm} package, haplotypes are made of 0's and 1's, the former indicating ancestral alleles and the latter derived alleles.
However, when one wants to simulate realistic data, one often converts haplotypes into genotypes encoded as allele dose.
At this step, one may not want to always count the number of copies of the derived allele.
It hence is useful to permute alleles beforehand.
}
\seealso{
\code{\link{simulCoalescent}}, \code{\link{segSites2allDoses}}, \code{\link{haplosAlleles2num}}
}
\author{
Timothee Flutre
}
