% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{phasedJoinMapCP2qtl}
\alias{phasedJoinMapCP2qtl}
\title{Genotype coding}
\usage{
phasedJoinMapCP2qtl(x, verbose = 1)
}
\arguments{
\item{x}{data frame in the JoinMap format, for instance from \code{\link{genoClasses2JoinMap}}; any column named "clas" will be discarded from the start; columns named "seg" and "phase" should be present; a column named "locus" should be present unless row names are given; all other columns are supposed to correspond to genotypes}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
matrix in the "qtl" format
}
\description{
Convert genotypes encoded in the \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} format for cross type "CP" with known parental phases into the format used by the \href{https://cran.r-project.org/package=qtl}{qtl} package.
}
\seealso{
\code{\link{correspondenceJoinMap2qtl}}, \code{\link{updateJoinMap}}
}
\author{
Timothee Flutre
}
