% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plantTrialLmmFitCompSel}
\alias{plantTrialLmmFitCompSel}
\title{Model fit, comparison and selection}
\usage{
plantTrialLmmFitCompSel(glob.form, dat, part.comp.sel = "fixed",
  saved.file = "lmmFitCompSel.RData", nb.cores = 1, cl = NULL,
  verbose = 1)
}
\arguments{
\item{glob.form}{formula for the global model}

\item{dat}{data frame with all the data}

\item{part.comp.sel}{part(s) of the model (fixed and/or random); if only "fixed", the lm/lme4 and MuMIn packages will be used, otherwise the lmerTest package will be used}

\item{saved.file}{name of the file in which to save all model fits or from which to load all model fits; ignored with lmerTest}

\item{nb.cores}{number of cores; ignored with lmerTest}

\item{cl}{an object of class "cluster"; if NULL, will be created automaticall based on \code{nb.cores}; ignored with lmerTest}

\item{verbose}{verbosity level (0/1)}
}
\value{
list
}
\description{
For a plant field trial, fit by maximum likelihood (ML) a global linear (mixed) model with all the specified terms, then fit by ML various sub-models and compare them (AIC) or test each term (F-test for fixed effects and likelihood ratio test for random variables), finally select the best model and, if there are random variables, re-fit it using restricted maximum likelihood (ReML).
}
