% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{plot.plate}
\alias{plot.plate}
\title{Plot the arrangement of samples on a given plate.}
\usage{
\method{plot}{plate}(plate, main = "Plate")
}
\arguments{
\item{plate}{matrix}

\item{main}{string containing the text for the main title}
}
\value{
nothing
}
\description{
Plot the arrangement of samples on a given plate.
}
\author{
Timothée Flutre
}

