% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.scale}
\alias{plot.scale}
\title{Plot a scale, e.g. to add on the side of image()}
\usage{
\method{plot}{scale}(z, zlim, col = heat.colors(12), breaks, horiz = TRUE,
  ylim = NULL, xlim = NULL, ...)
}
\arguments{
\item{z}{}

\item{zlim}{}

\item{col}{}

\item{breaks}{}

\item{horiz}{}

\item{ylim}{}

\item{xlim}{}

\item{...}{}
}
\description{
Takes some time to draw (there is one polygon per break...)
http://menugget.blogspot.de/2011/08/adding-scale-to-image-plot.html
}
\author{
Timothée Flutre
}

