% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{plotAligns}
\alias{plotAligns}
\title{Plot alignments}
\usage{
plotAligns(coords, main = "Alignments", xlab = "reference", xlim = NULL,
  col = "black")
}
\arguments{
\item{coords}{data.frame with at least three columns, "start", "end" and "names"}

\item{main}{main title}

\item{xlab}{label of the x-axis}

\item{xlim}{x-axis limits}

\item{col}{segment color(s)}
}
\value{
nothing
}
\description{
Plot alignments of several queries on the same reference.
}
\examples{
\dontrun{coords <- data.frame(start=c(2, 21, 29, 50),
                     end=c(10, 25, 45, 53),
                     names=c("seq1", "seq2", "seq2", "seq3"),
                     stringsAsFactors=FALSE)
par(mar=c(4, 7, 3, 1))
plotAligns(coords, xlab="chr1", col=c(1, 1, 2, 1))
}
}
\seealso{
\code{\link{plotGRanges}}
}
\author{
Timothee Flutre
}
