% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{plotEllipsis}
\alias{plotEllipsis}
\title{Plot vcov as ellipsis}
\usage{
plotEllipsis(
  mat,
  center = c(x = 0, y = 0),
  main = "Variance-covariance",
  xlab = "first dimension",
  ylab = "second dimension",
  asp = 1,
  lwd = 1,
  col = "black",
  axisType = NULL,
  axisCols = c(major = "black", minor = "black")
)
}
\arguments{
\item{mat}{input square matrix}

\item{center}{vector with the x and y coordinates of the center (population mean)}

\item{main}{main title}

\item{xlab}{label of the x-axis}

\item{ylab}{label of the y-axis}

\item{asp}{aspect ratio}

\item{lwd}{line width for the ellipsis}

\item{col}{color for the ellipsis}

\item{axisType}{types of major and minor axes to be added to the plot (NULL/lines/arrows)}

\item{axisCols}{colors for the axes (if they are plotted)}
}
\value{
invisible matrix with ellipsis coordinates and intermediary values in the attributes
}
\description{
Plots a variance-covariance matrix as an ellipsis.
}
\examples{
## without correlation
center <- c(0, 0)
mat <- matrix(c(0.5,0,0,0.5), 2, 2)
plotEllipsis(mat, center)
plotEllipsis(mat, center, axisType="line")

## with positive correlation
mat <- matrix(c(0.5,0.45,0.45,0.5), 2, 2)
tmp <- plotEllipsis(mat, center)
attr(tmp, "rotAngle_deg")
plotEllipsis(mat, center=c(1,10))
plotEllipsis(mat, center, axisType="line")
plotEllipsis(mat, center, axisType="arrow")
plotEllipsis(mat, center, lwd=2, col="darkblue", axisType="line",
             axisCols=c("major"="blue","minor"="red"))

## with negative correlation
mat <- matrix(c(2,-0.7,-0.7,0.5), 2, 2)
tmp <- plotEllipsis(mat, center)
attr(tmp, "rotAngle_deg")
plotEllipsis(mat, center, axisType="arrow")
}
\seealso{
\code{getEllipsisCoords}
}
\author{
Timothee Flutre
}
