% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotFalconer}
\alias{plotFalconer}
\title{Falconer's parameterization}
\usage{
plotFalconer(a = 1, d = 0.75, f = 0.25)
}
\arguments{
\item{a}{additive effect (gene action) of allele A2}

\item{d}{dominance effect}

\item{f}{frequency of allele A2}
}
\value{
invisible list
}
\description{
Plot Falconer's parameterization for a bi-allelic locus with allele A1 and A2.
}
\author{
Timothee Flutre
}
