% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotGridMissGenos}
\alias{plotGridMissGenos}
\title{Missing genotypes}
\usage{
plotGridMissGenos(X, main = "Missing genotypes", xlab = "Individuals",
  ylab = "SNPs")
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns; NA if missing}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}
}
\value{
nothing
}
\description{
Plot missing SNP genotypes as a grid.
Data will be represented in black if missing, white otherwise.
}
\author{
Timothee Flutre
}

