% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotHaplosMatrix}
\alias{plotHaplosMatrix}
\title{Haplotypes}
\usage{
plotHaplosMatrix(haplos, main = "Haplotypes")
}
\arguments{
\item{haplos}{matrix of haplotypes with haplotypes (2 per individual) in rows and sites in columns}

\item{main}{main title}
}
\value{
nothing
}
\description{
Plot haplotypes as an image.
}
\author{
Timothee Flutre
}

