% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{plotHistPval}
\alias{plotHistPval}
\title{P values}
\usage{
plotHistPval(pvalues, freq = FALSE, main = NULL, pi0 = NULL)
}
\arguments{
\item{pvalues}{vector of raw p values (missing values will be omitted)}

\item{freq}{as in \code{\link[graphics]{hist}}; if TRUE, the histogram graphic is a representation of frequencies, the "counts" component of the result; if FALSE, probability densities, component "density", are plotted (so that the histogram has a total area of one)}

\item{main}{an overall title for the plot (default: "Density of <nb of non-NA p values> p values")}

\item{pi0}{estimate of the proportion of null hypotheses}
}
\value{
invisible output of \code{\link[graphics]{hist}}
}
\description{
Plot the histogram of p values with freq=FALSE to imitate figure 1 of Storey & Tibshirani (2003).
The x-axis goes from 0 to 1 with equidistant bins of width 0.05, and the y-axis is in the density scale.
For a given bin, the density of p values in this bin corresponds to the relative frequency in this bin divided by the bin width, where the relative frequency is the absolute frequency (i.e. raw counts in this bin) divided by the sum of all counts (i.e. over all bins).
In the density scale, the area of a given bin is equal to its density times the bin width, thus the area of the whole histogram is equal to 1.
Assuming all null hypotheses are true, the "density" histogram represents the density of the discrete version of the Uniform distribution, where the height of each bin is 1, so that the whole histogram area remains equal to 1.
}
\author{
Timothee Flutre
}

