% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{plotInfoVariantCalls}
\alias{plotInfoVariantCalls}
\title{Plot information on variant-level calls}
\usage{
plotInfoVariantCalls(x, main = "")
}
\arguments{
\item{x}{data frame, e.g. from "bcftools query --format '\%CHROM\\t\%POS\\t\%TYPE\\t...'"}

\item{main}{main title}
}
\value{
nothing
}
\description{
Plot histograms of QUAL, DP, BaseQRankSum, MQRankSum, ReadPosRankSum and ClippingRankSum over 2 rows and 3 columns.
}
\seealso{
\code{\link{infoVariantCalls}}
}
\author{
Timothee Flutre
}
