% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotLd}
\alias{plotLd}
\title{Pairwise linkage disequilibrium}
\usage{
plotLd(x, y, main, use.density = TRUE, xlab = "physical distance (bp)",
  ylab = "linkage disequilibrium (r2)", span = 1/10, degree = 1,
  evaluation = 50, sample.size = NULL)
}
\arguments{
\item{x}{vector of distances between SNPs (see \code{\link{distSnpPairs}})}

\item{y}{vector of LD estimates (see \code{\link{estimLd}})}

\item{main}{main title}

\item{use.density}{if TRUE, uses smoothScatter; otherwise, use scatter.smooth}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{span}{the parameter alpha which controls the degree of smoothing (see \code{\link[stats]{loess}})}

\item{degree}{the degree of the polynomials to be used (see \code{\link[stats]{loess}})}

\item{evaluation}{number of points at which to evaluate the smooth curve (see \code{\link[stats]{loess.smooth}})}

\item{sample.size}{nb of sampled haplotypes, n, used to estimate the pairwise LD; if not NULL, n is used to calculate the r value above which the null hypothesis "D=0" is rejected, where r = D / sqrt(f_A f_a f_B f_b) and X^2 = n r^2 is the test statistic asymptotically following a Chi2(df=1) (see McVean, Handbook of Statistical Genetics, 2007, and Pritchard and Przewoski, AJHG, 2001)}
}
\value{
nothing
}
\description{
Plots the linkage disequilibrium between pairs of SNPs, as a blue density or black points, with a red loess.
}
\author{
Timothee Flutre
}

