% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotLdSry}
\alias{plotLdSry}
\title{Pairwise linkage disequilibrium}
\usage{
plotLdSry(
  x,
  y,
  bin.len = 10^3,
  max.dist = 10^5,
  main = "",
  xlab = "Physical distance (bp)",
  ylab = "Summarized LD"
)
}
\arguments{
\item{x}{vector of distances between SNPs (see \code{\link{distSnpPairs}})}

\item{y}{vector of LD estimates (see \code{\link{estimLd}})}

\item{bin.len}{length of each bin}

\item{max.dist}{maximum distance between SNPs to consider}

\item{main}{main title}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}
}
\value{
invisible data frame with the summary per bin
}
\description{
Plots a summary of linkage disequilibrium between pairs of SNPs per non-overlapping bin (mean, median and quartiles 1 and 3).
}
\seealso{
\code{\link{estimLd}}, \code{\link{distSnpPairs}}, \code{\link{plotLd}}
}
\author{
Timothee Flutre
}
