% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grapevine.R
\name{plotMapSxGDomaineChapitre}
\alias{plotMapSxGDomaineChapitre}
\title{Plot the map of SxG}
\usage{
plotMapSxGDomaineChapitre(map, lang = "FR", show.plant.types = TRUE,
  col.block = setNames(c("grey81", "bisque1"), c("A", "B")))
}
\arguments{
\item{map}{data.frame with columns rank, location, x and y}

\item{lang}{language for the title and legend (default=FR/EN)}

\item{show.plant.types}{show the plants corresponding to F1, parents and controls}

\item{col.block}{vector with block colors (set to NULL to ignore)}
}
\value{
nothing
}
\description{
reciprocal cross Syrah x Grenache at the Domaine du Chapitre (near Montpellier, France)
}
\author{
Timothee Flutre
}

