% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{plotMcmcChain}
\alias{plotMcmcChain}
\title{MCMC diagnostics}
\usage{
plotMcmcChain(res.mcmc, param.name, subplots = 1:4, pe = NULL, hi = NULL)
}
\arguments{
\item{res.mcmc}{object of class "mcmc" (not "mcmc.list"!)}

\item{param.name}{parameter name in the columns of res.mcmc}

\item{subplots}{vector indicating which sub-plot(s) to make (1 for trace, 2 for autocorrelation, 3 for running quantiles, 4 for density)}

\item{pe}{point estimate (optional, but required if "hi" is given)}

\item{hi}{half-interval (optional)}
}
\value{
nothing
}
\description{
Plot the trace, autocorrelation, running average and density side by side for a single chain.
TODO: show dark background for burn-in in traceplot
}
\author{
Timothee Flutre
}
