% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{plotMummer}
\alias{plotMummer}
\title{MUMmer}
\usage{
plotMummer(coords, main = "MUMmer", xlim = NULL, col = c("black", "red"))
}
\arguments{
\item{coords}{data.frame (see \code{\link{loadMummer}}); should only contain a single reference}

\item{main}{main title}

\item{xlim}{x-axis limits}

\item{col}{segment color(s)}
}
\value{
nothing
}
\description{
Plot alignment coordinates from \href{http://mummer.sourceforge.net/}{MUMmer}.
}
\examples{
\dontrun{## MUMmer should have been executed beforehand
coords <- loadMummer(file.coords)
pdf("plot_mummer.pdf", width=7, height=10)
par(mar=c(4, 7, 3, 1))
plotMummer(coords, col=c(1,2))
dev.off()
}
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{loadMummer}}
}

