% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotPhyDistVsLdSry}
\alias{plotPhyDistVsLdSry}
\title{Pairwise linkage disequilibrium}
\usage{
plotPhyDistVsLdSry(
  x,
  y,
  xlab,
  ylab,
  main = "",
  fitted.loess = NULL,
  fitted.lm = NULL,
  fitted.loglm = NULL,
  fitted.asyreg = NULL,
  fitted.biexp = NULL,
  ...
)
}
\arguments{
\item{x}{vector of pysical coordinates of mid bins (see output of \code{\link{summarizeLd}})}

\item{y}{vector of LD estimates (see output of \code{\link{summarizeLd}})}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{main}{main title}

\item{fitted.loess}{see the output of \code{\link{fitPhyDistVsLd}}}

\item{fitted.lm}{see the output of \code{\link{fitPhyDistVsLd}}}

\item{fitted.loglm}{see the output of \code{\link{fitPhyDistVsLd}}}

\item{fitted.asyreg}{see the output of \code{\link{fitPhyDistVsLd}}}

\item{fitted.biexp}{see the output of \code{\link{fitPhyDistVsLd}}}

\item{...}{arguments passed to \code{\link{plot}}}
}
\value{
nothing
}
\description{
Plots a summary of linkage disequilibrium between pairs of SNPs versus physical distance.
}
\seealso{
\code{\link{summarizeLd}}, \code{\link{fitPhyDistVsLd}}
}
\author{
Timothee Flutre
}
