% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotPhyVsGenDistances}
\alias{plotPhyVsGenDistances}
\title{Plot physical versus genetic distances}
\usage{
plotPhyVsGenDistances(x, chrs = NULL, type = "b", las = 1,
  xlab = "physical distance (in bp)", ylab = "genetic distance (in cM)",
  ...)
}
\arguments{
\item{x}{data frame with markers in rows and at least three columns named "chr", "physical.distance", "genetic.distance"}

\item{chrs}{chromosome(s) to plot (all if NULL); will be sorted by mixedsort if the "gtools" package is available; if several chromosomes are to be plotted, there will be one plot per chromosome, with two plots per row as specified to \code{\link[graphics]{par}}}

\item{type}{type of plot to draw (see \code{\link[graphics]{plot}})}

\item{las}{style of axis labels (see \code{\link[graphics]{par}})}

\item{xlab}{title for the x axis}

\item{ylab}{title for the y axis}

\item{...}{arguments to be passed to \code{\link[graphics]{plot}}, except "x" (will be the physical distance), "y" (will be the genetic distance), "main" (will be the chromosome identifier), "type", "las", "xlab" and "ylab"}
}
\value{
nothing
}
\description{
Plots physical versus genetic distances
}
\author{
Timothee Flutre
}
