% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{plotPlate}
\alias{plotPlate}
\title{Plot plate}
\usage{
plotPlate(plate, main = "Plate")
}
\arguments{
\item{plate}{matrix}

\item{main}{string containing the text for the main title}
}
\value{
nothing
}
\description{
Plot the arrangement of samples on the given plate.
}
\author{
Timothee Flutre
}

