% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotResidualsBtwYears}
\alias{plotResidualsBtwYears}
\title{Plot residuals between years}
\usage{
plotResidualsBtwYears(
  df,
  colname.res,
  years,
  cols.uniq.id,
  blocks = NULL,
  las = 1,
  lgd.pos = "topright",
  ...
)
}
\arguments{
\item{df}{data frame with a column given by \code{colname.res}, a column named "year", a column named "block" if \code{blocks} is specified, as well as any other column specified in \code{cols.uniq.id}}

\item{colname.res}{name of the column containing the residuals}

\item{years}{vector with two years as character}

\item{cols.uniq.id}{vector of column name(s) allowing to identify each plant and pair it between years or years and blocks, e.g. \code{c("geno","block","rank","location")} or \code{"geno"}}

\item{blocks}{if not NULL, vector of two blocks (e.g. \code{c("A","A")} or \code{c("A","B")})}

\item{las}{style of axis labels, see \code{\link[graphics]{par}}}

\item{lgd.pos}{position of the legend}

\item{...}{arguments passed on to \code{\link[graphics]{plot}}, such as \code{main}}
}
\value{
invisible data frame of the data used to make the plot
}
\description{
Plot residuals between years to check their temporal independence.
Especially useful as a diagnostic of model fit for statistical analysis of perennial plants.
}
\author{
Timothee Flutre
}
