% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{plotTukey}
\alias{plotTukey}
\title{Tukey plot}
\usage{
plotTukey(mat, title = NULL)
}
\arguments{
\item{mat}{square matrix with}

\item{title}{optional title}
}
\value{
nothing
}
\description{
Plots significant pairwise differences from Tukey HSD test as proposed in \href{https://acsess.onlinelibrary.wiley.com/doi/10.2134/agronj2017.10.0580}{Piepho (2018)}.
}
\examples{
## From Piepho (2018):
n <- 5
(tab1 <- data.frame(farmID=LETTERS[1:n],
                    mlc=c(1.86, 2.34, 1.82, 2.18, 1.38),
                    group=c("ab", "a", "ab", "a", "b")))
vec1 <- setNames(tab1$mlc, tab1$farmID)
vec1 <- sort(vec1, decreasing=TRUE)
(sortedFarmIDs <- names(vec1))
tab3 <- matrix(nrow=n, ncol=n,
               dimnames=list(names(vec1), names(vec1)))
tab3[1, 2:n] <- c(0.16, 0.48, 0.52, 0.96)
tab3[2, 3:n] <- c(0.32, 0.36, 0.80)
tab3[3, 4:n] <- c(0.04, 0.48)
tab3[4, n] <- c(.44)
tab3
tab3_notSignif <- tab3
tab3_notSignif["B","E"] <- NA
tab3_notSignif["D","E"] <- NA
tab3_notSignif
plotTukey(tab3_notSignif, title="Fig. I from Piepho (2018)")
}
\seealso{
\code{\link{simplifTukey}}
}
\author{
Timothee Flutre
}
