% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{plotVcfPercNa}
\alias{plotVcfPercNa}
\title{Plot VCF}
\usage{
plotVcfPercNa(
  vcf.file,
  yieldSize = NA_integer_,
  dict.file = NULL,
  seq.id = NULL,
  seq.start = NULL,
  seq.end = NULL,
  main = "Density of missing genotypes",
  plot.it = TRUE
)
}
\arguments{
\item{vcf.file}{path to the VCF file (if the bgzip index doesn't exist in the same directory, it will be created)}

\item{yieldSize}{number of records to yield each time the file is read from (see \code{?TabixFile}) if seq.id is NULL}

\item{dict.file}{path to the SAM dict file (see \url{https://broadinstitute.github.io/picard/command-line-overview.html#CreateSequenceDictionary}) if seq.id is specified with no start/end}

\item{seq.id}{sequence identifier to work on (e.g. "chr2")}

\item{seq.start}{start of the sequence to work on (if NULL, whole seq)}

\item{seq.end}{end of the sequence to work on (if NULL, whole seq)}

\item{main}{main title of the plot}

\item{plot.it}{if TRUE, \code{autoplot} from the ggbio package will be used}
}
\value{
GenomicRange
}
\description{
Plot markers from a VCF file along chromosomes, colored according to their percentage of missing genotypes.
}
\author{
Timothee Flutre
}
