% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotWithScale}
\alias{plotWithScale}
\title{Plot a scale, e.g. to add on the side of image()}
\usage{
plotWithScale(
  z,
  zlim,
  col = grDevices::heat.colors(12),
  breaks,
  horiz = TRUE,
  ylim = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{z}{vector}

\item{zlim}{lim}

\item{col}{color}

\item{breaks}{vector}

\item{horiz}{boolean}

\item{ylim}{lim}

\item{xlim}{lim}

\item{...}{arguments to be passed to plot()}
}
\description{
Takes some time to draw (there is one polygon per break...)
http://menugget.blogspot.de/2011/08/adding-scale-to-image-plot.html
}
\author{
Timothee Flutre
}
