% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pooledVar}
\alias{pooledVar}
\title{Pooled variance}
\usage{
pooledVar(v, n)
}
\arguments{
\item{v}{vector of sampled variances}

\item{n}{vector of sample sizes}
}
\value{
numeric
}
\description{
Returns the pooled variance.
}
\author{
Timothee Flutre
}
