% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{powerWX2019}
\alias{powerWX2019}
\title{GWAS power analysis}
\usage{
powerWX2019(n = 500, h2 = 0.05, lambda = 1, rho = 0.5, m = 1e+05, alpha = 0.05)
}
\arguments{
\item{n}{actual population size}

\item{h2}{heritability}

\item{lambda}{lambda}

\item{rho}{effective correlation coefficient}

\item{m}{number of markers}

\item{alpha}{alpha}
}
\value{
numeric
}
\description{
Useful for the power analysis of a GWAS from \href{http://www.nature.com/articles/s41437-019-0205-3}{Wang and Xu (2019)}.
Code from \href{https://onlinelibrary.wiley.com/doi/10.1002/csc2.20692}{White et al (2020)}.
}
\author{
Jon White
}
