% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{precMatAR1}
\alias{precMatAR1}
\title{AR(1)}
\usage{
precMatAR1(n, rho, sigma2)
}
\arguments{
\item{n}{dimension of the matrix (number of rows and columns)}

\item{rho}{correlation between successive variables}

\item{sigma2}{variance of the errors (also called "innovations" in the time-series literature)}
}
\value{
sparse matrix
}
\description{
Return a first-order auto-regressive precision matrix, as in section 1.1.1 of Rue and Held (2005).
}
\author{
Timothee Flutre
}
