% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prettyPrintBetterSummary}
\alias{prettyPrintBetterSummary}
\title{Print better summary}
\usage{
prettyPrintBetterSummary(x = NULL, smy = NULL, spec = "\%.2f",
  header = NULL)
}
\arguments{
\item{x}{vector of numbers; if NULL, \code{smy} should be provided}

\item{smy}{output of \code{\link{betterSummary}}; if NULL, \code{x} should be provided; if data frame, will be converted to vector}

\item{spec}{specifier to format floats, see \code{\link{sprintf}}}

\item{header}{optional character to be placed in front of what will be printed}
}
\value{
invisible summary
}
\description{
Print the output of \code{\link{betterSummary}} in a single line.
}
\author{
Timothee Flutre
}
