% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{prettyPrintSummary}
\alias{prettyPrintSummary}
\title{Summary}
\usage{
prettyPrintSummary(x, spec = "\%.2f")
}
\arguments{
\item{x}{vector of numbers}

\item{spec}{specifier, see \code{\link{sprintf}}}
}
\value{
nothing
}
\description{
Print the output of \code{\link{summary}} in a single line.
}
\author{
Timothee Flutre
}

