% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{pruneSnpsLd}
\alias{pruneSnpsLd}
\title{Prune SNPs based on LD}
\usage{
pruneSnpsLd(X = NULL, snp.coords = NULL, gds = NULL, ld.threshold = 0.2,
  remove.monosnp = TRUE, maf = NaN, missing.rate = NaN,
  method = c("composite", "r", "dprime", "corr"), slide.max.bp = 5e+05,
  slide.max.n = NA, seed = NULL, nb.cores = 1, verbose = 1)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded, for each SNP, in number of copies of its second allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; the "second" allele is arbitrary, it can correspond to the minor (least frequent) or the major (most frequent) allele; will be transformed into a "gds" object (see next argument)}

\item{snp.coords}{data frame which row names are SNP identifiers, the first column should contain chromosomes as integers, and the second column should contain positions; compulsory if the X argument is specified}

\item{gds}{object of class "SNPGDSFileClass" from the SNPRelate package}

\item{ld.threshold}{the LD threshold below which SNPs are discarded; will be passed to "snpgdsLDpruning"}

\item{remove.monosnp}{will be passed to "snpgdsLDpruning"}

\item{maf}{will be passed to "snpgdsLDpruning"}

\item{missing.rate}{will be passed to "snpgdsLDpruning"}

\item{method}{will be passed to "snpgdsLDpruning"}

\item{slide.max.bp}{will be passed to "snpgdsLDpruning"}

\item{slide.max.n}{will be passed to "snpgdsLDpruning"}

\item{seed}{seed for the pseudo-random number generator}

\item{nb.cores}{will be passed to "snpgdsLDpruning"}

\item{verbose}{verbosity level (0/1)}
}
\value{
vector of SNP identifiers
}
\description{
Prune SNPs based on their pairwise linkage disequilibriums via sliding windows using the "snpgdsLDpruning" function from the "SNPRelate" package..
}
\seealso{
\code{\link{thinSnps}}, \code{\link[SNPRelate]{snpgdsLDpruning}}
}
\author{
Timothee Flutre
}
