% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pseudoR2}
\alias{pseudoR2}
\title{Pseudo R squared}
\usage{
pseudoR2(dat, res, pred, method = "Efron")
}
\arguments{
\item{dat}{N-vector of data}

\item{res}{N-vector of residuals}

\item{pred}{N-vector of predictions}

\item{method}{method}
}
\value{
pseudo R squared
}
\description{
Compute a \href{https://stats.idre.ucla.edu/other/mult-pkg/faq/general/faq-what-are-pseudo-r-squareds/}{pseudo R squared}.
}
\author{
Timothee Flutre
}
