% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{pve2beta}
\alias{pve2beta}
\title{Proportion of variance explained}
\usage{
pve2beta(pve = 0.7, n = 200, maf = 0.3, sigma = 1)
}
\arguments{
\item{pve}{proportion of variance explained}

\item{n}{sample size}

\item{maf}{minor allele frequency}

\item{sigma}{error standard deviation}
}
\value{
numeric
}
\description{
For the model "y_i = mu + beta * x_i + epsilon_i", compute the additive effect of a bi-allelic SNP (beta) given its PVE, MAF and error standard deviation.
}
\author{
Timothee Flutre
}
