% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{qtlrelPerChr}
\alias{qtlrelPerChr}
\title{QTLRel per chromosome}
\usage{
qtlrelPerChr(y, X, snp.coords, chr.ids = NULL, W = NULL, Z = NULL,
  verbose = 0)
}
\arguments{
\item{y}{vector of phenotypes}

\item{X}{matrix of SNP genotypes encoded as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"}

\item{chr.ids}{set of chromosome identifiers to analyze (optional, all by default)}

\item{W}{incidence matrix of covariates (should not contain the column of 1's for the intercept)}

\item{Z}{incidence matrix relating phenotypes to individuals (if nrow(y) and nrow(X) are different; diagonal otherwise)}

\item{verbose}{verbosity level (default=1)}
}
\value{
a list with two data.frames as components, vc and scan
}
\description{
For each SNP p, the likelihood is y = W alpha + Z x_p beta_p + Z u + epsilon where all x_p are columns of X, and one want to test the null hypothesis "beta = 0". See Cheng et al (BMC Genetics, 2011).
}
\author{
Timothee Flutre
}

