% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{quant.norm}
\alias{quant.norm}
\title{Quantile-normalize a vector of numbers to a standard normal distribution.}
\usage{
quant.norm(x, break.ties.rand = TRUE, seed = 1859)
}
\arguments{
\item{x}{vector of numeric data}

\item{break.ties.rand}{break ties randomly (default=TRUE)}

\item{seed}{see for the pseudo-random number generator (default=1859)}
}
\value{
vector
}
\description{
TODO: add ref
}
\author{
Timothée Flutre
}

