% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{quantilesBinnedSnpData}
\alias{quantilesBinnedSnpData}
\title{Quantiles of allelic R2 after binning}
\usage{
quantilesBinnedSnpData(
  snp.data,
  snp.bins,
  quant.probs = seq(0, 1, 0.25),
  verbose = 1
)
}
\arguments{
\item{snp.data}{numeric vector of SNP data (one value per SNP) with SNP identifiers as names}

\item{snp.bins}{factor as returned by \code{\link{cut}} applied to SNP data (percentage of missing data, minor allele frequencies, etc) with SNP identifiers as names}

\item{quant.probs}{numeric vector of probabilities with values in [0,1] to be used with \code{\link[stats]{quantile}}}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with the quantiles per bin, as well as the midpoint of each bin
}
\description{
Calculate the quantiles of allelic R2 after binning.
}
\author{
Timothee Flutre
}
