% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{quantNorm}
\alias{quantNorm}
\title{Quantile normalization}
\usage{
quantNorm(x, break.ties.rand = TRUE, seed = 1859)
}
\arguments{
\item{x}{vector of numeric data}

\item{break.ties.rand}{break ties randomly (default=TRUE)}

\item{seed}{see for the pseudo-random number generator (default=1859)}
}
\value{
vector
}
\description{
Quantile-normalizes a vector of numbers to a standard normal distribution as described in \href{http://dx.doi.org/10.1093/bioinformatics/19.2.185}{Bolstad et al (2003)}.
}
\author{
Timothee Flutre
}
