% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{quass}
\alias{quass}
\title{Compute the inverse of the additive relationship matrix using Quass algorithm}
\usage{
quass(s, d)
}
\arguments{
\item{s}{vector of sires}

\item{d}{vector of dams}
}
\value{
matrix
}
\description{
Given the vectors of sire and dam, directly return inverse of additive relationship matrix 'A' without creating the 'A' itself.
This is a modification of Henderson's method and unlike createAinv.r, this can be used in inbred populations.
Reference:
\itemize{
\item Quass, R. L. 1976. Computing the Diagonal Elements and Inverse of a Large Numerator Relationship Matrix. Biometrics 32:949-953.
}
}
\note{
Unknown parents should be coded as zero. Last modified by Morota on April 2, 2010.
}
\author{
Gota Morota
}
