% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{rand.plate}
\alias{rand.plate}
\title{Randomize plate}
\usage{
rand.plate(plate, rand.scheme = "1x96", seed = NULL, file = NULL)
}
\arguments{
\item{plate}{matrix}

\item{rand.scheme}{string}

\item{seed}{integer}

\item{file}{path to file in which to write the randomised plate}
}
\value{
matrix
}
\description{
Randomize the given plate.
}
\author{
Timothee Flutre
}

