% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{read.biomercator}
\alias{read.biomercator}
\title{Returns the genetic map contained in a BioMercator TXT file.}
\usage{
read.biomercator(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from}
}
\value{
list
}
\description{
http://moulon.inra.fr/index.php/en/tranverse-team/atelier-de-bioinformatique/projects/projets/135
}
\author{
Timothee Flutre
}

