% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{readBcftoolsCounts}
\alias{readBcftoolsCounts}
\title{Read output from bcftools counts}
\usage{
readBcftoolsCounts(files)
}
\arguments{
\item{files}{vector of file names to read}
}
\value{
data frame
}
\description{
Reads the output from "bcftools plugin counts"
}
\author{
Timothee Flutre
}
