% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_contest.R
\name{readCheckBreedDataFile}
\alias{readCheckBreedDataFile}
\title{Read for breeding contest}
\usage{
readCheckBreedDataFile(f = NULL, df = NULL, max.nb.plots, subset.snps,
  max.nb.inds = 1000)
}
\arguments{
\item{f}{path to the input file}

\item{df}{data.frame (if the file was already read)}

\item{max.nb.plots}{maximum number of plots}

\item{subset.snps}{list with two components named "ld" and "hd" containing vector of SNP identifiers}

\item{max.nb.inds}{maximum number of unique individuals for which at least one request is made}
}
\value{
invisible data.frame
}
\description{
Read and check a file supposed to contain requests about pheno/geno data.
}
\author{
Timothee Flutre
}

