% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{readCheckBreedPlantFile}
\alias{readCheckBreedPlantFile}
\title{Read for breeding game}
\usage{
readCheckBreedPlantFile(f = NULL, df = NULL, max.nb.hd = 800)
}
\arguments{
\item{f}{path to the input file (columns should be separated by a tabulation)}

\item{df}{data.frame (if the file was already read)}

\item{max.nb.hd}{maximum number of haplodiploidization to request in a single file}
}
\value{
invisible data.frame
}
\description{
Read and check a file supposed to contain requests about plant material.
It should have 3 columns named \code{parent1}, \code{parent2} and \code{child}.
}
\seealso{
\code{\link{makeExamplePlantFile}}
}
\author{
Timothee Flutre
}
