% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{readGenoDoseFileFromBcftools}
\alias{readGenoDoseFileFromBcftools}
\title{Read SNP genotypes as dosage from bcftools}
\usage{
readGenoDoseFileFromBcftools(genos.file, get.coords = TRUE,
  get.alleles = TRUE, use.fread = TRUE)
}
\arguments{
\item{genos.file}{path to the input file containing the SNP genotype data}

\item{get.coords}{if TRUE, SNP coordinates will also be extracted}

\item{get.alleles}{if TRUE, SNP alleles will also be extracted}

\item{use.fread}{if TRUE, \code{\link[data.table]{fread}} will be used to speed-up}
}
\value{
list with a matrix of SNP genotypes and optional data frames of SNP coordinates and alleles
}
\description{
Read an input file with genotypes as allele dosage as written by \code{bcftools +dosage genos.vcf.gz | tr ' ' '\t' | datamash transpose | gzip > genos_dosage.tsv.gz}.
Caution, depending on the file size, this may require the allocation of too much RAM for your system.
}
\seealso{
\code{\link{convertVcfToGenoDoseWithBcftools}}
}
\author{
Timothee Flutre
}
