% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPHASE.R
\name{readGenosFastphase}
\alias{readGenosFastphase}
\title{Read genotypes from fastPHASE}
\usage{
readGenosFastphase(file, snp.ids)
}
\arguments{
\item{file}{character}

\item{snp.ids}{vector of SNP identifiers}
}
\value{
matrix with genotypes in rows (2 per genotype) and SNPs in columns
}
\description{
Extract the genotypes from the output file of fastPHASE.
}
\author{
Timothee Flutre
}
