% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPHASE.R
\name{readHaplosFastphase}
\alias{readHaplosFastphase}
\title{Read haplotypes from fastPHASE}
\usage{
readHaplosFastphase(file, snp.ids)
}
\arguments{
\item{file}{character}

\item{snp.ids}{vector of SNP identifiers}
}
\value{
matrix with haplotypes in rows (2 per genotype) and SNPs in columns
}
\description{
Extract the haplotypes from fastPHASE's output files ("hapguess_switch" and "hapguess_indiv").
}
\author{
Timothee Flutre
}
