% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FImpute.R
\name{readOutputsFimpute}
\alias{readOutputsFimpute}
\title{Read outputs from FImpute}
\usage{
readOutputsFimpute(out.dir)
}
\arguments{
\item{out.dir}{directory in which the output files are saved
Read output files from FImpute.}
}
\value{
nothing
}
\description{
Read outputs from FImpute
}
\author{
Timothee Flutre
}
