% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findhap.R
\name{readOutputsFindhap}
\alias{readOutputsFindhap}
\title{Read outputs from findhap}
\usage{
readOutputsFindhap(out.dir)
}
\arguments{
\item{out.dir}{directory in which the output files are saved}
}
\value{
list
}
\description{
Read output files from findhap.
}
\author{
Timothee Flutre
}
