% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{readOutSTRUCTURE}
\alias{readOutSTRUCTURE}
\title{STRUCTURE}
\usage{
readOutSTRUCTURE(out.file)
}
\arguments{
\item{out.file}{path to the output file}
}
\value{
list
}
\description{
Read the output file from STRUCTURE.
}
\author{
Timothee Flutre
}
