% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{readPlinkMendel}
\alias{readPlinkMendel}
\title{Read PLINK}
\usage{
readPlinkMendel(prefix, suffix = "", verbose = 1)
}
\arguments{
\item{prefix}{prefix of the file to read}

\item{suffix}{suffix of the file to read (mendel/imendel/fmendel/lmendel)}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame
}
\description{
Read the output files of \code{plink --mendel}.
}
\examples{
\dontrun{## run PLINK in the command-line
cmd <- "plink --mendel --bed family.bed --bim family.bim --fam family.fam --out family_plink"
system(cmd)

pl.mend <- readPlinkMendel(prefix="family_plink", suffix="mendel")
str(pl.mend)

pl.imend <- readPlinkMendel(prefix="family_plink", suffix="imendel")
head(pl.imend[order(pl.imend$N, decreasing=TRUE),])
}
}
\author{
Timothee Flutre
}
