% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{readSamDict}
\alias{readSamDict}
\title{Read SAM dict}
\usage{
readSamDict(file)
}
\arguments{
\item{file}{the name of the ".dict" file which the data are to be read from}
}
\value{
data.frame
}
\description{
Reads a file in the "dict" format. See \url{https://broadinstitute.github.io/picard/command-line-overview.html#CreateSequenceDictionary}.
}
\author{
Timothee Flutre
}

