% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{readSegregJoinMap}
\alias{readSegregJoinMap}
\title{Read genotypes for JoinMap/MapQTL}
\usage{
readSegregJoinMap(file, na.string = "--", verbose = 1)
}
\arguments{
\item{file}{name of the file from which the data will be read (can be compressed with gzip)}

\item{na.string}{a character to be interpreted as NA values}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame in the JoinMap format
}
\description{
Read genotype data from a file in the \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} format ("loc").
}
\note{
the code is heavily inspired from the "read.cross.mq.loc" function in the "qtl" package, which I also wrote
}
\seealso{
\code{\link{writeSegregJoinMap}}
}
\author{
Timothee Flutre
}
