% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{readVcfSubset}
\alias{readVcfSubset}
\title{Read VCF}
\usage{
readVcfSubset(vcf.file, genome = "", seq.id, seq.start, seq.end)
}
\arguments{
\item{vcf.file}{path to the VCF file (bgzip index should exist in same directory)}

\item{genome}{genome identifier (e.g. "VITVI_12x2")}

\item{seq.id}{sequence identifier to work on (e.g. "chr2")}

\item{seq.start}{start of the sequence to work on (1-based coordinate)}

\item{seq.end}{end of the sequence to work on (1-based coordinate)}
}
\value{
CollapsedVCF (see pkg \href{http://bioconductor.org/packages/VariantAnnotation/}{VariantAnnotation})
}
\description{
Read a subset of a VCF file
}
\author{
Timothee Flutre
}

