% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{rearrangeInputsForAssoGenet}
\alias{rearrangeInputsForAssoGenet}
\title{Plant Association genetics}
\usage{
rearrangeInputsForAssoGenet(ids, y, X, snp.coords, alleles, verbose = 1)
}
\arguments{
\item{ids}{data.frame of identifiers, with (at least) column names \code{cultivar.code} and \code{accession.code}; the outputs will be sorted according to this option}

\item{y}{vector, matrix or data.frame which row names should be present in \code{ids$cultivar.code} or \code{ids$accession.code}}

\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns; row names should be present in \code{ids$accession.code} or \code{ids$cultivar.code}, and column names in \code{rownames(snp.coords)}}

\item{snp.coords}{data.frame with 2 columns \code{coord} and \code{chr}, and SNP identifiers as row names}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (named "minor" and "major")}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with inputs after subsetting and sorting
}
\description{
Subset and sort inputs necessary to perform an analysis of plant association genetics, that is, the subset of cultivars with genotypes and phenotypes, and the subset of markers having genotypes and coordinates.
}
\author{
Timothee Flutre
}

