% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeFileExtension}
\alias{removeFileExtension}
\title{File}
\usage{
removeFileExtension(file, fileext)
}
\arguments{
\item{file}{character (will be passed to \code{\link{strsplit}})}

\item{fileext}{character containing a regular expression (will be passed to \code{\link{strsplit}})}
}
\value{
character
}
\description{
Remove extension(s) from a file name.
}
\examples{
f <- "data.txt.gz"
removeFileExtension(file=f, fileext="\\\\.gz")
removeFileExtension(file=f, fileext="\\\\.txt\\\\.gz")
}
\author{
Timothee Flutre
}
