% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{renameVcfSamples}
\alias{renameVcfSamples}
\title{Rename VCF samples}
\usage{
renameVcfSamples(
  in.vcf.file,
  samples,
  out.vcf.file,
  skip.check = FALSE,
  sort.variants = TRUE
)
}
\arguments{
\item{in.vcf.file}{path to the input VCF file (will be automatically indexed if necessary)}

\item{samples}{data frame with (at least) two columns, the first corresponding to the old names (same order as in the VCF file) and the second to the new names; spaces inside names are forbidden}

\item{out.vcf.file}{path to the output VCF file}

\item{skip.check}{if TRUE, no comparison is done to check that \code{samples} is coherent with the data in the VCF file}

\item{sort.variants}{if TRUE, variants are also sorted, with \code{bcftools sort} (requires bcftools >= 1.6)}
}
\value{
invisible \code{out.vcf.file}
}
\description{
Rename samples in a VCF file with \code{bcftools reheader}.
Variants can also be sorted.
}
\author{
Timothee Flutre
}
