% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{requireNamespaces}
\alias{requireNamespaces}
\title{Namespaces}
\usage{
requireNamespaces(packages)
}
\arguments{
\item{packages}{vector of package names}
}
\value{
nothing
}
\description{
Stops if any package's namespace can't be loaded.
}
\author{
Timothee Flutre
}

