% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rhoWX2019}
\alias{rhoWX2019}
\title{GWAS power analysis}
\usage{
rhoWX2019(n = 500, lambda = 1, n0 = 500)
}
\arguments{
\item{n}{actual population size}

\item{lambda}{lambda}

\item{n0}{effective population size}
}
\value{
numeric
}
\description{
Computes the effective correlation coefficient from the sample.
Useful for the power analysis of a GWAS from \href{http://www.nature.com/articles/s41437-019-0205-3}{Wang and Xu (2019)}.
Code from \href{https://onlinelibrary.wiley.com/doi/10.1002/csc2.20692}{White et al (2020)}.
}
\author{
Jon White
}
