% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmatnorm}
\alias{rmatnorm}
\title{Matrix-variate Normal distribution}
\usage{
rmatnorm(n = 1, M, U, V, pivot = c(U = FALSE, V = FALSE))
}
\arguments{
\item{n}{number of observations}

\item{M}{mean matrix}

\item{U}{between-row covariance matrix}

\item{V}{between-column covariance matrix}

\item{pivot}{use pivoting for Choleski decomposition of U and V (see \code{\link[base]{chol}}); useful when U and/or V are singular}
}
\value{
array
}
\description{
Random generation for the matrix-variate Normal distribution.
See \url{https://en.wikipedia.org/wiki/Matrix_normal_distribution}.
}
\author{
Timothee Flutre
}

