% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(error)
}
\arguments{
\item{error}{vector \eqn{\hat{\theta}_i - \theta_i}}
}
\value{
numeric
}
\description{
Returns the root mean squared error (RMSE).
}
\author{
Timothee Flutre
}
