% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmsre}
\alias{rmsre}
\title{Root mean squared relative error}
\usage{
rmsre(rel_error)
}
\arguments{
\item{rel_error}{vector \eqn{(\hat{\theta}_i - \theta_i) / \theta_i}}
}
\value{
numeric
}
\description{
Returns the root mean squared relative error (RMSRE) as defined in \href{https://doi.org/10.3354/cr030079}{Willmott and Matsuura (2005)}.
}
\seealso{
\code{\link{rmse}}, \code{\link{rrmse}}, \code{\link{mae}}
}
\author{
Timothee Flutre
}
