% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{rootfreeChol}
\alias{rootfreeChol}
\title{LDL' decomposition}
\usage{
rootfreeChol(x)
}
\arguments{
\item{x}{matrix}
}
\value{
matrix
}
\description{
This wrapper around \code{\link{chol}} factors a square, symmetric, positive (semi-)definite matrix into the product of a lower triangular matrix (L), a diagonal matrix (D), and the transpose of the lower triangular matrix (U=L').
The value returned is a lower triangular matrix with the elements of D on the diagonal.
This decomposition is notably used in quantitative genetics (Misztal and Perez-Enciso, 1993).
}
\seealso{
\code{\link{getSparseInv}}
}
