% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rrmse}
\alias{rrmse}
\title{Relative root mean squared error}
\usage{
rrmse(error, theta, perc = TRUE)
}
\arguments{
\item{error}{vector \eqn{(\hat{\theta}_i - \theta_i)}}

\item{theta}{vector \eqn{\theta_i}}

\item{perc}{if TRUE, the return value will be a percentage}
}
\value{
numeric
}
\description{
Returns the relative root mean squared error (RRMSE) as defined in \href{https://doi.org/10.3354/cr030079}{Willmott and Matsuura (2005)}.
}
\seealso{
\code{\link{rmse}}, \code{\link{rmsre}}
}
\author{
Timothee Flutre
}
