% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sampleWX2019}
\alias{sampleWX2019}
\title{GWAS power analysis}
\usage{
sampleWX2019(
  h2 = 0.05,
  lambda = 1,
  rho = 0.5,
  power = 0.85,
  m = 1e+05,
  alpha = 0.05
)
}
\arguments{
\item{h2}{QTL heritability}

\item{lambda}{lambda}

\item{rho}{effective correlation coefficient}

\item{power}{power}

\item{m}{number of markers}

\item{alpha}{alpha}
}
\value{
numeric
}
\description{
Computes the minimum sample for the desired QTL heritability.
Useful for the power analysis of a GWAS from \href{http://www.nature.com/articles/s41437-019-0205-3}{Wang and Xu (2019)}.
Code from \href{https://onlinelibrary.wiley.com/doi/10.1002/csc2.20692}{White et al (2020)}.
}
\author{
Jon White
}
