% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{seg.sites2all.doses}
\alias{seg.sites2all.doses}
\title{Convert the SFS of independent replicates into a matrix of allele dosage.}
\usage{
seg.sites2all.doses(seg.sites)
}
\arguments{
\item{seg.sites}{list returned by scrm()}
}
\value{
matrix with diploid individuals in rows and SNPs in columns
}
\description{
SFS stands for site frequency spectrum
}
\author{
Timothée Flutre
}

