% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{seg.sites2snp.coords}
\alias{seg.sites2snp.coords}
\title{Make a data.frame of SNP coordinates from the SFS of independent replicates}
\usage{
seg.sites2snp.coords(seg.sites, snp.ids)
}
\arguments{
\item{seg.sites}{list returned by scrm()}

\item{snp.ids}{vector of identifiers (one per SNP)}
}
\value{
data.frame with SNPs in rows and 2 columns (chr, pos)
}
\description{
SFS stands for site frequency spectrum
}
\author{
Timothee Flutre
}

