% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{setJoinMapPhasesFromParentalLinkGroups}
\alias{setJoinMapPhasesFromParentalLinkGroups}
\title{Genotype coding}
\usage{
setJoinMapPhasesFromParentalLinkGroups(x, lg.par1, lg.par2)
}
\arguments{
\item{x}{data frame in the JoinMap format, for instance from \code{\link{genoClasses2JoinMap}}; row names should contain locus names; the first column should be "seg"; any column(s) "phase" or "clas" already existing will be discarded; other columns should contain genotype data}

\item{lg.par1}{linkage groups of the first parent (usually the mother) as a data frame with one row per marker and at least two columns named "linkage.group" and "locus"; "mirror" markers should have suffix "_m" as done by \code{\link{joinMap2backcross}}}

\item{lg.par2}{linkage groups of the second parent (usually the father) as a data frame with one row per marker and at least two columns named "linkage.group" and "locus"; "mirror" markers should have suffix "_m" as done by \code{\link{joinMap2backcross}}}
}
\value{
data frame in the JoinMap format with a "phase" column
}
\description{
Set linkage phases in the \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} format from two sets of parental linkage groups.
This function is tested only for segregation types <hkxhk>, <lmxll> and <nnxnp>, that is, in the case of bi-allelic SNPs segregating in a cross of outbred parents.
}
\author{
Timothee Flutre
}
