% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_contest.R
\name{setUpBreedingContest}
\alias{setUpBreedingContest}
\title{Set up}
\usage{
setUpBreedingContest(root.dir, shared.dir = NULL, nb.breeders = 3,
  lang = "fr")
}
\arguments{
\item{root.dir}{path to the root directory}

\item{shared.dir}{path to the shared directory (e.g. via Dropbox; root.dir if NULL)}

\item{nb.breeders}{number of breeders (will be named "breeder<i>"; a "test" will also be created)}

\item{lang}{language to be used (default=fr/en)}
}
\value{
of paths
}
\description{
Set up the directories and names for the breeding contest.
}
\author{
Timothee Flutre
}

