% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{setUpBreedingGame}
\alias{setUpBreedingGame}
\title{Set up breeding game}
\usage{
setUpBreedingGame(root.dir, shared.dir = NULL, nb.breeders = 3,
  lang = "fr")
}
\arguments{
\item{root.dir}{path to the root directory}

\item{shared.dir}{path to the shared directory (e.g. via Dropbox; root.dir if NULL)}

\item{nb.breeders}{number of breeders (an additional "test" breeder will be created)}

\item{lang}{language to be used (fr/en)}
}
\value{
list
}
\description{
Set up the directories and names for the breeding game.
Already-existing directories are not re-created.
}
\author{
Timothee Flutre
}

