% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{setupQtlCrossObject}
\alias{setupQtlCrossObject}
\title{Set-up a R/qtl "cross" object}
\usage{
setupQtlCrossObject(gendat, cross.type, genmap = NULL, phenos = NULL)
}
\arguments{
\item{gendat}{genotype data in the qtl format as a matrix with genotypes in rows and markers in columns; if genmap is provided, only the markers also on the map will be kept}

\item{cross.type}{type of cross ("bc" for a backcross, "4way" for a cross between two outbred parents)}

\item{genmap}{genetic map as a data frame with one row per marker and at least two columns named "linkage.group" and "locus"; if a third one, "genetic.distance", is absent, it will be created and filled with fake, incremental values; if NULL, a fake genetic map will be created with all markers in the same linkage group; it is assumed that all linkage groups correspond to autosomes}

\item{phenos}{data frame containing the phenotypes; if not NULL, the genotype identifiers should be in a column named "Genotype" or as row names; otherwise, will be created with such a column}
}
\value{
object of class "cross" as defined in the qtl package
}
\description{
Set up a "cross" object from the \href{https://cran.r-project.org/package=qtl}{qtl} package.
}
\seealso{
\code{\link{joinMap2backcross}}
}
\author{
Timothee Flutre
}
