% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{simplifTukey}
\alias{simplifTukey}
\title{Simplify Tukey's groups}
\usage{
simplifTukey(
  emm,
  tuk,
  colEmmID,
  colEmmVal = "emmean",
  colContrID = "contrast",
  colContrVal = "estimate",
  colPv = "p.value",
  alpha = 0.05,
  plot.it = FALSE,
  title = NULL
)
}
\arguments{
\item{emm}{data frame with the estimated marginal means per level of the factor of interest}

\item{tuk}{data frame with the contrasts as the parwise differences between factor levels}

\item{colEmmID}{name of the column in \code{emm} that contains the names of levels of the factor of interest}

\item{colEmmVal}{name of the column in \code{emm} that contains the estimated marginal means}

\item{colContrID}{name of the column in \code{tuk} that contains the names of the contrasts}

\item{colContrVal}{name of the column in \code{tuk} that contains the values of the contrasts}

\item{colPv}{name of the column in \code{tuk} that contains the p values of the contrasts}

\item{alpha}{value of the type I error rate}

\item{plot.it}{logical; if TRUE, will call \code{\link{plotTukey}}}

\item{title}{optional title}
}
\value{
data frame \code{emm} with a new column \code{"group_simplif"}
}
\description{
Simplifies groups from Tukey's HSD test as proposed in \href{https://acsess.onlinelibrary.wiley.com/doi/10.2134/agronj2017.10.0580}{Piepho (2018)}.
}
\examples{
## From Piepho (2018):
n <- 5
(tab1 <- data.frame(farmID=LETTERS[1:n],
                    mlc=c(1.86, 2.34, 1.82, 2.18, 1.38),
                    group=c("ab", "a", "ab", "a", "b")))
vec1 <- setNames(tab1$mlc, tab1$farmID)
vec1 <- sort(vec1, decreasing=TRUE)
(sortedFarmIDs <- names(vec1))
tab3 <- matrix(nrow=n, ncol=n,
               dimnames=list(names(vec1), names(vec1)))
tab3[1, 2:n] <- c(0.16, 0.48, 0.52, 0.96)
tab3[2, 3:n] <- c(0.32, 0.36, 0.80)
tab3[3, 4:n] <- c(0.04, 0.48)
tab3[4, n] <- c(.44)
tab3


tuk <- matWide2Long(tab3, diag=FALSE, outColNs=c("lev1","lev2","diff"))
tuk <- data.frame(contrast = paste0(tuk$lev1, " - ", tuk$lev2),
                  estimate = tuk[,"diff"])
tuk$p.value <- runif(nrow(tuk), min=0.051, max=1)
idx <- which(tuk$contrast \%in\% c("B - E","D - E"))
tuk$p.value[idx] <- runif(length(idx), min=0, max=0.05)
tuk
simplifTukey(tab1, tuk, colEmmID="farmID", colEmmVal="mlc", plot.it=TRUE,
             title="Fig. I from Piepho (2018)")
}
\seealso{
\code{\link{plotTukey}}
}
\author{
Timothee Flutre
}
