% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{SIMQTL}
\alias{SIMQTL}
\title{QTL detection by SIM}
\usage{
SIMQTL(cross, response.in.cross = TRUE, response = NULL,
  numeric.chr.format = TRUE, geno.joinmap = NULL, phase = NULL,
  pheno.col = "y", nperm = 100, threshold.alpha = 0.05,
  QTL_position = NULL, method = "em", plot = FALSE, verbose = 0)
}
\arguments{
\item{cross}{object}

\item{response.in.cross}{logical}

\item{response}{character}

\item{numeric.chr.format}{logical}

\item{geno.joinmap}{genotypes at markers in the JoinMap format}

\item{phase}{phase}

\item{pheno.col}{character}

\item{nperm}{number of permutations to be done in \code{qtl::scanone}}

\item{threshold.alpha}{vector of length 1 or 2 (optional) with thresholds for (1) the significance of QTL presence (based on permutations) and (2) the significance of QTL effects}

\item{QTL_position}{useful for simulated data}

\item{method}{method to detect QTL in \code{qtl::scanone}}

\item{plot}{logical}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
list
}
\description{
Perform QTL detection by simple interval mapping
}
\seealso{
\code{\link{MIMQTL}}
}
\author{
Charlotte Brault [aut], Timothee Flutre [ctb]
}
