% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simul.animal.model}
\alias{simul.animal.model}
\title{Simulate a data set from a basic animal model.}
\usage{
simul.animal.model(n = 300, mu = 4, P = 1, b = 2, nb.snps = 1000,
  maf = 0.3, A = NULL, sigma2 = 5, lambda = 3)
}
\arguments{
\item{n}{number of individuals (default is 300)}

\item{mu}{global mean (default is 4)}

\item{P}{number of fixed effects (default is 1)}

\item{b}{fixed effects (default is 2)}

\item{nb.snps}{number of SNPs (default is 1000; ignored if A is given)}

\item{maf}{minor allele frequency (default is 0.3; ignored if A is given)}

\item{A}{matrix of additive relationships}

\item{sigma2}{variance component of the errors (default is 5)}

\item{lambda}{ratio of variance components as sigma_u^2 /sigma^2 (default is 3)}
}
\value{
list with all input variables and the data set ready to be analyzed
}
\description{
y = mu 1_n + X b + Z u + e = W a + Z u + e
y is n x 1; X is n x P; Z is n x Q; W is n x (P+1)
u ~ Norm_Q(0, sigma_u^2 A); e ~ Norm_n(0, sigma^2 I_n)
}
\author{
Timothée Flutre
}

