% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simul.bslmm}
\alias{simul.bslmm}
\title{Simulate phenotypes according to the BSLMM model.}
\usage{
simul.bslmm(X, Q = 1, pi = NULL, h = NULL, rho = NULL, seed = NULL)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses, with SNPs in
columns and individuals in rows (SNPs with missing values or low MAF
should be discarded beforehand)}

\item{Q}{number of covariates (including the intercept)}

\item{pi}{proportion of beta-tilde values that are non-zero}

\item{h}{approximation to E[PVE] (h and rho should be NULL or not together)}

\item{rho}{approximation to E[PGE]}

\item{seed}{seed for the pseudo-random number generator}
}
\value{
list
}
\description{
See Zhou, Carbonetto & Stephens (2013).
}
\author{
Timothee Flutre
}

