% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulBvsr}
\alias{simulBvsr}
\title{BVSR}
\usage{
simulBvsr(Q = 3, mu = 50, mean.c = 5, sd.c = 2, X, pi = 1,
  pve.A = 0.7, sigma.a2 = 10^(-3), perc.NA = 0, err.df = Inf,
  seed = NULL)
}
\arguments{
\item{Q}{number of fixed effects, i.e. the intercept plus the number of years during which individuals are phenotyped (starting in 2010)}

\item{mu}{overall mean}

\item{mean.c}{mean of the prior on c[2:Q]}

\item{sd.c}{std dev of the prior on c[2:Q]}

\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns (SNPs with missing values or low MAF should be discarded beforehand); will be used in the simulations as X_A which is the column-centered version of X encoded in {-1,0,1}}

\item{pi}{proportion of marker effects (a) that are non-zero; setting pi at 1 means simulating from the additive infinitesimal model (equivalent to ridge regression)}

\item{pve.A}{proportion of phenotypic variance explained by SNPs with non-zero effect ("heritability"); PVE_A = V[X_A a] / V[y]; used along with option sigma.a2 to choose a value for sigma^2}

\item{sigma.a2}{prior variance of the non-zero additive effects}

\item{perc.NA}{percentage of missing phenotypes, at random}

\item{err.df}{degrees of freedom of errors' Student's t-distribution}

\item{seed}{seed for the pseudo-random number generator}
}
\value{
list
}
\description{
Simulate phenotypes according to the following model: Y = W c + Z X_A a + epsilon where Y is N x 1; W is N x Q; c is Q x 1; Z is N x I; X_A is I x P and epsilon is N x 1 with epsilon ~ Normal_N(0, sigma^2 Id) and c ~ Normal(mean_a, sd_a) so that sd_a is large ("fixed effect"). For SNP p, a_p ~ Prob(gamma_p=1) Normal_1(0, sigma_a^2) + Prob(gamma_p=0) delta_0, where Prog(gamma_p=1) is named pi, and delta_0 is Dirac's delta function at 0. For the case where pi is small, see Guan & Stephens (2011), Carbonetto & Stephens (2012), Peltola et al (2012).
}
\author{
Timothee Flutre
}

