% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulCoalescent}
\alias{simulCoalescent}
\title{Coalescent with recombination}
\usage{
simulCoalescent(nb.inds = 100, ind.ids = NULL, nb.reps = 10,
  pop.mut.rate = 40, pop.recomb.rate = 40, chrom.len = 10^4,
  other = NULL, nb.pops = 1, mig.rate = 5, get.trees = FALSE,
  get.tmrca = FALSE, get.alleles = FALSE, verbose = 1)
}
\arguments{
\item{nb.inds}{diploids (thus nb of haplotypes is 2 * nb.inds)}

\item{ind.ids}{vector of identifiers (one per genotype)}

\item{nb.reps}{number of independent loci that will be produced (could be seen as distinct chromosomes)}

\item{pop.mut.rate}{theta = 4 N0 mu}

\item{pop.recomb.rate}{rho = 4 N0 r}

\item{chrom.len}{in bp}

\item{other}{character vector of length 1 with other parameters to the simulator (e.g. time-specific parameters such as "-G 6.93 -eG 0.2 0.0 -eN 0.3 0.5")}

\item{nb.pops}{number of populations (\code{\link{kmeans}} will then be used to pair haplotypes into diploid genotypes)}

\item{mig.rate}{migration rate = 4 N0 m (will be symmetric)}

\item{get.trees}{get gene genealogies in the Newick format}

\item{get.tmrca}{get time to most recent common ancestor and local tree lengths}

\item{get.alleles}{get fake alleles sampled in {A,T,G,C}}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
list with haplotypes (list), genotypes as allele doses (matrix) and SNP coordinates (data.frame)
}
\description{
Simulate haplotypes according to an approximation to the coalescent with recombination named the Sequential Coalescent with Recombination Model. Requires the scrm package (Staab et al, 2014).
}
\examples{
\dontrun{## simulate haplotypes and genotypes in a single population
nb.genos <- 200
Ne <- 10^4
chrom.len <- 10^5
mu <- 10^(-8)
c <- 10^(-8)
genomes <- simulCoalescent(nb.inds=nb.genos,
                           pop.mut.rate=4 * Ne * mu * chrom.len,
                           pop.recomb.rate=4 * Ne * c * chrom.len,
                           chrom.len=chrom.len)
}
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{makeCrosses}}
}

