% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{simulGeneralisedLogistic}
\alias{simulGeneralisedLogistic}
\title{Generalised logistic growth (Richards' curve)}
\usage{
simulGeneralisedLogistic(
  t = seq(-1.5, 3.5, 0.5),
  a = 0,
  k = 1,
  r = 3,
  nu = 0.5,
  b = 0.5,
  c = 1,
  sigma2 = 0
)
}
\arguments{
\item{t}{vector of time points}

\item{a}{lower asymptote}

\item{k}{upper asymptote (if a=0, k is called the carrying capacity)}

\item{r}{growth rate}

\item{nu}{affects near which asymptote maximum growth occurs (> 0)}

\item{b}{related to the value at t=0}

\item{c}{typically takes a value of 1}

\item{sigma2}{variance of the errors}
}
\value{
list
}
\description{
Simulate phenotypes via the following model (\href{https://doi.org/10.1093/jxb/10.2.290}{Richards (1959)}): y(t) = g(t) + epsilon(t) where g(t) = a + (k - a) / ((c + b exp(-r t))^(1/nu)) and epsilon(t) ~ N(0, sigma^2).
}
\examples{
\dontrun{set.seed(1859)
model <- simulGeneralisedLogistic(t=seq(-1.5, 3.5, 0.1), a=0, k=1, r=3, nu=0.5, b=0.5, c=1, sigma2=0.01)
plot(x=model$t, y=model$g.t, type="b", las=1, xlab="time (t)", ylab="g(t)",
     main="Generalised logistic without noise")
plot(x=model$t, y=model$y.t, type="b", las=1, xlab="time (t)", ylab="y(t)",
     main="Generalised logistic with noise")
}
}
\author{
Timothee Flutre
}
