% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulGenosDose}
\alias{simulGenosDose}
\title{Genotypes}
\usage{
simulGenosDose(nb.genos, nb.snps, geno.ids = NULL, snp.ids = NULL, mafs = NULL)
}
\arguments{
\item{nb.genos}{number of genotypes (i.e. individuals)}

\item{nb.snps}{number of SNPs}

\item{geno.ids}{vector of genotype identifiers (if NULL, will be "geno001", etc)}

\item{snp.ids}{vector of SNP identifiers (if NULL, will be "snp001", etc)}

\item{mafs}{vector of minor allele frequencies; by default, they are uniformly distributed between 0.05 and 0.5}
}
\value{
matrix with genotypes in rows and SNPs in columns
}
\description{
Simulate SNP genotypes as allele dose additively encoded, i.e. 0,1,2.
}
\author{
Peter Carbonetto [aut], Timothee Flutre [ctb]
}
