% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulLogistic}
\alias{simulLogistic}
\title{Logistic growth}
\usage{
simulLogistic(t = 1:20, a = 50, g.t0 = 1, r = 0.6, sigma2 = 0)
}
\arguments{
\item{t}{vector of time points}

\item{a}{asymptotic value of g when t tends to infinity}

\item{g.t0}{initial value of g at time t = 0}

\item{r}{relative rate of growth}

\item{sigma2}{variance of the errors}
}
\value{
list
}
\description{
Simulate phenotypes via the following model (\href{http://www.genetics.org/content/161/4/1751.abstract}{Ma et al (2002)}): y(t) = g(t) + epsilon(t) where g(t) = a / (1 + b exp(-r t))) and epsilon(t) ~ N(0, sigma^2).
TODO: add QTL effect(s) + add AR(1) process on errors
}
\examples{
\dontrun{## without noise
model <- simulLogistic(t=1:20, a=50, g.t0=1, r=0.6, sigma2=0)
plot(x=model$t, y=model$g.t, type="b", las=1, xlab="time (t)", ylab="g(t)")

## with noise
set.seed(1859)
model <- simulLogistic(t=1:20, a=50, g.t0=1, r=0.6, sigma2=1)
plot(x=model$t, y=model$g.t, type="b", las=1, xlab="time (t)", ylab="g(t)")
}
}
\author{
Timothee Flutre
}
