% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulRefAltSnpAlleles}
\alias{simulRefAltSnpAlleles}
\title{SNP alleles}
\usage{
simulRefAltSnpAlleles(nb.snps = NULL, snp.ids = NULL, colnames = c("ref",
  "alt"), verbose = 1)
}
\arguments{
\item{nb.snps}{number of SNPs}

\item{snp.ids}{SNP identifiers (if \code{NULL}, will be generated by default)}

\item{colnames}{names of the columns for the output data frame}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame with SNPs in rows
}
\description{
Simulate alleles for bi-allelic SNPs.
}
\seealso{
\code{\link{simulCoalescent}}, \code{\link{simulGenosDose}}
}
\author{
Timothee Flutre
}
