% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{simulRefseqCompatibleWithVcf}
\alias{simulRefseqCompatibleWithVcf}
\title{Fasta file}
\usage{
simulRefseqCompatibleWithVcf(
  vcf.file = NULL,
  vcf.rng.all = NULL,
  chr.lengths = 10^5,
  fa.file = NULL,
  seed = NULL,
  verbose = 1
)
}
\arguments{
\item{vcf.file}{path to the file in the VCF format (will be used only if \code{vcf.rng.all=NULL}}

\item{vcf.rng.all}{data frame output from \code{\link{rngVcf2df}}}

\item{chr.lengths}{vector of chromosome lengths (if a single number, will be repeated, i.e. all chromosomes will have the ame length)}

\item{fa.file}{path to the output file in the fasta format (will be compressed via \code{gzip} if the name ends with \code{gz}); if NULL, the reference sequence will be returned as an object but not written into a file}

\item{seed}{seed for the pseudo-random number generator (will be set only if non-NULL)}

\item{verbose}{verbosity level (0/1)}
}
\value{
invisible DNAStringSet
}
\description{
Simulate a reference genome sequence compatible with alleles from genotype data in the VCF format.
Useful with the GATK software.
}
\author{
Timothee Flutre
}
