% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{simulSnpEffectsTraits12}
\alias{simulSnpEffectsTraits12}
\title{Simul breeding game}
\usage{
simulSnpEffectsTraits12(snp.ids, sigma.beta2 = c(trait1 = 10^(-5), trait2 =
  10^(-5)), prop.pleio = 0.4, cor.pleio = -0.7, verbose = 1)
}
\arguments{
\item{snp.ids}{vector of SNP identifiers for which effects should be simulated}

\item{sigma.beta2}{vector of variances of additive SNP effects for both traits}

\item{prop.pleio}{proportion of SNPs having an effect on both traits}

\item{cor.pleio}{genotypic correlation at these SNPs}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with a matrix of SNP effects and a vector identifying which SNPs are pleiotropic
}
\description{
Simulate the additive SNP effects of traits 1 and 2 jointly, with some degree of pleiotropy.
}
\examples{
\dontrun{set.seed(1859)
X <- simulGenosDose(nb.genos=825, nb.snps=2000)
snp.effects <- simulSnpEffectsTraits12(snp.ids=colnames(X))
sum(snp.effects$is.pleiotropic)
regplot(snp.effects$Beta[,1], snp.effects$Beta[,2])
}
}
\author{
Timothee Flutre
}

