% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{simulTrait3}
\alias{simulTrait3}
\title{Simul breeding game}
\usage{
simulTrait3(dat, X, afs = NULL, subset.snps = NULL, qtn.id = NULL,
  resist.genos = NULL, prob.resist.no.qtl = 0.02, verbose = 1)
}
\arguments{
\item{dat}{data.frame specifying the structure of the initial data set of phenotypes}

\item{X}{matrix of bi-allelic SNP genotypes encoded in allele dose in {0,1,2}}

\item{afs}{vector of allele frequencies (ignored if \code{qtn.id} is not NULL)}

\item{subset.snps}{vector of SNP identifiers to which the QTN should belong (ignored if \code{qtn.id} is not NULL)}

\item{qtn.id}{identifier of the causal SNP (QTN)}

\item{resist.genos}{vector of genotypes in allele dose at the causal SNP providing the resistance (ignored if \code{qtn.id} is NULL)}

\item{prob.resist.no.qtl}{probability for a genotype without the QTL to be resistant a year with the pathogen}

\item{verbose}{verbosity level (0/1)}
}
\value{
list
}
\description{
Simulate phenotypes of trait 3.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{makeDfInitPhenos}}
}

