% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{simulTraits12}
\alias{simulTraits12}
\title{Simul breeding game}
\usage{
simulTraits12(dat, mu = c(trait1 = 40, trait2 = 14), sigma.alpha2 = c(trait1
  = 6, trait2 = 3), Alpha = NULL, X, Beta, h2 = c(trait1 = 0.3, trait2 =
  0.4), cor.E.inter.trait = 0, verbose = 1)
}
\arguments{
\item{dat}{data.frame specifying the structure of the initial data set of phenotypes}

\item{mu}{vector of global means, for each trait}

\item{sigma.alpha2}{vector of variance for the "year" effects, for each trait (ignored if \code{Alpha} is not NULL)}

\item{Alpha}{matrix of "year" effects, for each trait (if NULL, will be simulated using \code{sigma.alpha2})}

\item{X}{matrix of bi-allelic SNP genotypes encoded in allele dose in {0,1,2}}

\item{Beta}{matrix of additive SNP effects, for each trait}

\item{h2}{vector of heritabilities, for each trait}

\item{cor.E.inter.trait}{correlation of errors between both traits}

\item{verbose}{verbosity level (0/1)}
}
\value{
list
}
\description{
Simulate phenotypes of traits 1 and 2 jointly: Y = Z.J * Alpha + Z.I * G.A + E.
}
\examples{
\dontrun{set.seed(1859)
X <- simulGenosDose(nb.genos=825, nb.snps=2000)
snp.effects12 <- simulSnpEffectsTraits12(snp.ids=colnames(X))
dat <- makeDfInitPhenos(line.ids=rownames(X))
phenos <- simulTraits12(dat, X=X, Beta=snp.effects12$Beta)
regplot(phenos$G.A[,1], phenos$G.A[,2])
}
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{simulSnpEffectsTraits12}}, \code{\link{makeDfInitPhenos}}
}

