% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{simulTraits12Rnd}
\alias{simulTraits12Rnd}
\title{Simul breeding game}
\usage{
simulTraits12Rnd(h2 = c(0.3, 0.4), sigma.beta2 = c(10^(-5), 10^(-5)),
  prop.pleio = 0.4, cor.pleio = -0.7, X, center.genos = FALSE,
  plot = TRUE, verbose = 1)
}
\arguments{
\item{h2}{narrow-sense heritability of both traits}

\item{sigma.beta2}{variance of the SNP effects for both traits}

\item{prop.pleio}{proportion of SNPs having an effect on both traits}

\item{cor.pleio}{genotypic correlation at these SNPs}

\item{X}{matrix of SNP genotypes encoded in allele dose in {0,1,2}; will be encoded in {-1,0,1} afterwards}

\item{center.genos}{if TRUE, the genotypes will be centered per SNP (but for the breeding game, it should be set to FALSE as the SNP effects won't change from generation to generation)}

\item{plot}{if TRUE, the first-trait genotypic values will be plotted against the second-trait}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix of genotypic values with noise
}
\description{
Simulate the random part of traits 1 and 2 jointly.
}
\author{
Timothee Flutre
}

