% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{snpCoordsDf2Gr}
\alias{snpCoordsDf2Gr}
\title{SNP coordinates from data frame to GRanges}
\usage{
snpCoordsDf2Gr(x, si = NULL)
}
\arguments{
\item{x}{data frame of SNP coordinates}

\item{si}{output from the "seqinfo" function from the GenomeInfoDb package}
}
\value{
GRanges
}
\description{
Convert a data frame of SNP coordinates to GRanges.
}
\author{
Timothee Flutre
}
