% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{snpDistances}
\alias{snpDistances}
\title{Calculate the distances between SNPs, assuming they are sorted.}
\usage{
snpDistances(snp.coords, nb.cores = 1)
}
\arguments{
\item{snp.coords}{data.frame with SNP identifiers as row names, and with two columns "chr" and "pos"}

\item{nb.cores}{the number of cores to use (default=1)}
}
\value{
list with one component per chromosome
}
\description{
Useful before estimating pairwise linkage disequilibrium.
}
\author{
Timothee Flutre
}

