% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{solveMme}
\alias{solveMme}
\title{Solve MME}
\usage{
solveMme(y, X, Z, sigma.u2, Ainv, sigma2)
}
\arguments{
\item{y}{vector of responses (or matrix with one column)}

\item{X}{incidence matrix of fixed effects}

\item{Z}{incidence matrix of random variables}

\item{sigma.u2}{variance component of the random variables}

\item{Ainv}{inverse of A, the variance-covariance matrix of the random variables}

\item{sigma2}{variance component of the errors}
}
\value{
vector of length PQ containing the BLUEs of beta and the BLUPs of u
}
\description{
Given T=1 response, P fixed effects and Q random variables (for a total of N observations), compute the BLUEs and BLUPs of the following linear mixed model by solving Henderson's mixed model equation (MME): y = X beta + Z u + epsilon, where y is N x 1; X is N x P; Z is N x Q; u ~ Normal_Qx1(0, sigma_u^2 A); epsilon ~ Normal_Nx1(0, sigma^2 Id_N).
}
\examples{
\dontrun{## simulate genotypes
set.seed(1859)
X <- simulGenosDose(nb.genos=200, nb.snps=2000)

## estimate the additive genetic relationships
A <- estimGenRel(X, relationships="additive", method="vanraden1", verbose=0)

## simulate phenotypes
model <- simulAnimalModel(T=1, Q=3, A=A, V.G.A=15, V.E=5, seed=1859)

## calculate BLUEs and BLUPs
if(isSingular(A)){
   Ainv <- mpInv(A)
} else
   Ainv <- solve(A)
fit <- solveMme(y=model$Y[,1,drop=FALSE], X=model$W, Z=model$Z,
                sigma.u2=model$V.G.A, Ainv=Ainv, sigma2=model$V.E)
cbind(model$C, fit[1:3])
cor(model$G.A, fit[4:length(fit)])
}
}
\seealso{
\code{link{makeMmeElements}}, \code{\link{makeMmeRhs}}, \code{\link{makeMmeLhs}}
}
\author{
Timothee Flutre
}
