% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{sortVcfFile}
\alias{sortVcfFile}
\title{Sort a VCF file}
\usage{
sortVcfFile(in.vcf.file, out.vcf.file)
}
\arguments{
\item{in.vcf.file}{path to the input VCF file}

\item{out.vcf.file}{path to the output VCF file (if it is the same as \code{in.vcf.file}, the input VCF file will be sorted into a temporary file which will then be copied back to it)}
}
\value{
invisible path to the output file
}
\description{
Sort a VCF file with \code{bcftools sort} (requires bcftools >= 1.6)
}
\seealso{
\code{\link{indexVcfFile}}
}
\author{
Timothee Flutre
}
