% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{stanAMmvwriteModel}
\alias{stanAMmvwriteModel}
\title{Animal model (multivariate)}
\usage{
stanAMmvwriteModel(stan.file = NULL, verbose = 0)
}
\arguments{
\item{stan.file}{path to a file; if NULL, a temporary one will be created}

\item{verbose}{verbosity level (0/1)}
}
\value{
invisible path to stan.file
}
\description{
Animal model (multivariate)
}
\seealso{
\code{\link{stanAMmv}}
}
\author{
Timothee Flutre
}
