% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{stanAMwriteModel}
\alias{stanAMwriteModel}
\title{Animal model (univariate)}
\usage{
stanAMwriteModel(
  stan.file = NULL,
  include.dom = FALSE,
  errors.Student = FALSE,
  missing.phenos = FALSE
)
}
\arguments{
\item{stan.file}{path to a file; if NULL, a temporary one will be created}

\item{include.dom}{include a variance component for dominant genetic relationships}

\item{errors.Student}{use a Student's t distribution for the errors (useful to handle outliers)}

\item{missing.phenos}{indicating if there is any missing phenotypes to impute jointly with the other unknown variables}
}
\value{
invisible path to stan.file
}
\description{
Animal model (univariate)
}
\seealso{
\code{\link{stanAM}}
}
\author{
Timothee Flutre
}
