% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{stats.all.pair.aligns}
\alias{stats.all.pair.aligns}
\title{Extract statistics from all pairwise alignments}
\usage{
stats.all.pair.aligns(aligns, nb.sequences)
}
\arguments{
\item{aligns}{list of instances of class PairwiseAlignments (see all.pair.aligns())}

\item{nb.sequences}{number of sequences}
}
\value{
list of matrices
}
\description{
Requires the Biostrings package.
}
\author{
Timothee Flutre
}

