% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{statsAllPairAligns}
\alias{statsAllPairAligns}
\title{Extract statistics from all pairwise alignments}
\usage{
statsAllPairAligns(aligns, nb.sequences)
}
\arguments{
\item{aligns}{list of instances of class PairwiseAlignments (see \code{\link{allPairAligns}})}

\item{nb.sequences}{number of sequences}
}
\value{
list of matrices
}
\description{
Requires the Biostrings package.
}
\author{
Timothee Flutre
}

