% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{statsStrongAssign}
\alias{statsStrongAssign}
\title{Genetic clustering}
\usage{
statsStrongAssign(assign.probs, assign.grps, thresh.strong.assign = 0.8)
}
\arguments{
\item{assign.probs}{vector of assignment probabilities}

\item{assign.grps}{vector of assigned clusters}

\item{thresh.strong.assign}{threshold}
}
\value{
list
}
\description{
Describes the strong assignment of individuals per cluster.
}
\author{
Timothee Flutre
}
