% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{stopIfNotValidGenosDose}
\alias{stopIfNotValidGenosDose}
\title{Genotypes}
\usage{
stopIfNotValidGenosDose(
  X,
  check.hasColNames = TRUE,
  check.hasRowNames = TRUE,
  check.noNA = TRUE,
  check.isDose = TRUE,
  check.notImputed = FALSE
)
}
\arguments{
\item{X}{input matrix}

\item{check.hasColNames}{logical}

\item{check.hasRowNames}{logical}

\item{check.noNA}{logical}

\item{check.isDose}{logical}

\item{check.notImputed}{logical}
}
\description{
Check that the input is a valid matrix of bi-allelic SNP genotypes coded in allele doses.
}
\author{
Timothee Flutre
}
