% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{stopIfNotValidHaplos}
\alias{stopIfNotValidHaplos}
\title{Haplotypes}
\usage{
stopIfNotValidHaplos(haplos, check.hasColNames = FALSE, check.noNA = TRUE)
}
\arguments{
\item{haplos}{input list}

\item{check.hasColNames}{logical}

\item{check.noNA}{logical}
}
\description{
Check that the input is a valid list of matrices of bi-allelic marker haplotypes.
}
\author{
Timothee Flutre
}

