% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{structure2clumpp}
\alias{structure2clumpp}
\title{STRUCTURE to CLUMPP}
\usage{
structure2clumpp(struct.files, clumpp.file, verbose = 1)
}
\arguments{
\item{struct.files}{vector of file names}

\item{clumpp.file}{file name}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
invisible list of lists, one per STRUCTURE file
}
\description{
Gather output files from STRUCTURE (several runs at the same number of clusters, K) into a single file for input to CLUMPP.
}
\seealso{
\code{\link{readOutSTRUCTURE}}
}
\author{
Timothee Flutre
}
