% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{subsetDiffHaplosWithinParent}
\alias{subsetDiffHaplosWithinParent}
\title{Haplotypes}
\usage{
subsetDiffHaplosWithinParent(haplos.chr, snps.tokeep)
}
\arguments{
\item{haplos.chr}{matrix containing both haplotypes of a given parent, with haplotypes in rows and SNPs in columns}

\item{snps.tokeep}{vector of SNPs to keep in the subset (typically the ones at which the crossing-overs occurred)}
}
\value{
matrix with both haplotypes in rows and a subset of SNPs in columns
}
\description{
Return a subset of SNPs for both haplotypes within a given parent in order to easily visualize a given crossing-over.
}
\author{
Timothee Flutre
}

