% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{subsetPedigree}
\alias{subsetPedigree}
\title{Subset a pedigree}
\usage{
subsetPedigree(ped, inds)
}
\arguments{
\item{ped}{data frame containing the input pedigree with three columns named "parent1", "parent2" and "child"; if "parent1" is different than "parent2", it is an allofecundation; if "parent1" is the same as "parent2" it is an autofecundation; if "parent2" is missing it is a haplodiploidization}

\item{inds}{vector of individual names (present in the "child" column of "ped") for which their ancestors will be retrieved, up to the founders}
}
\value{
data frame
}
\description{
From a pedigree, extract a subset of individuals.
}
\seealso{
\code{\link{plotPedigree}}
}
\author{
Julien Diot [cre,aut], Timothee Flutre [ctb]
}
