% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{subsetVcfOnAllelicity}
\alias{subsetVcfOnAllelicity}
\title{Parse VCF}
\usage{
subsetVcfOnAllelicity(vcf, single.ref = TRUE, single.alt = TRUE)
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg \href{http://bioconductor.org/packages/VariantAnnotation/}{VariantAnnotation})}

\item{single.ref}{if TRUE, only records with a single 'ref' are kept}

\item{single.alt}{if TRUE, only records with a single 'alt' are kept}
}
\value{
CollapsedVCF
}
\description{
Returns a VCF object subsetted for allelicity.
}
\author{
Gautier Sarah, Timothee Flutre
}

