% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{summarizeLd}
\alias{summarizeLd}
\title{Pairwise linkage disequilibrium}
\usage{
summarizeLd(ld, coln.var = "cor2", coln.dist = "dist.locs",
  bin.width = 500, max.phy.len = 10^6, nb.cores = 1, span = 0.75)
}
\arguments{
\item{ld}{data frame returned by \code{\link{estimLd}}}

\item{coln.var}{name of the column in \code{ld} which contains the LD estimates}

\item{coln.dist}{name of the column in \code{ld} which contains the physical distance between SNPs belonging to the same pair}

\item{bin.width}{width of each successive bin (they won't overlap)}

\item{max.phy.len}{maximum physical length to consider}

\item{nb.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously (not on Windows)}

\item{span}{the parameter alpha which controls the degree of loess smoothing}
}
\value{
list with the LD summarized over bins as well as loess fits of LD vs mid bin
}
\description{
Summarize estimates of linkage disequilibrium between pairs of SNPs belonging to the same chromosome over non-overlapping bins.
}
\seealso{
\code{\link{estimLd}}, \code{\link{estimLdPerChr}}, \code{\link{fitPhyDistVsLd}}
}
\author{
Timothee Flutre
}
