% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{summarizeLdPerChr}
\alias{summarizeLdPerChr}
\title{Pairwise linkage disequilibrium}
\usage{
summarizeLdPerChr(
  ld,
  coln.var = "cor2",
  coln.dist = "dist.locs",
  coln.chr = "chr",
  bin.width = 500,
  max.phy.len = 10^6,
  span = 0.75,
  nb.cores = 1,
  verbose = 0
)
}
\arguments{
\item{ld}{data frame returned by \code{\link{estimLd}}}

\item{coln.var}{name of the column in \code{ld} which contains the LD estimates}

\item{coln.dist}{name of the column in \code{ld} which contains the physical distance between SNPs belonging to the same pair}

\item{coln.chr}{name of the column in \code{ld} which contains the chromosome of each SNP pair}

\item{bin.width}{width of each successive bin (they won't overlap)}

\item{max.phy.len}{maximum physical length to consider}

\item{span}{the parameter alpha which controls the degree of loess smoothing}

\item{nb.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously (not on Windows)}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with a list per chromosome with the LD summarized over bins as well as loess fits of LD vs mid bin
}
\description{
Summarize estimates of linkage disequilibrium between pairs of SNPs belonging to the same chromosome over non-overlapping bins per chromosome.
}
\seealso{
\code{\link{summarizeLd}}
}
\author{
Timothee Flutre
}
