% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{summaryFasta}
\alias{summaryFasta}
\title{Fasta file}
\usage{
summaryFasta(fa.file, letters = c("A", "T", "G", "C", "N"))
}
\arguments{
\item{fa.file}{path to the fasta file (can be gzip-compressed)}

\item{letters}{vector of characters to get their frequency}
}
\value{
list
}
\description{
Return key aspects of a fasta file as summary (number of records, length of each record, letter frequency of each record).
}
\author{
Timothee Flutre
}

