% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{summaryGq}
\alias{summaryGq}
\title{Summary of GQ per variant}
\usage{
summaryGq(vcf.file, genome, yieldSize = 10^4, verbose = 1)
}
\arguments{
\item{vcf.file}{path to the VCF file}

\item{genome}{genome identifier (e.g. "VITVI_12x2")}

\item{yieldSize}{number of records to yield each time the file is read from (see ?TabixFile)}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
matrix with one row per variant
}
\description{
Compute the min, Q1, med, mean, Q3, max of the genotype qualities per variant.
}
\author{
Timothee Flutre
}

