% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summaryMcmcChain}
\alias{summaryMcmcChain}
\title{MCMC results}
\usage{
summaryMcmcChain(res.mcmc, param.names)
}
\arguments{
\item{res.mcmc}{object of class "mcmc" (not "mcmc.list"!)}

\item{param.names}{parameter names in the columns of res.mcmc}
}
\value{
matrix with parameters in rows
}
\description{
Return effective sample size, posterior mean, time-series standard error of the posterior mean (i.e. without ignoring autocorrelation; based on an estimate of the spectral density at 0 via \code{\link[coda]{spectrum0.ar}}), posterior standard deviation, and posterior quantiles (at 0.025, 0.5, 0.975).
}
\author{
Timothee Flutre
}
