% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{thinSnps}
\alias{thinSnps}
\title{Thin SNPs}
\usage{
thinSnps(method, threshold, snp.coords, only.chr = NULL)
}
\arguments{
\item{method}{index or coord}

\item{threshold}{keep every "threshold" SNPs (if method="index"), keep SNPs with more than "threshold" base pairs between them (if method="coord")}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"; SNPs will be sorted according to their coordinates per chromosome (use \code{gtools::mixedsort} if you want to also sort chromosomes)}

\item{only.chr}{identifier of a given chromosome}
}
\value{
vector of SNP identifiers
}
\description{
Thin SNPs according to various methods: based on their index or based on their genomic coordinate.
}
\author{
Timothee Flutre
}

