% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tridiag}
\alias{tridiag}
\title{Tridiagonal matrix}
\usage{
tridiag(main, upper, lower, as.covar = FALSE)
}
\arguments{
\item{main}{vector for the main diagonal}

\item{upper}{vector for the upper diagonal}

\item{lower}{vector for the lower diagonal}

\item{as.covar}{if TRUE, the matrix is interpreted as a variance-covariance matrix, with \code{main} giving the variances, and \code{upper} giving the correlations (the covariance will be deduced)}
}
\value{
matrix
}
\description{
Make a tridiagonal matrix.
}
\examples{
\dontrun{main <- 3:6
upper <- 1:3
lower <- 2:4
tridiag(main, upper, lower)

main <- sample(x=1:2, size=4, replace=TRUE)
upper <- lower <- rep(-0.4, 3)
tridiag(main, upper, lower)
}
}
\author{
Jthorpe (https://stackoverflow.com/a/28974577/597069)
}
