% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{updateJoinMap}
\alias{updateJoinMap}
\title{JoinMap format}
\usage{
updateJoinMap(x, verbose = 1)
}
\arguments{
\item{x}{data.frame in the JoinMap v2 format; the first four lines of a JoinMap loc-file should obviously be absent; columns \code{<locus name>}, \code{[PHASE]} and \code{[CLAS]} are ignored; missing genotypes are coded as "--"}

\item{verbose}{verbosity level (0/1)}
}
\value{
data.frame
}
\description{
The \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} software needs a locus genotype file ("loc-file") containing the information (genotypes, phased or not) of the loci for a single segregating population.
This function converts a data.frame in the JoinMap v2 format into the JoinMap v3-v4 format.
It is assumed that the population is of type CP: "a population resulting from a cross between two heterogeneously heterozygous and homozygous diploid parents, linkage phases originally (possibly) unknown".
}
\author{
Timothee Flutre
}

