% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{varqual2summary}
\alias{varqual2summary}
\title{Summary per variant}
\usage{
varqual2summary(vcf, field = "GQ")
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg VariantAnnotation)}

\item{field}{genotype field of the VCF to parse (GQ/DP)}
}
\value{
matrix with one row per variant
}
\description{
Compute the mean, sd, min, Q1, med, mean, Q3, max of the genotype qualities per variant, also reporting the number of samples and the number of missing data.
}
\author{
Timothee Flutre
}

