% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{vcf2dosage}
\alias{vcf2dosage}
\title{Convert VCF to dose}
\usage{
vcf2dosage(vcf.file, dict.file, genome, seqs = NULL, gdose.file, amap.file,
  uncertain = FALSE, verbose = 1)
}
\arguments{
\item{vcf.file}{path to the VCF file (bgzip index should exist in same directory; should only contain SNPs and be already filtered for QD, FS, MQ, etc)}

\item{dict.file}{path to the SAM dict file (see \url{https://broadinstitute.github.io/picard/command-line-overview.html#CreateSequenceDictionary})}

\item{genome}{genome identifier (e.g. "VITVI_12x2")}

\item{seqs}{sequence identifier(s) to work on (e.g. "chr2")}

\item{gdose.file}{path to the output file to record genotypes as allele doses (will be gzipped)}

\item{amap.file}{path to the output file to record SNP positions and alleles (will be gzipped)}

\item{uncertain}{boolean indicating whether the genotypes to convert should come from the "GT" field (uncertain=FALSE) or the "GP" or "GL" field (uncertain=TRUE)}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
nothing
}
\description{
Convert genotypes from a VCF file into allele doses.
}
\author{
Timothee Flutre
}

