% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{vcf2segreg}
\alias{vcf2segreg}
\title{Convert VCF file}
\usage{
vcf2segreg(vcf.file, ped.file, out.file)
}
\arguments{
\item{vcf.file}{VCF file}

\item{ped.file}{file with the pedigree}

\item{out.file}{path to the output file}
}
\value{
nothing
}
\description{
Convert genotypes from the VCF encoding into the "segregation" encoding, e.g. used in MapQTL.
}
\author{
Gautier Sarah, Timothee Flutre
}

