% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{writeCartagene}
\alias{writeCartagene}
\title{Genotype coding}
\usage{
writeCartagene(x, file, type = "f2 backcross", aliases = NULL)
}
\arguments{
\item{x}{matrix of characters containing marker genotypes in the CarthaGene formatfor instance from \code{\link{joinMap2CarthaGene}}; row names should correspond to marker names}

\item{file}{path to the file to which the data will be written}

\item{type}{type of the data ("f2 backcross")}

\item{aliases}{aliases}
}
\value{
nothing
}
\description{
Write marker genotypes formatted for CarthaGene in a file.
}
\seealso{
\code{\link{joinMap2CarthaGene}}
}
\author{
Timothee Flutre
}
