% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{writeGenMapJoinMap}
\alias{writeGenMapJoinMap}
\title{Write genetic map for JoinMap/MapQTL}
\usage{
writeGenMapJoinMap(genmap, file, verbose = 1)
}
\arguments{
\item{genmap}{data frame with at least three columns named "linkage.group", "locus" and "genetic.distance", and one row per locus}

\item{file}{name of the file in which the data will be saved}

\item{verbose}{verbosity level (0/1)}
}
\value{
nothing
}
\description{
Writes the given genetic map in the JoinMap/MapQTL format.
}
\seealso{
\code{link{writePhenoJoinMap}}, \code{link{writeSegregJoinMap}}
}
\author{
Timothee Flutre
}
