% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FImpute.R
\name{writeGenosFimpute}
\alias{writeGenosFimpute}
\title{Write SNP genotypes for FImpute}
\usage{
writeGenosFimpute(X, genos.file, chips = NULL)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; missing values should be encoded as NA; the maximum length of genotypes identifiers is 30 characters}

\item{genos.file}{path to the file in which genotypes will be saved}

\item{chips}{if several microarrays were used, provide a vector with chip numbers which names are genotype identifiers (same as in \code{X})}
}
\value{
nothing
}
\description{
Write SNP genotypes into a file formatted for FImpute.
}
\author{
Timothee Flutre
}
