% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPHASE.R
\name{writeInputsFastphase}
\alias{writeInputsFastphase}
\title{Write inputs for fastPHASE}
\usage{
writeInputsFastphase(X, snp.coords, alleles, file, verbose = 0)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded, for each SNP, in number of copies of its second allele, i.e. as allele doses in \{0,1,2\}, with genotypes in rows and SNPs in columns; the "second" allele is arbitrary, it corresponds to the second column of \code{alleles}, which can be the minor or the major allele}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (exactly 2 columns are required); the second column should correspond to the allele which number of copies is counted at each SNP in \code{X}}

\item{file}{write the genotype data to this file (for instance 'genotypes_fastphase.txt' or 'genotypes_fastphase.txt.gz', but don't use \code{\link[base]{gzfile}})}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
nothing
}
\description{
Use the external software fcGENE from \href{dx.plos.org/10.1371/journal.pone.0097589}{Roshyara and Scholz (2014)}.
}
\author{
Timothee Flutre
}
