% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FImpute.R
\name{writeInputsFimpute}
\alias{writeInputsFimpute}
\title{Write inputs for FImpute}
\usage{
writeInputsFimpute(ctl.file, X = NULL, genos.file, chips = NULL,
  snp.coords = NULL, snp.info.file, vcf.file = NULL, yieldSize = 10000,
  ped = NULL, ped.file = NULL, title = "FImpute", out.dir,
  hap.lib.file = NULL, add.ungen = NULL, parentage.test = NULL,
  ped.depth = NULL, turn.off.fam = FALSE, turn.off.pop = FALSE,
  save.partial = FALSE, save.genos = FALSE, save.hap.lib = FALSE,
  random.fill = FALSE, nb.jobs = 1, verbose = 1)
}
\arguments{
\item{ctl.file}{path to the control file in which FImpute's configuration will be saved}

\item{X}{matrix of bi-allelic SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; missing values should be encoded as NA; the maximum length of genotypes identifiers is 30 characters; if not NULL, will be used in priority even if \code{vcf.file} is not NULL}

\item{genos.file}{path to the file in which genotypes will be saved}

\item{chips}{if several microarrays were used, provide a vector with chip numbers which names are genotype identifiers (same as in \code{X})}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two compulsory columns in that order, chromosome identifiers and coordinate/position identifiers; the maximum length of SNP identifiers is 50 characters; chromosome identifiers should be numeric; other optional column(s) contain the SNP order on the microarray(s) (maximum 10); compulsory if \code{X} is specified}

\item{snp.info.file}{path to the file in which SNP information will be saved}

\item{vcf.file}{path to the VCF file (if the bgzip index doesn't exist in the same directory, it will be created); used only if \code{X=NULL}}

\item{yieldSize}{number of records to yield each time the file is read from  (see \code{?TabixFile})}

\item{ped}{data frame of pedigree with four columns in that order, genotype identifiers (same as in \code{X}), parent1 identifiers (considered as father/sire), parent2 identifiers (considered as mother/dam), and sex (as M or F)}

\item{ped.file}{path to the file in which the pedigree will be saved}

\item{title}{character identifying the analysis}

\item{out.dir}{directory in which the output files will be saved}

\item{hap.lib.file}{path to a file containing a haplotype library}

\item{add.ungen}{to add ungenotyped genotypes in the imputation process, provide a list such as \code{list(min.fsize=4, output.min.fsize=4, output.min.call.rate=0.9)}}

\item{parentage.test}{check for parentage errors}

\item{ped.depth}{to set a maximum number of generations to be traced for family imputation, provide a number}

\item{turn.off.fam}{turn off family imputation}

\item{turn.off.pop}{turn off population imputation}

\item{save.partial}{save partial calls}

\item{save.genos}{save genotypes instead of haplotypes (encoded as {0,1,2,5/NA})}

\item{save.hap.lib}{save the haplotype library built from reference individuals}

\item{random.fill}{ramdom filling (imputation) based on allele frequency; useful to assess minimum accuracy}

\item{nb.jobs}{number of jobs}

\item{verbose}{verbosity level (0/1)}
}
\value{
nothing
}
\description{
Write inputs into files formatted for FImpute.
}
\author{
Timothee Flutre
}
