% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findhap.R
\name{writeInputsFindhap}
\alias{writeInputsFindhap}
\title{Write inputs for findhap}
\usage{
writeInputsFindhap(
  opt.file,
  X = NULL,
  genos.file,
  chips = NULL,
  snp.coords = NULL,
  snp.info.file,
  vcf.file = NULL,
  yieldSize = 10000,
  ped = NULL,
  ped.file = NULL,
  iters = 4,
  Xchrom = 0,
  maxlen = 1000,
  minlen = 100,
  steps = 3,
  maxhap = 5000,
  hapout = 1,
  genout = 1,
  damout = 1,
  listout = -1,
  errate = 0.003,
  verbose = 1
)
}
\arguments{
\item{opt.file}{path to the options file in which findhap's configuration will be saved}

\item{X}{matrix of bi-allelic SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; missing values should be encoded as NA; if not NULL, will be used in priority even if \code{vcf.file} is not NULL}

\item{genos.file}{path to the file in which genotypes will be saved}

\item{chips}{if several microarrays were used, provide a vector with chip numbers which names are genotype identifiers (same as in \code{X})}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two compulsory columns in that order, chromosome identifiers and coordinate/position identifiers; chromosome identifiers should be numeric; other optional column(s) contain the SNP order on the microarray(s) (maximum 10); compulsory if \code{X} is specified}

\item{snp.info.file}{path to the file in which SNP information will be saved}

\item{vcf.file}{path to the VCF file (if the bgzip index doesn't exist in the same directory, it will be created); used only if \code{X=NULL}}

\item{yieldSize}{number of records to yield each time the file is read from  (see \code{?TabixFile})}

\item{ped}{data frame of pedigree with four columns in that order, genotype identifiers (same as in \code{X}), parent1 identifiers (considered as father/sire), parent2 identifiers (considered as mother/dam), and sex (as M or F)}

\item{ped.file}{path to the file in which the pedigree will be saved}

\item{iters}{maximum iterations for haplotype imputation}

\item{Xchrom}{final chromosome number for X, or 0 if no X}

\item{maxlen}{maximum length of segments to check when haplotyping}

\item{minlen}{minimum length of segments to check when haplotyping}

\item{steps}{number of steps to get from maxlen to minlen}

\item{maxhap}{maximum different haplotypes within any segment}

\item{hapout}{output full haplotypes or not}

\item{genout}{output filled genotypes or not}

\item{damout}{output imputed genotypes of parents or not}

\item{listout}{output haplotypes list or not}

\item{errate}{fraction of miscalled genotypes allowed when matching}

\item{verbose}{verbosity level (0/1)}
}
\value{
nothing
}
\description{
Write inputs into files formatted for findhap
}
\author{
Timothee Flutre
}
