% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{writeInputsForFastPhase}
\alias{writeInputsForFastPhase}
\title{Convert genotypes}
\usage{
writeInputsForFastPhase(X, snp.coords, alleles, file)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (first is "minor", second is "major")}

\item{file}{write the genotype data to this file (for instance 'genotypes_fastphase.txt' or 'genotypes_fastphase.txt.gz', but don't use \code{\link[base]{gzfile}})}
}
\value{
nothing
}
\description{
Convert genotypes
}
\author{
Timothee Flutre
}

