% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{writePhenoJoinMap}
\alias{writePhenoJoinMap}
\title{Write phenotypes for JoinMap/MapQTL}
\usage{
writePhenoJoinMap(phenos, file, alias.miss = ".", verbose = 1)
}
\arguments{
\item{phenos}{data frame with one genotype per row and traits in columns (one colun can contain the genotype identifiers)}

\item{file}{name of the file in which the data will be saved}

\item{alias.miss}{alias used to replace missing values (which should be encoded as NA in "genmap")}

\item{verbose}{verbosity level (0/1)}
}
\value{
nothing
}
\description{
Writes the given phenotypes in the JoinMap/MapQTL format.
}
\seealso{
\code{\link{writeGenMapJoinMap}}, \code{\link{writeSegregJoinMap}}
}
\author{
Timothee Flutre
}
