% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{writeSegregJoinMap}
\alias{writeSegregJoinMap}
\title{Write genotypes}
\usage{
writeSegregJoinMap(pop.name, pop.type = "CP", locus, segregs, phases,
  classifs, genos, file, save.ind.names = TRUE)
}
\arguments{
\item{pop.name}{name of the population}

\item{pop.type}{type of the population}

\item{locus}{vector of locus names}

\item{segregs}{vector of segregation types}

\item{phases}{vector of phase types}

\item{classifs}{vector of classification types}

\item{genos}{data frame containing genotypes encoded in the JoinMap format,  similar to the output from \code{\link{genoClasses2JoinMap}}}

\item{file}{name of the file in which the data will be saved (will be compressed if ends with ".gz" and "gzip" is available in the PATH)}

\item{save.ind.names}{if TRUE, individual names will be saved at the end of the file}
}
\value{
nothing
}
\description{
Write "segregation" data into a "loc" file in the \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap/MapQTL} format.
}
\examples{
\dontrun{
nb.snps <- 6
x <- data.frame(par1=c("AA", "GC", "CG", "AT", NA, "AA"),
                par2=c("AT", "GC", "GG", "AT", "AT", "AT"),
                off1=c("AA", "GG", "CG", "AA", "AA", "AT"),
                off2=c("AT", "GG", "CG", "AT", "AT", "AA"),
                off3=c("AT", "GG", "GG", "TT", "TT", NA),
                off4=c(NA, NA, NA, NA, NA, NA),
                row.names=paste0("snp", 1:nb.snps),
                stringsAsFactors=FALSE)
jm <- genoClasses2JoinMap(x=x, reformat.input=TRUE, thresh.na=2, verbose=1)
idx <- which(! is.na(jm$seg))
writeSegregJoinMap(pop.name="test", pop.type="CP", locus=rownames(jm[idx,]),
                   segregs=jm[idx,"seg"], phases=rep("{01}", length(idx)),
                   classifs=rep("(a,c)", length(idx)),
                   genos=jm[idx,-c(1:9)], file="test.txt")
}
}
\seealso{
\code{\link{genoClasses2JoinMap}}
}
\author{
Timothee Flutre
}
