% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FImpute.R
\name{writeSnpInfoFimpute}
\alias{writeSnpInfoFimpute}
\title{Write SNP information for FImpute}
\usage{
writeSnpInfoFimpute(snp.coords, snp.info.file)
}
\arguments{
\item{snp.coords}{data.frame with SNP identifiers as row names, and two compulsory columns in that order, chromosome identifiers and coordinate/position identifiers; the maximum length of SNP identifiers is 50 characters; chromosome identifiers should be numeric; other optional column(s) contain the SNP order on the microarray(s) (maximum 10)}

\item{snp.info.file}{path to the file in which SNP information will be saved}
}
\value{
nothing
}
\description{
Write SNP information into a file formatted for FImpute.
}
\author{
Timothee Flutre
}
