% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findhap.R
\name{writeSnpInfoFindhap}
\alias{writeSnpInfoFindhap}
\title{Write SNP information for findhap}
\usage{
writeSnpInfoFindhap(snp.coords, snp.info.file)
}
\arguments{
\item{snp.coords}{data.frame with SNP identifiers as row names, and two compulsory columns in that order, chromosome identifiers and coordinate/position identifiers; chromosome identifiers should be numeric (the highest number for X-specific SNPs, Y-specific SNPs being unsupported); the "within" and "overall" columns will be added, as well as "n_chips" and "chip1" if there is a single microarray}

\item{snp.info.file}{path to the file in which SNP information will be saved}
}
\value{
nothing
}
\description{
Write SNP information into a file formatted for findhap.
}
\author{
Timothee Flutre
}
