% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{barplotInsertSizes}
\alias{barplotInsertSizes}
\title{Bar plot of insert sizes}
\usage{
barplotInsertSizes(file, main = NULL, add.text = FALSE)
}
\arguments{
\item{file}{path to the output file from \code{java -jar picard.jar CollectInsertSizeMetrics}}

\item{main}{overall title for the plot}

\item{add.text}{add total count, as well as Q25, median and Q75 for insert sizes, to the topright of the plot}
}
\value{
invisible data frame of the content of the file
}
\description{
Creates a bar plot with vertical bars of insert sizes from histogram data as calculated by the CollectInsertSizeMetrics program from the \href{https://broadinstitute.github.io/picard/}{Picard} software.
}
\author{
Timothee Flutre
}

