% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{betterSummary}
\alias{betterSummary}
\title{Better summary}
\usage{
betterSummary(x)
}
\arguments{
\item{x}{vector (or 1-column matrix) of numbers}
}
\value{
vector of 12 values
}
\description{
Format the output of \code{\link{summary}} in a single line, as well as the total number of observations, the number and percentage of missing data, the variance and standard deviation, and the coefficient of variation.
}
\seealso{
\code{\link{prettyPrintBetterSummary}}
}
\author{
Timothee Flutre
}
