% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{binaryClassif}
\alias{binaryClassif}
\title{Binary classification}
\usage{
binaryClassif(known.nulls, called.nulls)
}
\arguments{
\item{known.nulls}{vector of booleans (TRUE if the null is true)}

\item{called.nulls}{vector of booleans (TRUE if the null is accepted)}
}
\value{
vector with names
}
\description{
Return the number of true positives, false positives, true negatives,
false negatives, true positive proportion (sensitivity), false positive
proportion, accuracy, true negative proportion (specificity), false
discovery proportion, false negative proportion and positive predictive
value (precision)
}
\details{
Both input vectors should be sorted beforehand
}
\author{
Timothee Flutre
}

