% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{boxplotCandidateQtl}
\alias{boxplotCandidateQtl}
\title{Boxplot of QTL}
\usage{
boxplotCandidateQtl(y, X, snp, xlab = "Genotype", ylab = "Phenotype",
  verbose = 1, ...)
}
\arguments{
\item{y}{vector of phenotypes with genotype names}

\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in [0,2], with genotypes in rows and SNPs in columns; missing values should be encoded as NA; SNP genotypes at the given snp should not be imputed}

\item{snp}{character with SNP name corresponding to the candidate QTL to plot}

\item{xlab}{label ox the x-axis}

\item{ylab}{label of the y-axis}

\item{verbose}{verbosity level (0/1)}

\item{...}{other arguments to \code{\link[graphics]{boxplot}}}
}
\value{
invisible list with \code{y} and \code{x} used to make the boxplot (same order, with no missing data)
}
\description{
Make a boxplot of a candidate QTL.
}
\author{
Timothee Flutre
}

