% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcAsymptoticBayesFactorWakefield}
\alias{calcAsymptoticBayesFactorWakefield}
\title{Asymptotic Bayes factor}
\usage{
calcAsymptoticBayesFactorWakefield(theta.hat, V, W, log10 = TRUE)
}
\arguments{
\item{theta.hat}{MLE of the additive genetic effect}

\item{V}{variance of theta.hat}

\item{W}{variance of the prior on theta}

\item{log10}{to return the log10 of the ABF (default=TRUE)}
}
\value{
numeric
}
\description{
Calculate the asymptotic Bayes factor proposed by Wakefield in Genetic Epidemiology 33:79-86 (2009, \url{http://dx.doi.org/10.1002/gepi.20359}).
}
\author{
Timothee Flutre
}

