% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcExactBayesFactorServinStephens}
\alias{calcExactBayesFactorServinStephens}
\title{Exact Bayes factor}
\usage{
calcExactBayesFactorServinStephens(G, Y, sigma.a, sigma.d, log10 = TRUE)
}
\arguments{
\item{G}{vector of genotypes}

\item{Y}{vector of phenotypes}

\item{sigma.a}{variance of the prior on the additive genetic effect}

\item{sigma.d}{variance of the prior on the dominance genetic effect}

\item{log10}{to return the log10 of the ABF (default=TRUE)}
}
\value{
numeric
}
\description{
Calculate the exact Bayes factor proposed by Servin and Stephens in PLoS Genetics 3,7 (2007, \url{http://dx.doi.org/10.1371/journal.pgen.0030114}).
}
\author{
Bertrand Servin [cre,aut], Timothee Flutre [ctb]
}

