% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{calcFreqMissSnpGenosPerSnp}
\alias{calcFreqMissSnpGenosPerSnp}
\title{Missing genotypes}
\usage{
calcFreqMissSnpGenosPerSnp(X)
}
\arguments{
\item{X}{matrix of marker genotypes, with genotypes in rows and markers in columns; missing values should be encoded as NA}
}
\value{
vector
}
\description{
Calculate the frequency of missing genotypes for each marker.
}
\author{
Timothee Flutre
}
