% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{caretFitQtl}
\alias{caretFitQtl}
\title{Fit with qtl for caret}
\usage{
caretFitQtl(x, y, wts, param, lev, last, weights, classProbs, genmap,
  pop.type = "CP", nperm = 10, threshold.alpha = 0.05, plot = FALSE,
  phase, QTLmethod = "SIM", p2d = "", nb.cores = 1, QTLposition = NULL,
  verbose = 0)
}
\arguments{
\item{x}{current predictors used to fit the model}

\item{y}{current outcome used to fit the model}

\item{wts}{optional instance weights}

\item{param}{current tuning parameter values}

\item{lev}{class levels of the outcome (or NULL in regression)}

\item{last}{logical for whether the current fit is the final fit}

\item{weights}{?}

\item{classProbs}{logical for whether class probabilities should be computed}

\item{genmap}{genetic map}

\item{pop.type}{population type in the JoinMap format}

\item{nperm}{number of permutations}

\item{threshold.alpha}{vector}

\item{plot}{logical}

\item{phase}{?}

\item{QTLmethod}{character (SIM/MIM)}

\item{p2d}{path to directory with results of \code{scantwo} (passed to MIMQTL)}

\item{nb.cores}{number of cores (passed to MIMQTL)}

\item{QTLposition}{?}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
output of \code{\link{SIMQTL}} or \code{\link{MIMQTL}}
}
\description{
Fits the simple or multiple interval mapping model for usage with caret.
First, the genotypes in the JoinMap format are saved into a temporary file.
Second, a \code{cross} object is made.
Third, SIMQTL or MIMQTL is run (calc.genoprob, scanone or scantwo).
Fourth, allelic effects are estimated.
}
\author{
Timothee Flutre
}
