% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{caretSummary}
\alias{caretSummary}
\title{Summary for caret}
\usage{
caretSummary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{data frame with a column named \code{"obs"} and another named \code{"pred"}}

\item{lev}{levels (unused here)}

\item{model}{model (unused here)}
}
\value{
vector with the root measn square error, Pearson and Spearman correlations between all data points, the 50% best and the 25% best, as well as the regression intercept and slope
}
\description{
Returns various metrics comparing predicted data with the values of the hold-out fold.
}
\author{
Timothee Flutre
}
