% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{chromNames2integers}
\alias{chromNames2integers}
\title{Rename chromosomes}
\usage{
chromNames2integers(x, prefix = "chr", thresh.max.chr.int = +Inf)
}
\arguments{
\item{x}{vector of chromosome names (if factor, will be converted to character)}

\item{prefix}{characters to be removed at the start of each chromosome name (case will be ignored)}

\item{thresh.max.chr.int}{threshold on the maximum chromosome integer, above which the renaming will simply be the sequence from 1 to the number of chromosomes}
}
\value{
data.frame with original and new names
}
\description{
Rename chromosomes into integers, especially useful with FImpute, GEMMA, qqman, etc.
}
\examples{
\dontrun{## example from grapevine
chroms <- c("chr1", "chr1_random", "chr10", "chr10_random", "chrUn", "chr2")
chromNames2integers(x=chroms)

## example from apple
chroms <- c("Chr15", "Chr01", "Chr02", "Chr00", "Chr02")
chromNames2integers(x=chroms)

## example from cherry
chroms <- c("Super-Scaffold_14", "Super-Scaffold_4374", "Super-Scaffold_27")
chromNames2integers(x=chroms, prefix="Super-Scaffold_", thresh.max.chr.int=500)
}
}
\author{
Timothee Flutre
}
