% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{controlBayesFdr}
\alias{controlBayesFdr}
\title{FDR}
\usage{
controlBayesFdr(log10.bfs, pi0, fdr.level = 0.05, verbose = 1)
}
\arguments{
\item{log10.bfs}{vector containing the log10(BF) of each test}

\item{pi0}{estimate of the proba for a null hypothesis to be true}

\item{fdr.level}{threshold below which a null is rejected}

\item{verbose}{verbosity level (0/1)}
}
\value{
logical vector with TRUE if null is rejected (thus called significant)
}
\description{
Call significant tests by controlling the Bayesian FDR via the procedure from Newton et al (Biostatistics, 2004) also described in Muller et al (JASA, 2006).
}
\author{
Timothee Flutre
}

