% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{convertVcfToGenoDoseWithBcftools}
\alias{convertVcfToGenoDoseWithBcftools}
\title{Convert VCF file to dosage file}
\usage{
convertVcfToGenoDoseWithBcftools(in.vcf.file, out.txt.file, nb.cores = 1)
}
\arguments{
\item{in.vcf.file}{path to the input VCF file (gzip-compressed)}

\item{out.txt.file}{path to the output file (gzip-compressed)}

\item{nb.cores}{number of threads for bcftools}
}
\value{
nothing
}
\description{
Convert a VCF file to a genotype dosage file with \code{bcftools +dosage} and \code{datamash}.
}
\author{
Timothee Flutre
}
