% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{correctSpatialHeterogeneity}
\alias{correctSpatialHeterogeneity}
\title{Correct spatial heterogeneity}
\usage{
correctSpatialHeterogeneity(dat, response, fix.eff = NULL,
  min.ctls.per.year = 10, cressie = TRUE, vgm.model = c("Exp", "Sph",
  "Gau", "Ste"), nb.folds = 5, verbose = 1)
}
\arguments{
\item{dat}{data frame with, at least, columns named "geno", "control" (TRUE/FALSE), "rank", "location", "year" and <response>}

\item{response}{column name of dat corresponding to the response for which spatial heterogeneity will be corrected}

\item{fix.eff}{if not NULL, vector of column names of data corresponding to fixed effects to control for in the kriging (e.g. "block")}

\item{min.ctls.per.year}{minimum number of control data points in a given year to proceed}

\item{cressie}{if TRUE, the variogram function from the gstat package uses Cressie's robust variogram estimate, else it uses the classical method of moments}

\item{vgm.model}{type(s) of variogram model(s) given to the vgm function of the gstat package; if several, the best one (smaller sum of squared errors) will be used}

\item{nb.folds}{number of folds for the cross-validation}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
data frame as dat but with an additional column named <response>.csh
}
\description{
Use kriging to correct spatial heterogeneity in a plant field trial.
Kriging (i.e. prediction) is performed per year.
Then, the predicted responses that controls would have had if they had been planted everywhere are subtracted from the observed responses from the other genotypes.
}
\author{
Timothee Flutre
}
