% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{df2gr}
\alias{df2gr}
\title{Convert data frame to GRanges}
\usage{
df2gr(x, seq = "chr", start = "start", end = "end", strand = NULL,
  si = NULL, keep.all.seqlevels = FALSE, names2use = "row")
}
\arguments{
\item{x}{data frame of genomic coordinates}

\item{seq}{name of the column containing the sequence of the intervals}

\item{start}{name of the column containing the start of the intervals}

\item{end}{name of the column containing the end of the intervals}

\item{strand}{name of the column containing the strand of the intervals}

\item{si}{output from the "seqinfo" function from the GenomeInfoDb package}

\item{keep.all.seqlevels}{if TRUE, all sequence levels from \code{si} are kept, otherwise, only those present in \code{x}}

\item{names2use}{if NULL, the output will have no names; if "row", it will have the row names of the input (if any); else, it will have the content of the specified column (if it exists)}
}
\value{
GRanges
}
\description{
Convert a data frame of genomic coordinates to GRanges.
}
\author{
Timothee Flutre
}
