% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{discardMarkersMissGenos}
\alias{discardMarkersMissGenos}
\title{Missing genotypes}
\usage{
discardMarkersMissGenos(X, verbose = 1)
}
\arguments{
\item{X}{matrix of marker genotypes, with genotypes in rows and markers in columns; missing values should be encoded as NA}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix similar to X but possibly with less columns
}
\description{
Discard all markers with at least one missing genotype.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{imputeGenosWithMean}}
}

