% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{discardSnpsLowMaf}
\alias{discardSnpsLowMaf}
\title{Minor allele frequencies}
\usage{
discardSnpsLowMaf(X, mafs = NULL, thresh = 0.01, rmv.mono = FALSE,
  verbose = 1)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in [0,2], with genotypes in rows and SNPs in columns; missing values should be encoded as NA}

\item{mafs}{vector of minor allele frequencies; if NULL, will be estimated with \code{\link{estimSnpMaf}}}

\item{thresh}{threshold on minor allele frequencies strictly below which SNPs are ignored}

\item{rmv.mono}{if TRUE, monomorphic SNPs are also removed}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix similar to X but possibly with less columns
}
\description{
Discard the SNPs with a minor allele frequency below the given threshold, as well as monomorphic SNPs.
}
\author{
Timothee Flutre
}
