% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{drawLocCrossovers}
\alias{drawLocCrossovers}
\title{Crossing-overs}
\usage{
drawLocCrossovers(crosses, nb.snps, lambda = 2)
}
\arguments{
\item{crosses}{data.frame with three columns, parent1, parent2, child; if parent 1 and 2 are the same, it will be an autofecondation; if parent2 is NA, it will be a haplodiploidization}

\item{nb.snps}{vector with the nb of SNPs per chromosome, which names are chromosome names}

\item{lambda}{mean number of crossing-overs (parameter of a Poisson)}
}
\value{
list of lists (one per cross, then one per parent, then one per chromosome) whose names are crosses$child, in the same order
}
\description{
Draw the number and location of crossing-overs per gamete.
}
\examples{
\dontrun{## make inputs
nb.offs <- 100
names.offs <- paste0("par1-par2-",
                     sprintf(fmt=paste0("\%0", floor(log10(nb.offs))+1, "i"),
                             1:nb.offs))
crosses <- data.frame(parent1=rep("par1", nb.offs),
                      parent2=rep("par2", nb.offs),
                      child=names.offs,
                      stringsAsFactors=FALSE)

## draw locations of crossing-overs
loc.crossovers <- drawLocCrossovers(crosses=crosses,
                                    nb.snps=setNames(c(2739, 2811),
                                                     c("chr1", "chr2")))

## quick check at their distributions per chr
table(do.call(c, lapply(loc.crossovers, function(lc.off){
  do.call(c, lapply(lc.off, function(lc.par){
    sapply(lc.par, length)
  }))
})))
}
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{makeGameteSingleIndSingleChrom}}
}

