% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{effNbIndepTestsPerChr}
\alias{effNbIndepTestsPerChr}
\title{Multiple testing}
\usage{
effNbIndepTestsPerChr(X, snp.coords, chr, method = "galwey", pkg = "base", k)
}
\arguments{
\item{X}{SNP genotypes in allele dose (i.e., 0, 1, 2) with individuals in rows and SNPs in columns (optional if \code{corX} is supplied)}

\item{snp.coords}{data frame of SNP coordinates with at least two columns, "chr" and "coord" (or "pos")}

\item{chr}{name of the chromosome to analyze}

\item{method}{for now, only "galwey" is implemented}

\item{pkg}{package to compute the eigenvalues of \code{corX} if \code{evCorX} is not supplied (base/RSpectra)}

\item{k}{number of eigenvalues requested (used by the RSpectra package)}
}
\value{
numeric
}
\description{
Return the effective number of independent tests according to the method of \href{https://dx.doi.org/10.1002/gepi.20408}{Galwey (2009)} for a given chromosome after estimating the LD along it.
}
\seealso{
\code{\link{effNbIndepTests}}, \code{\link{adjustThreshGalwey}}
}
\author{
Maxence Remerand, Timothee Flutre
}
