% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimAf}
\alias{estimAf}
\title{Allele frequencies}
\usage{
estimAf(X)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}
}
\value{
vector
}
\description{
Estimate the frequency of the second allele for each SNP. Missing values should be encoded as NA in order to be ignored. Note that the "second" allele may not be the "minor" allele (the least frequent).
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{estimMaf}}
}

