% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimGenRel}
\alias{estimGenRel}
\title{Genomic relatedness}
\usage{
estimGenRel(X, mafs = NULL, thresh = 0.01, relationships = "additive",
  method = "vanraden1", theta = 0.5, verbose = 1)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses ({0,1,2}), with SNPs in columns and individuals in rows; SNPs with missing data are ignored}

\item{mafs}{vector with minor allele frequencies (calculated with `estimMaf` if NULL)}

\item{thresh}{threshold on allele frequencies below which SNPs are ignored (default=0.01, NULL to skip this step)}

\item{relationships}{relationship to estimate (default=additive/dominance/gauss) where "gauss" corresponds to the Gaussian kernel from Endelman (2011)}

\item{method}{if additive relationships, can be "vanraden1" (first method in VanRaden, 2008), "habier" (similar to 'vanraden1' without giving more importance to rare alleles; from Habier et al, 2007), "astle-balding" (two times equation 2.2 in Astle & Balding, 2009), "yang" (similar to 'astle-balding' but without ignoring sampling error per SNP; from Yang et al, 2010), "zhou" (centering the genotypes and not assuming that rare variants have larger effects; from Zhou et al, 2013) or "center-std"; if dominance relationships, can be "vitezica" (classical/statistical parametrization from Vitezica et al, 2013) or "su" (from Su et al, PLoS One, 2012)}

\item{theta}{smoothing parameter for "gauss" (default=0.5)}

\item{verbose}{verbosity level (default=1)}
}
\value{
matrix
}
\description{
Estimate genetic relationships between individuals from their SNP genotypes. Note that "relationships" are estimated, and not "coancestries" (2 x relationhips).
}
\author{
Timothee Flutre
}

