% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimH2means}
\alias{estimH2means}
\title{Broad-sense heritability}
\usage{
estimH2means(dat, colname.resp, colname.trial = "year", vc,
  geno.var.blups = NULL)
}
\arguments{
\item{dat}{data frame of input data after missing data have been excluded, with columns named \code{colname.resp}, "geno" and \code{colname.trial}}

\item{colname.resp}{name of the column containing the response}

\item{colname.trial}{name of the column identifying the trials (e.g. \code{"year"}, \code{"year_irrigation"}, etc)}

\item{vc}{data frame of variance components with columns "grp" and "vcov" (i.e. formatted as \code{as.data.frame(VarCorr())} from the "lme4" package); \code{grp="Residual"} for "var.e"}

\item{geno.var.blups}{vector of variances of empirical BLUPs of the genotypic effects, g, assuming g ~ MVN(0, G) where G = sigma_g^2 I_m; if not provided, the estimator of oakey et al won't be computed}
}
\value{
list with the mean number of trials, the mean number of replicates per trial, the broad-sense heritability (classical estimator from Falconer and Mackay, as well as optionally the one from Oakey et al), and a function to compute summary statistics whch can be used for estimating confidence intervals by bootstrap
}
\description{
Estimate broad-sense heritability (squared correlation between predicted and true genotypic effects) on an entry-mean basis:
\enumerate{
\item via the classical formula for balanced data sets (see Falconer and Mackay, or the introduction of \href{http://www.genetics.org/cgi/doi/10.1534/genetics.107.074229}{Piepho and Mohring (2007)}): H2 = var.g / var.p, where var.p = var.g + var.ge / m + var.e / (r x m) with "m" the number of trials and "r" the number of replicates per trial (for unbalanced data sets, the mean number of trials and replicates per trial are used);
\item via the formula of \href{http://dx.doi.org/10.1007/s00122-006-0333-z}{Oakey et al (2006)} for unbalanced data sets: H2 = 1 - trace(G^-1 C_zz) / m, with "m" the number of genotypes.
}
}
\author{
Timothee Flutre
}
