% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimLd}
\alias{estimLd}
\title{Return estimates of linkage disequilibrium between pairs of SNPs.}
\usage{
estimLd(X, K = NULL, pops = NULL, snp.coords, only.chr = NULL,
  only.pop = NULL, verbose = 0)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses, with SNPs in columns and individuals in rows}

\item{K}{matrix of kinship}

\item{pops}{vector of characters indicating the population of each individual}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "pos"}

\item{only.chr}{identifier of a given chromosome}

\item{only.pop}{identifier of a given population}

\item{verbose}{verbosity level (default=0/1)}
}
\value{
data frame
}
\description{
Requires package LDcorSV
}
\author{
Timothee Flutre
}

