% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimMaf}
\alias{estimMaf}
\title{Estimate minor allele frequencies of SNPs.}
\usage{
estimMaf(X)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}
}
\value{
vector
}
\description{
Missing values should be encoded as NA.
}
\author{
Timothee Flutre
}

