% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimSnpAf}
\alias{estimSnpAf}
\title{Allele frequencies}
\usage{
estimSnpAf(X)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in [0,2], with genotypes in rows and SNPs in columns; missing values should be encoded as NA}
}
\value{
vector
}
\description{
Estimate allele frequencies of bi-allelic SNPs.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{estimSnpMaf}}
}

