% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{estimSnpMaf}
\alias{estimSnpMaf}
\title{Minor allele frequencies}
\usage{
estimSnpMaf(X = NULL, afs = NULL)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in [0,2], with genotypes in rows and SNPs in columns; missing values should be encoded as NA}

\item{afs}{vector of allele frequencies; if NULL, X should be specified, and the allele frequencies will be estimated with \code{\link{estimSnpAf}}}
}
\value{
vector
}
\description{
Estimate minor allele frequencies of bi-allelic SNPs.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{estimSnpAf}}
}

