% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{fastphase}
\alias{fastphase}
\title{Genotype imputation}
\usage{
fastphase(X, snp.coords, alleles, out.dir = getwd(), task.id = "fastphase",
  nb.starts = 20, nb.iters = 25, nb.samp.haplos = 50,
  estim.haplos = FALSE, nb.clusters = 10, seed = 1859, clean = FALSE,
  verbose = 1)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (named "minor" and "major")}

\item{out.dir}{directory in which the output files will be saved}

\item{task.id}{identifier of the task (used in temporary and output file names)}

\item{nb.starts}{number of random starts of the EM algorithm}

\item{nb.iters}{number of iterations of the EM algorithm}

\item{nb.samp.haplos}{number of haplotypes sampled from the posterior distribution obtained from a particular random start of the EM algorithm}

\item{estim.haplos}{estimate haplotypes by minimizing genotype error}

\item{nb.clusters}{number of clusters}

\item{seed}{seed for the pseudo-random number generator for fastPHASE}

\item{clean}{remove files}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with matrix of genotypes before imputation, and matrix of haplotypes (2 per genotype, in rows) after imputation
}
\description{
Impute SNP genotypes via fastPHASE (Scheet and Stephens, 2006).
}
\examples{
\dontrun{## simulate haplotypes and genotypes in a single population
set.seed(1859)
nb.inds <- 100
Ne <- 10^4
chrom.len <- 10^5
mu <- 10^(-8)
c <- 10^(-7)
genomes <- simulCoalescent(nb.inds=nb.inds,
                           pop.mut.rate=4 * Ne * mu * chrom.len,
                           pop.recomb.rate=4 * Ne * c * chrom.len,
                           chrom.len=chrom.len)
nb.snps <- nrow(genomes$snp.coords)
plotHaplosMatrix(genomes$haplos[[1]]) # quick view of the amount of LD

## discard some genotypes according to a "microarray" design:
## some inds with high density of genotyped SNPs, and the others with
## low density of SNPs, these being on both microarrays
ind.names <- rownames(genomes$genos)
inds.high <- sample(x=ind.names, size=floor(0.5 * nb.inds))
inds.low <- setdiff(ind.names, inds.high)
snp.names <- colnames(genomes$genos)
snps.high.only <- sample(x=snp.names, size=0.5 * nb.snps)
X.na <- genomes$genos
X.na[inds.low, snps.high.only] <- NA
plotGridMissGenos(X=X.na)

## perform imputation
alleles <- as.data.frame(matrix(cbind(rep("A", nb.snps), rep("T", nb.snps)),
                                ncol=2, dimnames=list(snp.names,
                                                      c("minor", "major"))))
out.imp <- fastphase(X=X.na, snp.coords=genomes$snp.coords,
                     alleles=alleles, nb.starts=3, clean=TRUE)
X.imp <- segSites2allDoses(seg.sites=list(haplosAlleles2num(haplos=out.imp$haplos,
                                                            alleles=alleles)),
                           ind.ids=rownames(genomes$genos))

## assess imputation accuracy
sum(X.imp != genomes$genos)
}
}
\author{
Timothee Flutre
}

