% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{filterSegreg}
\alias{filterSegreg}
\title{Filter for segregation}
\usage{
filterSegreg(x, thresh.pval = 0.05, return.counts = FALSE, is.F2 = FALSE,
  nb.cores = 1, verbose = 1)
}
\arguments{
\item{x}{data.frame similar to the output from \code{\link{genoClasses2JoinMap}}, with locus in rows and where the first columns corresponding to parents are discarded, i.e. the first column should be named "seg" and the following should correspond to offsprings}

\item{thresh.pval}{significance threshold at which to control the FWER (Bonferroni-adjusted p values) and the FDR (Benjamini-Hochberg-adjusted p values); the number of non-rejected null hypotheses will be shown only if \code{verbose} is non-zero}

\item{return.counts}{if TRUE, counts used to calculate the chi2 statistic are returned}

\item{is.F2}{if TRUE, it is assumed that the two outbred parents were crossed once to make a F1 which was then autofecundated multiple times to make the offsprings; note that the SNP genotypes of the F1 should not be given (and they are often unavailable)}

\item{nb.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously (not on Windows)}

\item{verbose}{verbosity level (0/1)}
}
\value{
data.frame with one row per locus and several columns (chi2, p value, Bonferroni-adjusted p value, Benjamini-Hochberg-adjusted p value), as well as counts (if \code{return.counts=TRUE})
}
\description{
Filter genotypes based on the chi2 statistic to test for segregation distortion.
}
\seealso{
\code{\link{genoClasses2JoinMap}}, \code{\link{updateJoinMap}}, \code{\link{plotHistPval}}, \code{\link{qqplotPval}}
}
\author{
Timothee Flutre
}
