% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{filterVariantCalls}
\alias{filterVariantCalls}
\title{Filter variant calls}
\usage{
filterVariantCalls(vcf.file, genome, out.file, yieldSize = NA_integer_,
  dict.file = NULL, seq.id = NULL, seq.start = NULL, seq.end = NULL,
  is.snv = NULL, is.biall = NULL, min.var.dp = NULL, max.var.dp = NULL,
  min.alt.af = NULL, max.alt.af = NULL, min.spl.dp = NULL,
  min.prop.spl.dp = NULL, min.spl.gq = NULL, min.prop.spl.gq = NULL,
  max.var.nb.gt.na = NULL, max.var.prop.gt.na = NULL, verbose = 1)
}
\arguments{
\item{vcf.file}{path to the VCF file (bgzip index should exist in same directory; should be already filtered for INFO tags such as QD, FS, MQ, etc)}

\item{genome}{genome identifier (e.g. "VITVI_12x2")}

\item{out.file}{path to the output VCF file (a bgzip index will be created in the same directory)}

\item{yieldSize}{number of records to yield each time the file is read from (see ?TabixFile) if seq.id is NULL}

\item{dict.file}{path to the SAM dict file (see \url{https://broadinstitute.github.io/picard/command-line-overview.html#CreateSequenceDictionary}) if seq.id is specified with no start/end}

\item{seq.id}{sequence identifier to work on (e.g. "chr2")}

\item{seq.start}{start of the sequence to work on (if NULL, whole seq)}

\item{seq.end}{end of the sequence to work on (if NULL, whole seq)}

\item{is.snv}{if not NULL but TRUE, filter out the variants which are not SNVs}

\item{is.biall}{if not NULL but TRUE, filter out the variants with more than one alternative allele}

\item{min.var.dp}{minimum variant-level DP below which variants are filtered out}

\item{max.var.dp}{maximum variant-level DP above which variants are filtered out}

\item{min.alt.af}{minimum variant-level AF below which variants are filtered out}

\item{max.alt.af}{maximum variant-level AF above which variants are filtered out}

\item{min.spl.dp}{minimum sample-level DP}

\item{min.prop.spl.dp}{minimum proportion of samples with DP above threshold}

\item{min.spl.gq}{minimum sample-level GQ}

\item{min.prop.spl.gq}{minimum proportion of samples with GQ above threshold}

\item{max.var.nb.gt.na}{maximum number of samples with missing GT}

\item{max.var.prop.gt.na}{maximum proportion of samples with missing GT}

\item{verbose}{verbosity level (0/default=1)}
}
\value{
the destination file path as a character(1)

TRUE if SNV (single nucleotide variant)

TRUE if at most one alternate allele

TRUE if variant-level DP inside of given range

TRUE if allele frequency inside of given range

TRUE if high-enough proportion of samples with DP above threshold

TRUE if high-enough proportion of samples with GQ above threshold

TRUE if high-enough number of samples without missing genotypes

TRUE if high-enough proportion of samples without missing genotypes
}
\description{
Filter out variant calls from VCF file according to several criteria (bi-allelic, single nucleotide variant, proper amount of missing genotypes, overall depth and allele frequency).
}
\author{
Timothee Flutre
}

