% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{fitPhyDistVsLd}
\alias{fitPhyDistVsLd}
\title{Pairwise linkage disequilibrium}
\usage{
fitPhyDistVsLd(ld, phy.dist, method = "loess", span = 0.2, newdata)
}
\arguments{
\item{ld}{numeric vector of LD values}

\item{phy.dist}{numeric vector of physical distance}

\item{method}{loess/loglm/asyreg/biexp}

\item{span}{the parameter alpha which controls the degree of loess smoothing}

\item{newdata}{an optional data frame in which to look for variables with which to predict}
}
\value{
list with data, model fit and prediction(s)
}
\description{
Fit a model of physical distance versus LD.
}
\seealso{
\code{\link{summarizeLd}}
}
\author{
Timothee Flutre
}
