% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{formatReadCountsPerLane}
\alias{formatReadCountsPerLane}
\title{Reformat read counts per lane}
\usage{
formatReadCountsPerLane(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
matrix
}
\description{
Reformat the data.frame from \code{\link{loadReadCountsPerIndAndLane}} in a matrix of counts
}
\author{
Timothee Flutre
}

