% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{gemma}
\alias{gemma}
\title{Launch GEMMA}
\usage{
gemma(model = "ulmm", y, X, snp.coords, alleles = NULL, maf = 0.01,
  K.c = NULL, W, out.dir = getwd(), task.id = "gemma", verbose = 1,
  clean = "none", seed = 1859, burnin = 1000, nb.iters = 7000,
  thin = 10)
}
\arguments{
\item{model}{name of the model to fit (ulmm/bslmm)}

\item{y}{vector of phenotypes with genotype names}

\item{X}{matrix of bi-allelic SNP genotypes encoded, for each SNP, in number of copies of its second allele, i.e. as allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; the "second" allele is arbitrary, it corresponds to the second column of \code{alleles}, which can be the minor or the major allele}

\item{snp.coords}{data.frame with 3 columns (snp, coord, chr)}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (exactly 2 columns are required); the second column should correspond to the allele which number of copies is counted at each SNP in \code{X}; if NULL, fake alleles will be generated}

\item{maf}{SNPs with minor allele frequency strictly below this threshold will be discarded}

\item{K.c}{kinship matrix; if NULL, will be estimated using X via \code{\link{estimGenRel}} with \code{relationships="additive"} and \code{method="zhou"}}

\item{W}{matrix of covariates with genotypes in rows (names as row names), a first column of 1 and a second column of covariates values}

\item{out.dir}{directory in which the output files will be saved}

\item{task.id}{identifier of the task (used in temporary and output file names)}

\item{verbose}{verbosity level (0/1)}

\item{clean}{remove files: none, some (temporary only), all (temporary and results)}

\item{seed}{seed for the pseudo-random number generator for GEMMA}

\item{burnin}{number of iterations to discard as burn-in (if model="bslmm")}

\item{nb.iters}{number of iterations (if model="bslmm")}

\item{thin}{thining (if model="bslmm")}
}
\value{
invisible list
}
\description{
See Zhou & Stephens (Nature Genetics, 2012), and Zhou et al (PLoS Genetics, 2013).
}
\examples{
\dontrun{## simulate genotypes
set.seed(1859)
I <- 200
P <- 2000
X <- simulGenosDose(nb.genos=I, nb.snps=P)

## simulate phenotypes (only additive effects)
modelA <- simulBvsr(Q=1, X=X, pi=0.01, pve=0.7, sigma.a2=1)

## test SNPs one by one with the univariate LMM
snp.coords <- data.frame(coord=1:ncol(X),
                         chr="chr1",
                         row.names=colnames(X),
                         stringsAsFactors=FALSE)
alleles <- data.frame(major=rep("A", ncol(X)),
                      minor="a",
                      row.names=colnames(X),
                      stringsAsFactors=FALSE)
fit.u <- gemma(model="ulmm", modelA$Y[,1], X, snp.coords, alleles,
               W=modelA$W, out.dir=tempdir(), clean="all")
cor(modelA$a[modelA$gamma == 1], fit.u$tests$beta[modelA$gamma == 1])
cols <- rep("black",ncol(X)); cols[modelA$gamma==1] <- "red"
pv.adj <- qqplotPval(fit.u$tests$p_wald, col=cols, ctl.fdr.bh=TRUE,
                     plot.signif=TRUE)

## fit all SNPs jointly with the BSLMM
burnin <- 10^3
nb.iters <- 10^4
thin <- 10^2
fit.bs <- gemma(model="bslmm", y=modelA$Y[,1], X, snp.coords, alleles,
                W=modelA$W, out.dir=tempdir(), clean="all",
                burnin=burnin, nb.iters=nb.iters, thin=thin)
posterior.samples <- coda::mcmc(data=fit.bs$hyperparams, start=burnin + 1,
                                end=burnin + nb.iters, thin=thin)
summary(posterior.samples)

## simulate phenotypes (only dominant effects)
set.seed(1859)
modelD <- simulBvsr(Q=1, X=X, pi=0.01, pve=0.7, sigma.a2=0, sigma.d2=1)

## test additive SNPs one by one with the univariate LMM
A.z <- estimGenRel(X=X, relationships="additive", method="zhou")
fit.u.DA <- gemma(model="ulmm", y=modelD$Y[,1], X=X,
                  snp.coords, alleles,
                  K.c=A.z, W=modelD$W, out.dir=tempdir(), clean="all")
cols <- rep("black",ncol(X)); cols[modelD$gamma==1] <- "red"
pvadj.DA <- qqplotPval(fit.u.DA$tests$p_wald, col=cols, ctl.fdr.bh=TRUE,
                       plot.signif=TRUE)
t(binaryClassif(known.nulls=modelD$gamma == 0,
                called.nulls=pvadj.DA$pv.bh <= 0.05))

## test dominant SNPs one by one with the univariate LMM
X.D <- recodeIntoDominant(X=X)
fit.u.DD <- gemma(model="ulmm", y=modelD$Y[,1], X=X.D,
                  snp.coords, alleles,
                  K.c=A.z, W=modelD$W, out.dir=tempdir(), clean="all")
cols <- rep("black",ncol(X)); cols[modelD$gamma==1] <- "red"
pvadj.DD <- qqplotPval(fit.u.DD$tests$p_wald, col=cols, ctl.fdr.bh=TRUE,
                       plot.signif=TRUE)
t(binaryClassif(known.nulls=modelD$gamma == 0,
                called.nulls=pvadj.DD$pv.bh <= 0.05))
}
}
\author{
Timothee Flutre [aut,cre], Dalel Ahmed [ctb]
}
\seealso{
\code{link{gemmaUlmmPerChr}}
}

