% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{gemma}
\alias{gemma}
\title{Launch GEMMA}
\usage{
gemma(model = "ulmm", y, X, snp.coords, alleles, K.c = NULL, W, out.dir,
  task.id, verbose = 1, clean = "none", burnin = 1000, nb.iters = 7000,
  thin = 10)
}
\arguments{
\item{model}{name of the model to fit (default=ulmm/bslmm)}

\item{y}{vector of phenotypes}

\item{X}{matrix of SNP genotypes encoded as allele doses, with SNPs in
columns and individuals in rows}

\item{snp.coords}{data.frame with 3 columns (snp, coord, chr)}

\item{alleles}{data.frame with SNPs in rows (names as row names) and
alleles in columns (first is "minor", second is "major")}

\item{K.c}{kinship centered matrix of SNPs}

\item{W}{matrix of covariates with individuals in rows (names as row names), a first column of 1 and a second column of covariates values}

\item{out.dir}{directory in which the output files will be saved}

\item{task.id}{identifier of the task (used in output file names)}

\item{verbose}{verbosity level (default=1)}

\item{clean}{remove files: none, some (temporary only), all (temporary and results)}

\item{burnin}{number of iterations to discard as burn-in}

\item{nb.iters}{number of iterations}

\item{thin}{thining}
}
\value{
invisible data.frame
}
\description{
Launch GEMMA
}
\author{
Timothee Flutre [cre,aut], Dalel Ahmed [ctb]
}

