% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{gemmaUlmmPerChr}
\alias{gemmaUlmmPerChr}
\title{Launch GEMMA uLMM chromosome per chromosome}
\usage{
gemmaUlmmPerChr(y, X, snp.coords, alleles, chr.ids = NULL, W, out.dir,
  task.id = "", clean = "none", verbose = 1)
}
\arguments{
\item{y}{vector of phenotypes}

\item{X}{matrix of SNP genotypes encoded as allele doses, with SNPs in
columns and individuals in rows}

\item{snp.coords}{data.frame with 3 columns (snp, coord, chr)}

\item{alleles}{data.frame with SNPs in rows (names as row names) and
alleles in columns (first is "minor", second is "major")}

\item{chr.ids}{set of chromosome identifiers to analyze (optional, all by default)}

\item{W}{matrix of covariates with individuals in rows (names as row names), a first column of 1 and a second column of covariates values}

\item{out.dir}{directory in which the data files will be found}

\item{task.id}{identifier of the task (used in output file names)}

\item{clean}{remove files: none, some (temporary only), all (temporary and results)}

\item{verbose}{verbosity level (default=1)}
}
\value{
a list of the GEMMA outputs for all chromosomes
}
\description{
Launch GEMMA uLMM chromosome per chromosome
}
\author{
Timothee Flutre [cre,aut], Dalel Ahmed [ctb]
}

