% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{gemmaUlmmPerChr}
\alias{gemmaUlmmPerChr}
\title{GEMMA uLMM per chromosome}
\usage{
gemmaUlmmPerChr(y, X, snp.coords, alleles = NULL, maf = 0.01,
  chr.ids = NULL, W, out.dir, task.id = "gemma", clean = "none",
  verbose = 1)
}
\arguments{
\item{y}{vector of phenotypes with individual names}

\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{snp.coords}{data.frame with 3 columns (snp, coord, chr)}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (first is "minor", second is "major"); will be made of A's and B's if missing}

\item{maf}{SNPs with minor allele frequency strictly below this threshold will be discarded}

\item{chr.ids}{set of chromosome identifiers to analyze (optional, all by default)}

\item{W}{matrix of covariates with individuals in rows (names as row names), a first column of 1 and a second column of covariates values}

\item{out.dir}{directory in which the data files will be found}

\item{task.id}{identifier of the task (used in output file names)}

\item{clean}{remove files: none, some (temporary only), all (temporary and results)}

\item{verbose}{verbosity level (0/1)}
}
\value{
a data.frame with GEMMA's output for all chromosomes
}
\description{
See Zhou & Stephens (Nature Genetics, 2012).
}
\author{
Timothee Flutre [aut,cre], Dalel Ahmed [ctb]
}
\seealso{
\code{link{gemma}}, \code{\link{plotHistPval}}, \code{\link{qqplotPval}}
}

