% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{genoClasses2genoDoses}
\alias{genoClasses2genoDoses}
\title{Convert genotypes}
\usage{
genoClasses2genoDoses(
  x,
  na.string = "--",
  snpIDs1stCol = TRUE,
  errorIfNotBi = TRUE,
  verbose = 1
)
}
\arguments{
\item{x}{matrix or data.frame with bi-allelic SNPs in rows and genotypes in columns}

\item{na.string}{a character to be interpreted as NA values}

\item{snpIDs1stCol}{indicates if the SNP identifiers are in the first column; otherwise they should be in the row names}

\item{errorIfNotBi}{raises an error if a SNP is not bi-allelic; otherwise, raises a warning and set the SNP genotypes to missing}

\item{verbose}{verbosity level (0/1)}
}
\value{
list of a matrix of allele doses with SNPs in columns and genotypes in rows, and a matrix with major and minor alleles
}
\description{
Convert SNP genotypes from "genotypic classes" into "allele doses" by counting the number of minor alleles.
The code is not particularly efficient, but at least it exists.
}
\author{
Timothee Flutre
}
