% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{getHaplosInd}
\alias{getHaplosInd}
\title{Haplotypes of an genotype}
\usage{
getHaplosInd(haplos, ind.name)
}
\arguments{
\item{haplos}{list with one matrix per chromosome, with haplotype in rows and SNPs in columns, as returned by \code{\link{simulCoalescent}}}

\item{ind.name}{identifier of the genotype to retrieve}
}
\value{
list similar to "haplos"
}
\description{
Retrieve both haplotypes for all chromosomes of a given genotype.
}
\author{
Timothee Flutre
}
