% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{getIndNamesFromHaplos}
\alias{getIndNamesFromHaplos}
\title{Individual names}
\usage{
getIndNamesFromHaplos(haplos)
}
\arguments{
\item{haplos}{list of matrices (one per chromosome, with individuals in rows)}
}
\value{
vector
}
\description{
Return the identifiers of all individuals
}
\author{
Timothee Flutre
}

