% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getNbPCsMinimAvgSqPartCor}
\alias{getNbPCsMinimAvgSqPartCor}
\title{Return the number of PCs that minimizes the average squared partial
correlation}
\usage{
getNbPCsMinimAvgSqPartCor(X)
}
\arguments{
\item{X}{genotype matrix (0,1,2) with SNPs in rows and individuals in columns}
}
\value{
integer
}
\description{
Shriner (Heredity, 2011)
}
\author{
Daniel Shriner
}

