% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{grSummaryPerBin}
\alias{grSummaryPerBin}
\title{Genomic bins}
\usage{
grSummaryPerBin(
  gr,
  colname = NULL,
  binwidth = 200,
  which.summary = "sum",
  plot.it = FALSE,
  ...
)
}
\arguments{
\item{gr}{\code{GRanges} object}

\item{colname}{character specifying a given column in \code{mcols(gr)}; if NULL, the number of ranges per bin will be counted}

\item{binwidth}{fixed width of each bin}

\item{which.summary}{names of the summaryzing function (one or several among \code{c("sum", "mean", "min", "max")})}

\item{plot.it}{if TRUE, a karyogram will be plotted}

\item{...}{other arguments passed to \code{autoplot}, such as \code{main}}
}
\value{
invisible \code{GRanges} containing the bins with the summary
}
\description{
Summarize a given metric per \code{GRanges} over genomic bins.
}
\examples{
\dontrun{## dummy data
set.seed(1)
C <- 2
P <- C * 10^3
df <- data.frame(CHR=c(rep("chr1", 10^3), rep("chr2", 10^3)),
                 SNP=paste0("snp", 1:P),
                 POS=NA,
                 N=abs(rnorm(P)))
df$POS[df$CHR == "chr1"] <- sort(sample.int(10^5, 10^3))
df$POS[df$CHR == "chr2"] <- sort(sample.int(10^5, 10^3))
head(df)
str(df)

library(GenomicRanges)
df.gr <- GRanges(seqnames=df$CHR, ranges=IRanges(start=df$POS, width=1),
                 metric=df$N,
                 seqlengths=c(chr1=10^5, chr2=10^5))
df.gr

(bins <- grSummaryPerBin(gr=df.gr, colname=NULL, binwidth=200))
(bins <- grSummaryPerBin(gr=df.gr, colname="metric", binwidth=200,
                         which.summary="sum"))
(bins <- grSummaryPerBin(gr=df.gr, colname="metric", binwidth=200,
                         which.summary="sum", which.plot="sum", plot.it=TRUE))
}
}
\author{
Timothee Flutre
}
