% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{grSummaryPerBin}
\alias{grSummaryPerBin}
\title{Genomic bins}
\usage{
grSummaryPerBin(gr, colname, binwidth = 200, which.summary = "sum",
  which.plot = NULL)
}
\arguments{
\item{gr}{\code{GRanges} object}

\item{colname}{character specifying a given column in \code{mcols(gr)}}

\item{binwidth}{fixed width of each bin}

\item{which.summary}{names of the summaryzing function (one or several among \code{c("sum", "mean", "min", "max")})}

\item{which.plot}{name of the summaryzing function to plot along the karyogram (if not NULL; via the \href{http://www.bioconductor.org/packages/ggbio/}{ggbio} package)}
}
\value{
invisible \code{GRanges} containing the bins with the summary
}
\description{
Summarize a given metric per \code{GRanges} over genomic bins.
}
\author{
Timothee Flutre
}
