% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{haplosAlleles2num}
\alias{haplosAlleles2num}
\title{Haplotypes}
\usage{
haplosAlleles2num(haplos, alleles, nb.cores = 1)
}
\arguments{
\item{haplos}{matrix with haplotypes in rows (2 consecutive per genotype) and SNPs in columns}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (named "minor" and "major")}

\item{nb.cores}{the number of cores to use, i.e. at most how many child processes will be run simultaneously (not on Windows)}
}
\value{
matrix with 0 for major alleles and 1 for minor alleles
}
\description{
Reformat haplotypes from alleles as string into numeric.
}
\seealso{
\code{\link{segSites2allDoses}}
}
\author{
Timothee Flutre
}
