% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{haplosList2Matrix}
\alias{haplosList2Matrix}
\title{Convert haplotypes}
\usage{
haplosList2Matrix(haplos)
}
\arguments{
\item{haplos}{list of matrices (one per chromosome, with genotypes in rows and SNPs in columns)}
}
\value{
matrix
}
\description{
Convert a list of haplotypes into a matrix.
}
\author{
Timothee Flutre
}

