% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{imageWithScale}
\alias{imageWithScale}
\title{Plot a matrix as a heatmap in its natural orientation, with a colored
scale on the right side, and optionally using its dimension names for
rows and columns}
\usage{
imageWithScale(z, main = NULL, idx.rownames = NULL, idx.colnames = NULL,
  breaks = NULL, nb.breaks = 20, left.text.at = NULL, cex.txt = 1,
  cex.sc = 1, col.pal = grDevices::colorRampPalette(c("black", "red",
  "yellow"), space = "rgb"))
}
\arguments{
\item{z}{matrix to be plotted}

\item{main}{title to appear above the heatmap}

\item{idx.rownames}{vector giving the indices of the row names of z to be added on the left side of the plot}

\item{idx.colnames}{vector giving the indices of the column names of z to be added on top of the plot}

\item{breaks}{vector of breaks (if NULL, will be \code{seq(min(z), max(z), length.out=nb.breaks)})}

\item{nb.breaks}{number of breaks}

\item{left.text.at}{vector which names and values will be used to label the left side of the plot; if not NULL, takes precedence over idx.rownames}

\item{cex.txt}{numeric character expansion factor used with "left.text.at"}

\item{cex.sc}{numeric character expansion factor used with the scale}

\item{col.pal}{output of \code{\link[grDevices]{colorRampPalette}}}
}
\description{
To print all row names, choose idx.rownames=1:nrow(z). To print a subset
of 10 row names, choose idx.rownames=floor(seq(1, nrow(z), length.out=10)).
Similarly for column names.
}
\examples{
\dontrun{set.seed(1859)
genomes <- simulCoalescent(nb.inds=200, nb.pops=3, mig.rate=3)
X <- genomes$genos
A <- estimGenRel(X=X, relationships="additive", method="vanraden1")
imageWithScale(z=A, main="Additive genetic relationships", breaks=seq(0,1,length.out=20),
               left.text.at=setNames(c(0.83, 0.5, 0.17), c("pop1", "pop2", "pop3")))
}
}
\author{
Timothee Flutre
}
