% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{imputeGenosWithMeanPerPop}
\alias{imputeGenosWithMeanPerPop}
\title{Genotype imputation}
\usage{
imputeGenosWithMeanPerPop(X, pops, min.maf.pop = 0.1, max.miss.pop = 0.3,
  rm.still.miss = TRUE, verbose = 1)
}
\arguments{
\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in {0,1,2}, with genotypes in rows and SNPs in columns; missing values should be encoded as NA}

\item{pops}{data.frame with 2 columns, "ind" and "pop"}

\item{min.maf.pop}{minimum minor allele frequency per population (before imputation) below which SNPs are discarded}

\item{max.miss.pop}{maximum amount of missing genotypes per population (before imputation) above which SNPs are discarded}

\item{rm.still.miss}{if TRUE, remove SNP(s) still with missing genotype(s) in at least one population (depending on \code{min.maf.pop} and \code{max.miss.pop})}

\item{verbose}{verbosity level (0/1)}
}
\value{
matrix with imputed genotypes
}
\description{
Impute missing SNP genotypes with the mean of the non-missing, population by population.
}
\seealso{
\code{\link{imputeGenosWithMean}}
}
\author{
Timothee Flutre
}
