% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{inlaAM}
\alias{inlaAM}
\title{Animal model}
\usage{
inlaAM(dat, relmat, family = "gaussian", nb.threads = 1, verbose = 0,
  silent = TRUE)
}
\arguments{
\item{dat}{data.frame containing the data corresponding to relmat; should have a column grep-able for "response" as well as a column "geno.add" used with matrix A; if a column "geno.dom" exists, it will be used with matrix D; any other column will be interpreted as corresponding to "fixed effects"}

\item{relmat}{list containing the matrices of genetic relationships (see \code{\link{estimGenRel}}); additive relationships (matrix A) are compulsory, with name "geno.add"; dominant relationships (matrix D) are optional, with name "geno.dom"; can be in the "matrix" class (base) or the "dsCMatrix" class (Matrix package)}

\item{family}{a string indicating the likelihood family (see \code{\link[INLA]{inla}})}

\item{nb.threads}{maximum number of threads the inla-program will use (see \code{\link[INLA]{inla}})}

\item{verbose}{verbosity level (0/1)}

\item{silent}{if equal to TRUE, then the inla-program would be silent (see \code{\link[INLA]{inla}})}
}
\value{
\code{\link[INLA]{inla}} object
}
\description{
Given I genotypes, Q covariates and N=I*Q phenotypes for the trait, fit an "animal model" with the INLA package via the following likelihood: y = W c + Z g_A + Z g_D + epsilon, where y is Nx1; W is NxQ; Z is NxI; g_A ~ Normal_I(0, sigma_A^2 A) with A the known matrix of additive genetic relationships; g_D ~ Normal_I(0, sigma_D^2 D) with D the known matrix of dominant genetic relationships; epsilon ~ Normal_N(0, sigma^2 Id_N); Cov(g_A,g_D)=0; Cov(g_A,e)=0; Cov(g_D,e)=0.
}
\examples{
\dontrun{## simulate genotypes
set.seed(1859)
I <- 100 # genotypes
P <- 2000 # SNPs
X <- matrix(sample(0:2, size=I*P, replace=TRUE), nrow=I, ncol=P,
            dimnames=list(paste0("geno", 1:I), paste0("snp", 1:P)))

## simulate phenotypes with only additive part of genotypic values
A <- estimGenRel(X, relationships="additive", method="vanraden1", verbose=0)
A <- as.matrix(Matrix::nearPD(A)$mat) # not always necessary
modelA <- simulAnimalModel(T=1, Q=3, A=A, V.G.A=15, V.E=5)

## infer with INLA
library(INLA)
modelA$dat$geno.add <- modelA$dat$geno; modelA$dat$geno <- NULL
fitA <- inlaAM(dat=modelA$dat, relmat=list(geno.add=A))
summary(fitA)
c(modelA$C); 1/modelA$V.G.A; 1/modelA$V.E

## simulate phenotypes with additive and dominant parts of genotypic values
D <- estimGenRel(X, relationships="dominant", method="vitezica", verbose=0)
D <- as.matrix(Matrix::nearPD(D)$mat) # not always necessary
modelAD <- simulAnimalModel(T=1, Q=3, A=A, V.G.A=15, V.E=5,
                            D=D, V.G.D=3)

## infer with INLA
modelAD$dat$geno.add <- modelAD$dat$geno
modelAD$dat$geno.dom <- modelAD$dat$geno; modelAD$dat$geno <- NULL
fitAD <- inlaAM(dat=modelAD$dat, relmat=list(geno.add=A, geno.dom=D))
summary(fitAD)
c(modelAD$C); 1/modelAD$V.G.A; 1/modelAD$V.E; 1/modelAD$V.G.D
}
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{lmerAM}}, \code{\link{jagsAM}}
}

