% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/molbio_plates.R
\name{lengthen.plate}
\alias{lengthen.plate}
\title{Lengthen plate}
\usage{
lengthen.plate(plate.w)
}
\arguments{
\item{plate.w}{matrix of a plate in the "wide" format}
}
\value{
data.frame of a plate in the "long" format (1 well per row)
}
\description{
Lengthen a "wide" plate into 3 columns for easier processing.
}
\author{
Timothee Flutre
}

