% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{lmerAM}
\alias{lmerAM}
\title{Animal model}
\usage{
lmerAM(formula, data, relmat, REML = TRUE, verbose = 1)
}
\arguments{
\item{formula}{formula (see \code{\link[lme4]{lmer}})}

\item{data}{data.frame containing the data corresponding to formula and relmat (see \code{\link[lme4]{lmer}})}

\item{relmat}{list containing the matrices of genetic relationships (A is compulsory but D is optional); should use the same names as the colnames in data; can be in the "matrix" class (base) or the "dsCMatrix" class (Matrix package); see \code{\link{estimGenRel}}}

\item{REML}{default is TRUE (use FALSE to compare models with different fixed effects)}

\item{verbose}{verbosity level (0/1)}
}
\value{
\code{\link[lme4]{merMod}} object
}
\description{
Given I individuals, Q covariates and N=I*Q phenotypes for the trait, fit an "animal model" with the lme4 package via the following likelihood: y = W c + Z g_A + Z g_D + epsilon, where y is Nx1; W is NxQ; Z is NxI; g_A ~ Normal_I(0, sigma_A^2 A) with A the known matrix of additive genetic relationships; g_D ~ Normal_I(0, sigma_D^2 D) with D the known matrix of dominance genetic relationships; epsilon ~ Normal_N(0, sigma^2 Id_N); Cov(g_A,g_D)=0; Cov(g_A,e)=0; Cov(g_D,e)=0.
}
\note{
If A is not positive definite, an error will be raised (via \code{\link[base]{chol}}); in such cases, using \code{\link[Matrix]{nearPD}} can be useful.
}
\examples{
set.seed(1859)
I <- 100 # individuals
P <- 2000 # SNPs
X <- matrix(sample(0:2, size=I*P, replace=TRUE), nrow=I, ncol=P,
            dimnames=list(paste0("ind", 1:I), paste0("snp", 1:P)))
A.vr <- estimGenRel(X, relationships="additive", method="vanraden1", verbose=0)
A.vr <- as.matrix(Matrix::nearPD(A.vr)$mat) # not always necessary
model <- simulAnimalModelMultivar(T=1, Q=3, A=A.vr, sigma.A2=15, V.E=5)
c(model$C)
res <- lmerAM(formula=response1 ~ year + (1|ind), data=model$dat,
              relmat=list(ind=A.vr))
summary(res)
fixef(res)
vc <- as.data.frame(VarCorr(res))
c(vc[vc$grp == "ind", "vcov"], vc[vc$grp == "Residual", "vcov"])
if(interactive()){
   confint(res, method="profile", oldNames=FALSE) # faster
   confint(res, method="boot", oldNames=FALSE) # better
}
}
\author{
Timothee Flutre (inspired by Ben Bolker at http://stackoverflow.com/q/19327088/597069)
}

