% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{loadMummer}
\alias{loadMummer}
\title{MUMmer}
\usage{
loadMummer(file.coords, algo = "nucmer", asGRanges = TRUE,
  keep.nlong.qry = NULL, ref.lim = NULL, verbose = 1)
}
\arguments{
\item{file.coords}{path to the file with alignment coordinates obtained via \code{show-coords} (see the example)}

\item{algo}{nucmer or promer (the latter isn't available yet)}

\item{asGRanges}{if TRUE, returns a \code{\link[GenomicRanges]{GRanges}} object}

\item{keep.nlong.qry}{keep the N queries having the longest cumulated alignments on the reference (keep all of them by default)}

\item{ref.lim}{limits on reference coordinates outside which alignments are discarded (keep all of them by default)}

\item{verbose}{verbosity level (0/1)}
}
\value{
GRanges or data.frame
}
\description{
Load alignment coordinates from \href{http://mummer.sourceforge.net/}{MUMmer}.
}
\examples{
\dontrun{## MUMmer should be run beforehand, for instance:
## nucmer --maxmatch -p out-nucmer reference.fa queries.fa
## delta-filter -l 1000 -q out-nucmer.delta > out-nucmer_filter.delta
## show-coords -c -l -L 1000 -r -T out-nucmer_filter.delta | gzip > out-nucmer_filter_coords.txt.gz
coords <- loadMummer("out-nucmer_filter_coords.txt.gz")
library(GenomicRanges)
loc <- GRanges(seqnames=Rle("chr2"),
               ranges=IRanges(start=5, end=17))
idx <- subjectHits(findOverlaps(query=sl, subject=coords.gr))
plotAligns(coords=as.data.frame(ranges(coords.gr[idx,])))
}
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{mummer2granges}}
}

