% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{loadReadCountsPerIndAndLane}
\alias{loadReadCountsPerIndAndLane}
\title{Load read counts}
\usage{
loadReadCountsPerIndAndLane(lanes.dir)
}
\arguments{
\item{lanes.dir}{vector of paths to directories "lane_..." from "demultiplex.py --step 2"}
}
\value{
data.frame with (individual,lane) in rows and counts in columns
}
\description{
Load read counts per individual and lane
}
\author{
Timothee Flutre
}

