% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{log10WeightedSum}
\alias{log10WeightedSum}
\title{Log of weighted sum}
\usage{
log10WeightedSum(x, weights = NULL)
}
\arguments{
\item{x}{vector}

\item{weights}{weights}
}
\value{
numeric
}
\description{
Stable computation of \eqn{log_{10}(\sum_i w_i 10^x_i)}
}
\details{
Use equal weights if not specified.
}
\author{
Timothee Flutre
}
