% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(error)
}
\arguments{
\item{error}{vector \eqn{\hat{\theta}_i - \theta_i}}
}
\value{
numeric
}
\description{
Returns the mean absolute error (MAE) as defined in \href{https://doi.org/10.3354/cr030079}{Willmott and Matsuura (2005)}.
It can have some advantages over the RMSE (\href{https://doi.org/10.3354/cr030079}{Willmott and Matsuura (2005)}).
}
\seealso{
\code{\link{rmse}}
}
\author{
Timothee Flutre
}
