% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(error)
}
\arguments{
\item{error}{vector \eqn{\hat{\theta}_i - \theta_i}}
}
\value{
numeric
}
\description{
Return the Mean Absolute Error
}
\author{
Timothée Flutre
}

