% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeCross}
\alias{makeCross}
\title{Cross}
\usage{
makeCross(
  haplos.par1,
  loc.crossovers.par1,
  haplos.par2 = NULL,
  loc.crossovers.par2 = NULL,
  child.name = NULL,
  howto.start.haplo = 0,
  verbose = 1
)
}
\arguments{
\item{haplos.par1}{list of matrices (one per chromosome) for the first parent}

\item{loc.crossovers.par1}{list of vectors (one per chromosome) for the first parent}

\item{haplos.par2}{list of matrices (one per chromosome) for the second parent}

\item{loc.crossovers.par2}{list of vectors (one per chromosome) for the second parent}

\item{child.name}{identifier of the child (if NULL, <parent1>"-x-"<parent2> or <parent1>"-hd")}

\item{howto.start.haplo}{if 0, haplotypes with which to start the gametes will be chosen at random; but one can also specify 1 (always the first haplotype) or 2 (always the second haplotype)}

\item{verbose}{verbosity level (0/1)}
}
\value{
list of matrices (one per chromosome) for the child
}
\description{
Make a cross. If two different genotypes are given, then a fecundation is made with one gamete from each genotype. If the same genotype is given twice, an autofecondation is made with two different gametes from this genotype. If a single genotype is given, a haplodiploidization is made with a single gamete from this genotype.
}
\seealso{
\code{\link{makeCrosses}}, \code{\link{drawLocCrossovers}}, \code{\link{makeGameteSingleInd}}, \code{\link{fecundation}}
}
\author{
Timothee Flutre
}
