% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_contest.R
\name{makeExamplePlantFile}
\alias{makeExamplePlantFile}
\title{Example for breeding contest}
\usage{
makeExamplePlantFile(out.dir)
}
\arguments{
\item{out.dir}{path to the directory in which the file will be saved}
}
\value{
invisible data.frame
}
\description{
Make a file with examples of plant material to request (autofecundation, allofecundation, haplodiploidization)
}
\author{
Timothee Flutre
}

