% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{makeExamplePlantFile}
\alias{makeExamplePlantFile}
\title{Example for breeding game}
\usage{
makeExamplePlantFile(out.dir, lang = "fr")
}
\arguments{
\item{out.dir}{path to the directory in which the file will be saved}

\item{lang}{language to be used (fr/en)}
}
\value{
invisible data.frame
}
\description{
Make a file with examples of plant material to request (autofecundation, allofecundation, haplodiploidization)
}
\seealso{
\code{\link{readCheckBreedPlantFile}}
}
\author{
Timothee Flutre
}
