% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeGameteSingleIndSingleChrom}
\alias{makeGameteSingleIndSingleChrom}
\title{Gamete}
\usage{
makeGameteSingleIndSingleChrom(haplos.par.chr, loc.crossovers, start.haplo = 1)
}
\arguments{
\item{haplos.par.chr}{matrix containing both haplotypes of a parent for a given chromosome (must have dimnames, such as "ind37_h1" in rows and "snp00265" in columns)}

\item{loc.crossovers}{positions of the crossing overs in terms of SNP indices and not physical coordinates (the coordinate of the first nucleotide is assumed to be 1, and crossing-overs are assumed to occur right after the given positions)}

\item{start.haplo}{identifier of the haplotype with which to start the gamete (1/2)}
}
\value{
vector
}
\description{
Make a gamete for a given chromosome by recombining two parental haplotypes.
}
\seealso{
\code{\link{makeGameteSingleInd}}, \code{\link{drawLocCrossovers}}
}
\author{
Timothee Flutre
}
