% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{makeGameteSingleIndSingleChrom}
\alias{makeGameteSingleIndSingleChrom}
\title{Gamete}
\usage{
makeGameteSingleIndSingleChrom(haplos.par.chr, loc.crossovers)
}
\arguments{
\item{haplos.par.chr}{matrix containing both haplotypes of a parent for a given chromosome (must have dimnames, such as "ind37_h1" in rows and "snp00265" in columns)}

\item{loc.crossovers}{positions of the crossing overs (the coordinate of the first nucleotide is assumed to be 1, and crossing-overs are assumed to occur right after the given localization)}
}
\value{
vector
}
\description{
Make a gamete for a given chromosome by recombining two parental haplotypes (always starting with the first haplotype).
}
\author{
Timothee Flutre
}

