% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{MIMQTL}
\alias{MIMQTL}
\title{QTL detection by MIM}
\usage{
MIMQTL(cross, response.in.cross = TRUE, response = NULL,
  numeric.chr.format = FALSE, geno.joinmap = NULL, phase = NULL,
  pheno.col = "y", nperm = 100, threshold.alpha = 0.05,
  QTL_position = NULL, method = "hk", plot = c(FALSE, FALSE),
  range_nb_qtl_max = seq(1:5), nrun = 10, additive.only = TRUE,
  p2d = "", scan2file = "", type.CI = "LOD-1", nb.cores = 1,
  verbose = 0)
}
\arguments{
\item{cross}{object}

\item{response.in.cross}{logical}

\item{response}{character}

\item{numeric.chr.format}{logical}

\item{geno.joinmap}{genotypes at markers in the JoinMap format}

\item{phase}{phase}

\item{pheno.col}{character}

\item{nperm}{number of permutations to be done in \code{qtl::scantwo}}

\item{threshold.alpha}{vector of length 1 or 2 (optional) with thresholds for (1) the significance of QTL presence (based on permutations) and (2) the significance of QTL effects}

\item{QTL_position}{useful for simulated data}

\item{method}{method to detect QTL in \code{qtl::scantwo}}

\item{plot}{logical}

\item{range_nb_qtl_max}{vector}

\item{nrun}{integer}

\item{additive.only}{logical}

\item{p2d}{character}

\item{scan2file}{optional path to file}

\item{type.CI}{type of confidence interval}

\item{nb.cores}{number of cores}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
list
}
\description{
Perform QTL detection by multiple interval mapping
}
\author{
Agnes Doligez [aut], Charlotte Brault [ctbt], Timothee Flutre [ctb]
}
