% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mpInv}
\alias{mpInv}
\title{Pseudo-inverse}
\usage{
mpInv(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix
}
\description{
Return the Moore-Penrose pseudo-inverse of a matrix
}
\details{
Golub & Van Loan, Matrix Computations, 3rd edition, ch5, p257
}
\author{
Timothee Flutre
}

