% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{openCarthagene}
\alias{openCarthagene}
\title{Open CarthaGene}
\usage{
openCarthagene(file = NULL)
}
\arguments{
\item{file}{path to the file containing the dataset (will be loaded if not NULL)}
}
\value{
list
}
\description{
Open FIFOs to interact with CarthaGene.
}
\author{
Matthieu Falque [aut], Timothee Flutre [ctb]
}
