% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{parseCgGroup}
\alias{parseCgGroup}
\title{Parse CarthaGene's output}
\usage{
parseCgGroup(out.group, mrk.info = NULL, mrk2chr = NULL, verbose = 1)
}
\arguments{
\item{out.group}{vector of character output from \code{\link{runCarthagene}} corresponding to "group"}

\item{mrk.info}{data frame output from \code{\link{parseCgMrkinfo}} (optional; useful to have the marker names instead of only the marker identifiers)}

\item{mrk2chr}{named vector which values are markers and names are chromosomes; before usage, marker names ending in "_m" (an indication of parental coding) will be edited to remove such a suffix}

\item{verbose}{verbosity level (0/1)}
}
\value{
data frame with at least two columns, "linkage.group" and "id", another column, "locus", if mrk.info is provided, and another column, "chr", if mrk2chr is provided
}
\description{
Parse the output of CarthaGene's command "group".
}
\author{
Timothee Flutre
}
