% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pca}
\alias{pca}
\title{Principal component analysis}
\usage{
pca(X, ct = TRUE, sc = FALSE, plot = NULL, main = "PCA",
  cols = rep("black", nrow(X)))
}
\arguments{
\item{X}{data matrix with N rows and P columns}

\item{ct}{use TRUE to center the columns of X (recommended), FALSE otherwise}

\item{sc}{use TRUE to scale the columns of X (if different units), FALSE otherwise}

\item{plot}{if not NULL, use "points" to show a plot with \code{\link[graphics]{points}} of PC1 versus PC2, and "text" to use \code{\link[graphics]{text}} with row names of X as labels}

\item{main}{main title of the plot}

\item{cols}{N-vector of colors}
}
\value{
list with the rotated matrix (= X V) which columns corresponds to "principal components", and with the proportion of variance explained per PC
}
\description{
Via the singular value decomposition (SVD): X = U D V^T.
This is mostly for teaching purposes, see \code{\link[stats]{prcomp}} otherwise.
}
\author{
Timothee Flutre
}

