% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{plotGRanges}
\alias{plotGRanges}
\title{Plot GRanges}
\usage{
plotGRanges(gr, main = "Alignments", xlab = "reference", xlim = NULL,
  col = "black", shape = "segments")
}
\arguments{
\item{gr}{GRanges}

\item{main}{main title}

\item{xlab}{label of the x-axis}

\item{xlim}{x-axis limits}

\item{col}{segment color(s)}

\item{shape}{shape used to represent the alignments (segments/arrows)}
}
\value{
nothing
}
\description{
Plot GRanges of several queries on the same reference.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{plotAligns}}
}

