% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotGridMissGenos}
\alias{plotGridMissGenos}
\title{Plot missing SNP genotypes as a grid.}
\usage{
plotGridMissGenos(x, main = "Missing genotypes", xlab = "Individuals",
  ylab = "SNPs")
}
\arguments{
\item{x}{matrix with SNP genotypes as allele doses (NA if missing) with
SNPs in columns and individuals in rows}

\item{main}{an overall title for the plot (default="Missing genotypes")}

\item{xlab}{a title for the x axis (default="Individuals")}

\item{ylab}{a title for the y axis (default="SNPs")}
}
\value{
nothing
}
\description{
Data will be represented in black if missing, white otherwise.
}
\author{
Timothée Flutre
}

