% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotGridMissGenos}
\alias{plotGridMissGenos}
\title{Missing genotypes}
\usage{
plotGridMissGenos(
  X,
  main = "Missing marker genotypes",
  xlab = "Genotypes",
  ylab = "markers"
)
}
\arguments{
\item{X}{matrix of marker genotypes, with genotypes in rows and markers in columns; missing values should be encoded as NA}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}
}
\value{
nothing
}
\description{
Plot missing marker genotypes as a grid, represented in black if missing, white otherwise.
}
\author{
Timothee Flutre
}
