% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotHistMinAllelFreq}
\alias{plotHistMinAllelFreq}
\title{Histogram of minor allele frequencies}
\usage{
plotHistMinAllelFreq(X = NULL, maf = NULL, main = "", xlim = c(0, 0.5),
  col = "grey", border = "white", las = 1, breaks = "FD", verbose = 1,
  ...)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded as allele doses, with SNPs in
columns and individuals in rows (optional if maf is not null)}

\item{maf}{vector of minor allele frequencies (optional if X is not null)}

\item{main}{string for the main title (default="")}

\item{xlim}{default=c(0,0.5)}

\item{col}{color for the bars}

\item{border}{color for the border of the bars}

\item{las}{see ?par (default=1)}

\item{breaks}{see ?hist (default="FD")}

\item{verbose}{verbosity level (0/default=1)}

\item{...}{arguments to be passed to hist()}
}
\value{
nothing
}
\description{
Plot the histogram of the minor allele frequency per SNP
}
\details{
Missing values (encoded as NA) are discarded.
}
\author{
Timothee Flutre
}

