% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotHistMinAllelFreq}
\alias{plotHistMinAllelFreq}
\title{Minor allele frequencies}
\usage{
plotHistMinAllelFreq(X = NULL, maf = NULL, main = NULL, xlim = c(0,
  0.5), col = "grey", border = "white", las = 1, breaks = "FD",
  add.ml.beta = FALSE, verbose = 1, ...)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns (optional if maf is not null); if X is not NULL, minor allele frequencies will be estimated via \code{\link{estimMaf}}}

\item{maf}{vector of minor allele frequencies (optional if X is not null)}

\item{main}{string for the main title}

\item{xlim}{limits of the x-axis}

\item{col}{color for the bars}

\item{border}{color for the border of the bars}

\item{las}{see ?par}

\item{breaks}{see ?hist}

\item{add.ml.beta}{if TRUE, add a curve corresponding to a Beta distribution with parameters inferred by maximum likelihood}

\item{verbose}{verbosity level (0/1)}

\item{...}{arguments to be passed to hist()}
}
\value{
nothing
}
\description{
Plot the histogram of the minor allele frequency per SNP.
}
\author{
Timothee Flutre
}

