% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotLd}
\alias{plotLd}
\title{Pairwise linkage disequilibrium}
\usage{
plotLd(x, y, estim = "r2", main, use.density = TRUE,
  xlab = "Physical distance (bp)", ylab = paste0("Linkage disequilibrium (",
  estim, ")"), span = 1/10, degree = 1, evaluation = 50,
  sample.size = NULL, add.ohta.kimura = TRUE, add.sved = TRUE,
  Ne = NULL, c = NULL)
}
\arguments{
\item{x}{vector of distances between SNPs (see \code{\link{distSnpPairs}})}

\item{y}{vector of LD estimates (see \code{\link{estimLd}})}

\item{estim}{estimator of pairwise LD corresponding to the values in y (r2/r)}

\item{main}{main title}

\item{use.density}{if TRUE, uses smoothScatter; otherwise, use scatter.smooth}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{span}{the parameter alpha which controls the degree of smoothing (see \code{\link[stats]{loess}})}

\item{degree}{the degree of the polynomials to be used (see \code{\link[stats]{loess}})}

\item{evaluation}{number of points at which to evaluate the smooth curve (see \code{\link[stats]{loess.smooth}})}

\item{sample.size}{nb of sampled haplotypes, n, used to estimate the pairwise LD; if not NULL, n is used to plot the horizontal line at the r value above which the null hypothesis "D=0" is rejected, where r = D / sqrt(f_A f_a f_B f_b) and X^2 = n r^2 is the test statistic asymptotically following a Chi2(df=1) (see McVean, Handbook of Statistical Genetics, 2007, and Pritchard and Przewoski, AJHG, 2001)}

\item{add.ohta.kimura}{add the analytical approximation by Ohta and Kimura (1971); requires Ne and c}

\item{add.sved}{add the analytical approximation by Sved (1971); requires Ne and c}

\item{Ne}{effective population size}

\item{c}{recomb rate in events per base per generation}
}
\value{
nothing
}
\description{
Plots the linkage disequilibrium between pairs of SNPs, as a blue density or black points, with a red loess.
Possibility to add two analytical approximations of E[r^2] at equilibrium (see McVean, Handbook of Stat Gen, 2007): 1 / (1 + 4 Ne c x) by Sved (1971) and (10 + 4 Ne c x) / (22 + 13 * 4 Ne c x + (4 Ne c x)^2) by Ohta and Kimura (1971).
}
\author{
Timothee Flutre
}

