% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPca}
\alias{plotPca}
\title{Principal component analysis}
\usage{
plotPca(rotation, prop.vars, plot = "points", main = "PCA",
  cols = rep("black", nrow(rotation)))
}
\arguments{
\item{rotation}{rotated matrix which columns corresponds to "principal components"}

\item{prop.vars}{vector with the proportion of variance explained per PC}

\item{plot}{use "points" to show a plot with \code{\link[graphics]{points}} of PC1 versus PC2, and "text" to use \code{\link[graphics]{text}} with row names of \code{rotation} as labels}

\item{main}{main title of the plot}

\item{cols}{N-vector of colors}
}
\value{
nothing
}
\description{
Plot the two first principal components from a PCA.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{pca}}
}

