% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{plotPedigree}
\alias{plotPedigree}
\title{Plot pedigree}
\usage{
plotPedigree(inds, mothers, fathers, sexes, generations,
  edge.col.mother = "black", edge.col.father = "darkgrey",
  vertex.cols = "white", vertex.label.family = "Helvetica",
  mult.edge.curve = 0.25, ...)
}
\arguments{
\item{inds}{identifiers of the individuals}

\item{mothers}{identifiers of their mother; can be NA}

\item{fathers}{identifiers of their father; can be NA}

\item{sexes}{"F" for female (circle), "M" for male (square) and "H" for hermaphrodite (triangle); can also be NA (no shape)}

\item{generations}{should start at 0}

\item{edge.col.mother}{default="black"}

\item{edge.col.father}{default="darkgrey"}

\item{vertex.cols}{default="white"}

\item{vertex.label.family}{default="Helvetica"}

\item{mult.edge.curve}{default=0.25}

\item{...}{other plotting options; see plot.igraph()}
}
\value{
invisible pedigree as an "igraph" object
}
\description{
Plot a pedigree using the "igraph" package. Inspired by plot.pedigree() from the "synbreed" package (under GPL-3). Add options for monoecious species.
}
\author{
Timothee Flutre
}

