% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prettyPrintSummary}
\alias{prettyPrintSummary}
\title{Summary}
\usage{
prettyPrintSummary(x, spec = "\%.2f")
}
\arguments{
\item{x}{vector of numbers}

\item{spec}{specifier, see \code{\link{sprintf}}}
}
\value{
invisible summary
}
\description{
Print the output of \code{\link{betterSummary}} in a single line, as well as the total number of observations and the number of missing data.
}
\author{
Timothee Flutre
}
