% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{pve2beta}
\alias{pve2beta}
\title{Proportion of variance explained}
\usage{
pve2beta(pve = 0.7, maf = 0.3, sigma = 1)
}
\arguments{
\item{pve}{proportion of variance explained}

\item{maf}{minor allele frequency}

\item{sigma}{error standard deviation}
}
\value{
numeric
}
\description{
Computes the additive effect of a bi-allelic SNP (beta) given its PVE, MAF and error standard deviation for the simple linear regression model: for all i in {1,...,n}, y_i = mu + beta * x_i + epsilon_i with epsilon_i ~ N(0, sigma^2).
Indeed, for this model: var(y) = beta^2 var(x) + sigma^2.
Assuming Hardy-Weinberg equilibrium: x ~ Binomial(2, maf); hence: var(x) = 2 f (1 - f).
Moreover: PVE = var(beta x) / var(y).
As a consequence, by fixing the PVE, the MAF and sigma, we can deduce a value for beta.
}
\examples{
\dontrun{## compare two different PVEs
pve2beta(pve=0.7, maf=0.3, sigma=1)
pve2beta(pve=0.2, maf=0.3, sigma=1)

## compare two different MAFs, depending on the PVE
pve2beta(pve=0.7, maf=0.4, sigma=1)
pve2beta(pve=0.7, maf=0.1, sigma=1)
pve2beta(pve=0.2, maf=0.4, sigma=1)
pve2beta(pve=0.2, maf=0.1, sigma=1)
}
}
\author{
Timothee Flutre
}
