% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_contest.R
\name{readCheckBreedPlantFile}
\alias{readCheckBreedPlantFile}
\title{Read for breeding contest}
\usage{
readCheckBreedPlantFile(f = NULL, df = NULL, max.nb.hd = 1000)
}
\arguments{
\item{f}{path to the input file}

\item{df}{data.frame (if the file was already read)}

\item{max.nb.hd}{maximum number of haplodiploidization to request in a single file}
}
\value{
invisible data.frame
}
\description{
Read and check a file supposed to contain requests about plant material.
}
\author{
Timothee Flutre
}

