% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{readGenosFastPhase}
\alias{readGenosFastPhase}
\title{Imputation}
\usage{
readGenosFastPhase(file, snp.ids)
}
\arguments{
\item{file}{character}

\item{snp.ids}{vector of SNP identifiers}
}
\value{
matrix with genotypes in rows (2 per individual) and SNPs in columns
}
\description{
Extract the haplotypes from the output file of fastPHASE.
}
\author{
Timothee Flutre
}

