% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{rearrangeInputsForAssoGenet}
\alias{rearrangeInputsForAssoGenet}
\title{Plant association genetics}
\usage{
rearrangeInputsForAssoGenet(
  ids,
  y,
  X,
  snp.coords,
  alleles,
  rename.chr.prefix = NULL,
  verbose = 1
)
}
\arguments{
\item{ids}{data.frame of identifiers, with (at least) column names \code{cultivar.code} and \code{accession.code}; the outputs will be sorted according to this option}

\item{y}{vector, matrix or data.frame which row names should be present in \code{ids$cultivar.code} or \code{ids$accession.code}}

\item{X}{matrix of bi-allelic SNP genotypes encoded in allele doses in [0,2], with genotypes in rows and SNPs in columns; the "second" allele is arbitrary, it corresponds to the second column of \code{alleles}, which can be the minor or the major allele; row names should be present in \code{ids$accession.code} or \code{ids$cultivar.code}, and column names in \code{rownames(snp.coords)}}

\item{snp.coords}{data.frame with 2 columns \code{coord} and \code{chr}, and SNP identifiers as row names}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (exactly 2 columns are required); the second column should correspond to the allele which number of copies is counted at each SNP in \code{X}}

\item{rename.chr.prefix}{if not NULL, the value of this parameter will be passed to \code{\link{chromNames2integers}} so that chromosome names from snp.coords will be renamed as integers}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with inputs after subsetting and sorting
}
\description{
Subset and sort inputs necessary to perform an analysis of plant association genetics, that is, the subset of cultivars with genotypes and phenotypes, and the subset of markers having genotypes and coordinates.
}
\seealso{
\code{\link{chromNames2integers}}
}
\author{
Timothee Flutre
}
