% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{recodeGenosMinorSnpAllele}
\alias{recodeGenosMinorSnpAllele}
\title{Minor allele frequencies}
\usage{
recodeGenosMinorSnpAllele(X, alleles, verbose = 1)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in [0,2], with individuals in rows and SNPs in columns; missing values should be encoded as NA; genotypes can be imputed}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (named "minor" and "major")}

\item{verbose}{verbosity level (0/1)}
}
\value{
list with a matrix of SNP genotypes encoded in number of copies of the minor allele with individuals in rows and SNPs in columns, and a data.frame of alleles which columns are named "minor" and "major"
}
\description{
Recode SNP genotypes in terms of the number of copies of the minor allele.
Let be a bi-allelic SNP with two alleles, A, the first allele, and B, the second allele, "first" and "second" being arbitrary.
The SNP genotypes are originally coded in number of copies of the 2nd allele: if a genotype is "AA", then its allele dose is \code{0}; if "AB", then \code{1}; and if "BB", then \code{2}.
Let us now assume that "A" is the minor (least frequent) allele, then, when recoding SNP genotypes in terms of minor allele, "AA" will correspond to \code{2}, "AB" to \code{1} and "BB" to \code{0}.
}
\author{
Timothee Flutre
}

