% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{regplot}
\alias{regplot}
\title{Scatter plot with regression lines}
\usage{
regplot(x, y, reg = "lm", ...)
}
\arguments{
\item{x}{vector (missing data coded as NA will be automatically discarded)}

\item{y}{vector (missing data coded as NA will be automatically discarded)}

\item{reg}{specifies which model to use to add regression lines(s) (lm/loess/rlm)}

\item{...}{arguments to be passed to \code{\link[graphics]{plot}}}
}
\value{
object returned by the function specified via \code{reg}
}
\description{
Make a scatter plot of y as a function of x, along with regression line(s).
}
\examples{
set.seed(1859)
n <- 100
x <- rnorm(n=n, mean=37, sd=3)
y <- 50 + 1.2 * x + rnorm(n=n, mean=0, sd=3)
fit <- regplot(x=x, y=y, las=1, main="Linear regression")
text(x=40, y=84, labels=paste0("R2 = ", format(summary(fit)$r.squared, digits=2)))
}
\author{
Timothee Flutre
}

