% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{regplot}
\alias{regplot}
\title{Scatter plot with regression lines}
\usage{
regplot(
  x,
  y,
  reg = "lm",
  col = c(lm = "red"),
  show.crit = c("corP", "corS"),
  ci.int = TRUE,
  pred.int = TRUE,
  legend.x = "right",
  legend.y = NULL,
  las = 1,
  ...
)
}
\arguments{
\item{x}{vector or 1-column matrix (missing data coded as NA will be automatically discarded)}

\item{y}{vector or 1-column matrix (missing data coded as NA will be automatically discarded)}

\item{reg}{specifies which model(s) to use to add regression lines(s) (lm/loess/rlm; can be \code{c("lm", "rlm")})}

\item{col}{named vector specifying the color(s) of the regression line(s) specified via \code{reg}}

\item{show.crit}{specifies which criterion to show in the top left corner (corP/corS/R2adj); can be \code{c("R2adj", "corP")}) or \code{NULL}}

\item{ci.int}{if \code{reg="lm"}, add lines corresponding to confidence intervals}

\item{pred.int}{if \code{reg="lm"}, add lines corresponding to prediction intervals}

\item{legend.x}{the x coordinate to be used to position the legend (see \code{\link[graphics]{legend}})}

\item{legend.y}{the y coordinate to be used to position the legend (see \code{\link[graphics]{legend}})}

\item{las}{see \code{\link[graphics]{par}}}

\item{...}{arguments to be passed to \code{\link[graphics]{plot}}}
}
\value{
list of object(s) returned by the function(s) specified via \code{reg}
}
\description{
Make a scatter plot of y as a function of x, along with regression line(s).
}
\examples{
set.seed(1859)
n <- 500
x <- rnorm(n=n, mean=37, sd=3)
y <- 50 + 1.2 * x + rnorm(n=n, mean=0, sd=3)
fit <- regplot(x=x, y=y, reg="lm", las=1, main="Linear regression")
fit <- regplot(x=x, y=y, reg="loess", las=1, col=c(loess="red"),
               main="Locally weighted scatterplot smoothing (loess)")
y2 <- y + sample(x=c(rep(0, 0.7*floor(n)),
                     rnorm(n=ceiling(0.3*n), mean=c(7,13), sd=20)), size=n)
fit <- regplot(x=x, y=y2, reg=c("lm","rlm"), las=1,
               col=c(lm="red", rlm="blue"), legend.x="bottomright",
               main="(Robust) linear regressions")
}
\author{
Timothee Flutre
}
