% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{regplot}
\alias{regplot}
\title{Make a scatter plot of y as a function of x, along with the regression
line from lm() as well as both confidence and both prediction lines from
predict().}
\usage{
regplot(x, y, ...)
}
\arguments{
\item{x}{vector of points}

\item{y}{vector of points}

\item{...}{arguments to be passed to plot()}
}
\value{
nothing
}
\description{
Make a scatter plot of y as a function of x, along with the regression
line from lm() as well as both confidence and both prediction lines from
predict().
}
\author{
Timothee Flutre
}

