% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeFileExtension}
\alias{removeFileExtension}
\title{File}
\usage{
removeFileExtension(file, fileext)
}
\arguments{
\item{file}{character}

\item{fileext}{character}
}
\value{
character
}
\description{
Remove the last extension in a file name.
}
\examples{
f <- "data.txt.gz"
removeFileExtension(file=f, fileext=".gz")
removeFileExtension(file=f, fileext=".txt.gz")
}
\author{
Timothee Flutre
}
