% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmatnorm}
\alias{rmatnorm}
\title{The matrix-variate Normal distribution}
\usage{
rmatnorm(n = 1, M, U, V)
}
\arguments{
\item{n}{number of observations}

\item{M}{mean matrix}

\item{U}{between-row covariance matrix}

\item{V}{between-column covariance matrix}
}
\value{
array
}
\description{
Random generation for the matrix-variate Normal distribution.
See \url{https://en.wikipedia.org/wiki/Matrix_normal_distribution}.
}
\author{
Timothee Flutre
}

