% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(error)
}
\arguments{
\item{error}{vector \eqn{\hat{\theta}_i - \theta_i}}
}
\value{
numeric
}
\description{
Returns the root mean squared error (RMSE) as defined in \href{https://doi.org/10.3354/cr030079}{Willmott and Matsuura (2005)}.
}
\seealso{
\code{\link{rmsre}}, \code{\link{rrmse}}, \code{\link{mae}}
}
\author{
Timothee Flutre
}
