% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(error)
}
\arguments{
\item{error}{vector \eqn{\hat{\theta}_i - \theta_i}}
}
\value{
numeric
}
\description{
Return the Root Mean Squared Error
}
\author{
Timothee Flutre
}

