% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{rngVcf2df}
\alias{rngVcf2df}
\title{Convert ranges to data.frame}
\usage{
rngVcf2df(vcf)
}
\arguments{
\item{vcf}{CollapsedVCF (see pkg VariantAnnotation)}
}
\value{
data.frame
}
\description{
Convert ranges to data.frame
}
\author{
Timothee Flutre
}

