% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{rootfreeChol}
\alias{rootfreeChol}
\title{LDL' decomposition}
\usage{
rootfreeChol(x)
}
\arguments{
\item{x}{matrix}
}
\value{
matrix
}
\description{
This wrapper around \code{\link{chol}} factors a square, symmetric, positive (semi-)definite matrix into the product of a lower triangular matrix (L), a diagonal matrix (D), and the transpose of the lower triangular matrix (U=L').
The value returned is a lower triangular matrix with the elements of D on the diagonal.
This decomposition is notably used in quantitative genetics (Misztal and Perez-Enciso, 1993).
The determinant of the input matrix is then easily computed as the product of the diagonal entries of the decomposition.
}
\examples{
\dontrun{## example matrix from Misztal and Perez-Enciso (1993):
W <- matrix(NA, nrow=5, ncol=5)
W[1,] <- c(5, 0, 0, 3, 0)
W[2,-1] <- c(3, 0, 0, 1)
W[3,-c(1:2)] <- c(4, 1, 1)
W[4,-c(1:3)] <- c(4, 0)
W[5,-c(1:4)] <- 2
W
W[lower.tri(W)] <- t(W)[lower.tri(W)]
W

## square root free Cholesky decomposition:
decomp <- rootfreeChol(W)

## determinant:
prod(diag(decomp)) # ~= 162, as in the paper
}
}
\seealso{
\code{\link{getSparseInv}}
}
