% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarthaGene.R
\name{runCarthagene}
\alias{runCarthagene}
\title{Run a CarthaGene command}
\usage{
runCarthagene(cg, cmd)
}
\arguments{
\item{cg}{list output from \code{\link{openCarthagene}}}

\item{cmd}{character specifying the command to run}
}
\value{
vector of lines
}
\description{
Send a command to CarthaGene and wait for the result.
}
\author{
Matthieu Falque [aut], Timothee Flutre [ctb]
}
