% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{segregJoinMap2Qtl}
\alias{segregJoinMap2Qtl}
\title{JoinMap/MapQTL to R/qtl}
\usage{
segregJoinMap2Qtl()
}
\value{
data.frame
}
\description{
Return the correspondence in terms of genotype coding between the "segregation" format used by \href{https://www.kyazma.nl/index.php/JoinMap/}{JoinMap} and the one used by \href{https://cran.r-project.org/package=qtl}{qtl}.
}
\author{
Timothee Flutre
}
