% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{segSites2allDoses}
\alias{segSites2allDoses}
\title{Convert the SFS of independent replicates into a matrix of allele dosage.}
\usage{
segSites2allDoses(seg.sites)
}
\arguments{
\item{seg.sites}{list returned by scrm()}
}
\value{
matrix with diploid individuals in rows and SNPs in columns
}
\description{
SFS stands for site frequency spectrum
}
\author{
Timothee Flutre
}

