% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{segSites2snpCoords}
\alias{segSites2snpCoords}
\title{Site frequency spectrum}
\usage{
segSites2snpCoords(seg.sites, snp.ids, chrom.len, prefix = "chr",
  verbose = 0)
}
\arguments{
\item{seg.sites}{list of haplotypes returned by \code{scrm}, each component of which corresponds to a matrix with haplotypes in rows and SNP in columns}

\item{snp.ids}{vector of identifiers (one per SNP)}

\item{chrom.len}{chromosome length (same for all)}

\item{prefix}{character string}

\item{verbose}{verbosity level (0/1)}
}
\value{
data.frame with SNPs in rows and 2 columns (chr, pos)
}
\description{
Make a data.frame of SNP coordinates from the SFS of independent replicates.
}
\author{
Timothee Flutre
}
