% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{setUpBreedingGame}
\alias{setUpBreedingGame}
\title{Set up}
\usage{
setUpBreedingGame(root.dir, shared.dir = NULL, nb.breeders = 3,
  lang = "fr")
}
\arguments{
\item{root.dir}{path to the root directory}

\item{shared.dir}{path to the shared directory (e.g. via Dropbox; root.dir if NULL)}

\item{nb.breeders}{number of breeders (in EN, will be named "breeder<i>"; in FR, will be named "selectionneur<i>"; a "test" will also be created)}

\item{lang}{language to be used (default=fr/en)}
}
\value{
list
}
\description{
Set up the directories and names for the breeding game.
Already-existing directories are not re-created.
}
\author{
Timothee Flutre
}

