% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl.R
\name{SIMQTL}
\alias{SIMQTL}
\title{QTL detection by SIM}
\usage{
SIMQTL(cross, numeric.chr.format = TRUE, response.in.cross = TRUE,
  pheno.col = "y", response = NULL, method = "em", geno.joinmap = NULL,
  phase, threshold = NULL, nperm = 100, alpha = c(0.05, NA),
  plot = FALSE, QTL_position = NULL, verbose = 0)
}
\arguments{
\item{cross}{object}

\item{numeric.chr.format}{logical to indicate if chromosome names are numeric}

\item{response.in.cross}{logical to indicate if response studied is in \code{cross$pheno}, default is TRUE}

\item{pheno.col}{character indicating column to study in \code{cross$pheno}}

\item{response}{named numeric or vector for response if not in \code{cross$pheno}, default is NULL}

\item{method}{method to detect QTL in \code{qtl::scanone}, default is "em"}

\item{geno.joinmap}{genotypes at markers in the JoinMap format, if NULL (default), no estimation of allelic effects is given}

\item{phase}{marker phases}

\item{threshold}{genomewide significance LOD threshold, if NULL (default), is found by permutations (with nperm parameter).}

\item{nperm}{number of permutations to be done in \code{qtl::scanone}, default is 100}

\item{alpha}{vector of length 1 or 2 (optional) with thresholds for (1) the significance of QTL presence (based on permutations) and 
(2) the significance of linear regression on QTL effect, if NA, no threshold is applied and all allelic effects are kept.}

\item{plot}{logical, default is FALSE.}

\item{QTL_position}{matrix with genetic.distance and linkage.group as columns indicating QTL positions for plotting}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
list of 3 elements: qtl.df is a data frame with QTL informations (linkage.group, position, LOD, interval.inf and interval.sup) / 
selected markers is a character vector for markers inside confidence interval /allelic effetcs is a data frame with a column predictor and a column effect with estimated allelic effects.
}
\description{
Perform QTL detection by simple interval mapping
}
\seealso{
\code{\link{MIMQTL}}
}
\author{
Charlotte Brault [aut], Timothee Flutre [ctb]
}
