% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulAnimalModel}
\alias{simulAnimalModel}
\title{Simulate phenotypes from a basic "animal model" (LMM).}
\usage{
simulAnimalModel(Q = 3, mu = 50, mean.a = 5, sd.a = 2, A, lambda = 3,
  sigma.u2 = NULL, scale.halfCauchy = NULL, perc.NA = 0, err.df = Inf,
  seed = NULL)
}
\arguments{
\item{Q}{number of years}

\item{mu}{overall mean}

\item{mean.a}{mean of the prior on alpha[2:Q]}

\item{sd.a}{std dev of the prior on alpha[2:Q]}

\item{A}{matrix of additive genetic relationships}

\item{lambda}{ratio of variances (sigma_u^2 /sigma^2)}

\item{sigma.u2}{genetic variance (e.g. 15)}

\item{scale.halfCauchy}{scale of the half-Cauchy prior for sigma_u (e.g. 5)}

\item{perc.NA}{percentage of missing phenotypes, at random}

\item{err.df}{degrees of freedom of errors' Student's t-distribution}

\item{seed}{seed for the pseudo-random number generator}
}
\value{
list
}
\description{
y = W alpha + Z u + epsilon
where y is N x 1; W is N x Q; Z is N x I
u ~ Norm_I(0, G=sigma_u^2 A); epsilon_n ~ Student(nu, 0); Cov(u,e)=0
}
\note{
If A is not positive definite, an error will be raised; in such cases, using the nearPD function from the Matrix package can be useful.
}
\author{
Timothee Flutre
}

