% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulAnimalModel}
\alias{simulAnimalModel}
\title{Animal model}
\usage{
simulAnimalModel(T = 1, Q = 3, mu = rep(50, T), mean.C = 5, sd.C = 2,
  A, V.G.A = NULL, scale.hC.G.A = NULL, nu.G.A = T, D = NULL,
  V.G.D = NULL, scale.hC.G.D = NULL, nu.G.D = T, V.E = NULL,
  scale.hC.E = NULL, nu.E = T, err.df = Inf, perc.NA = 0, seed = NULL)
}
\arguments{
\item{T}{number of traits}

\item{Q}{number of covariates (as "fixed effects", e.g. replicates)}

\item{mu}{T-vector of overall means (one per trait), i.e. C[1,1:T]}

\item{mean.C}{mean of the univariate Normal prior on C[2:Q,1:T] (ignored if Q=1)}

\item{sd.C}{std dev of the univariate Normal prior on C[2:Q,1:T] (ignored if Q=1)}

\item{A}{IxI matrix of additive genetic relationships between genotypes (see \code{\link{estimGenRel}} with VanRaden's estimator)}

\item{V.G.A}{TxT matrix of additive genetic variance-covariance between traits (e.g. 15 when T=1)}

\item{scale.hC.G.A}{scale of the half-Cauchy prior for sqrt{V_{G_A}} (e.g. 5; used if V.G.A=NULL and T=1)}

\item{nu.G.A}{degrees of freedom of the Wishart prior for V_{G_A} (used if V.G.A=NULL and T>1)}

\item{D}{IxI matrix of dominant genetic relationships between genotypes (see \code{\link{estimGenRel}} with Vitezica's estimator)}

\item{V.G.D}{TxT matrix of dominant genetic variance-covariance between traits (e.g. 3 when T=1; used if D!=NULL)}

\item{scale.hC.G.D}{scale of the half-Cauchy prior for sqrt{V_{G_D}} (e.g. 5; used if D!=NULL, V.G.D=NULL and T=1)}

\item{nu.G.D}{degrees of freedom of the Wishart prior for V_{G_D} (used if D!=NULL, V.G.D=NULL and T>1)}

\item{V.E}{TxT matrix of error variance-covariance between traits (used if T=1 and err.df=Inf)}

\item{scale.hC.E}{scale of the half-Cauchy prior for sqrt{V_E} (e.g. 5; used if V.E=NULL and T=1 and err.df=Inf)}

\item{nu.E}{degrees of freedom of the Wishart prior for V_E (used if V.E=NULL and T>1)}

\item{err.df}{degrees of freedom of the Student's t-distribution of the errors (e.g. 3; Inf means Normal distribution; will be Inf if T>1)}

\item{perc.NA}{percentage of missing phenotypes, at random}

\item{seed}{seed for the pseudo-random number generator}
}
\value{
list
}
\description{
Given T traits, I genotypes, Q covariates and N=I*Q phenotypes per trait, simulate phenotypes via the following "animal model": \eqn{Y = W C + Z G_A + Z G_D + E}, where Y is N x T; W is N x Q; Z is N x I; G_A ~ Normal_IxT(0, A, V_{G_A}) with A being IxI and V_{G_A} being TxT; G_D ~ Normal_IxT(0, D, V_{G_D}) with D being IxI and V_{G_D} being TxT; E ~ Normal_NxT(0, Id_N, V_E) with Id_N being NxN and V_E being TxT. Note that using the matrix Normal (MN) is equivalent to using the multivariate Normal (MVN) via the vec operator and Kronecker product: \eqn{Y ~ MN(M, U, V) <=> vec(Y) ~ MVN(vec(M), V \otimes U)}.
}
\examples{
## simulate genotypes
set.seed(1859)
X <- simulGenosDose(nb.genos=100, nb.snps=2000)

## estimate the additive genetic relationships
A <- estimGenRel(X, relationships="additive", method="vanraden1", verbose=0)
A <- as.matrix(Matrix::nearPD(A)$mat) # not always necessary

# one trait with heritability h2=0.75, no covariate, Normal errors, no NA
model <- simulAnimalModel(T=1, Q=1, A=A, V.G.A=15, V.E=5)

# one trait with heritability h2=0.75, three covariates, Normal errors, no NA
model <- simulAnimalModel(T=1, Q=3, A=A, V.G.A=15, V.E=5)

# one trait with heritability h2=0.75, no covariate, Student errors, no NA
model <- simulAnimalModel(T=1, Q=1, A=A, V.G.A=15, err.df=3)

# one trait with heritability drawn at random, no covariate, Normal errors, no NA
model <- simulAnimalModel(T=1, Q=1, A=A, scale.hC.G.A=10, V.E=5)

# two traits with heritabilities drawn at random, no covariate, Normal errors, no NA
model <- simulAnimalModel(T=2, Q=1, A=A, nu.G.A=5, nu.E=3)
}
\author{
Timothee Flutre
}

