% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{simulAr1Ar1}
\alias{simulAr1Ar1}
\title{AR1xAR1 simulation}
\usage{
simulAr1Ar1(n = 1, R, C, sigma2, rho.r = 0, rho.c = 0)
}
\arguments{
\item{R}{number of rows}

\item{C}{number of columns}

\item{sigma2}{variance of the errors}

\item{rho.r}{correlation between rows}

\item{rho.c}{correlation between columns}

\item{N}{number of samples}
}
\value{
array
}
\description{
Simulate random samples from a separable AR1xAR1 process as in \href{https://dx.doi.org/10.1016/0378-3758(95)00066-6}{Martin (1996)}.
}
\author{
Timothee Flutre
}
