% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulBslmm}
\alias{simulBslmm}
\title{Simulate phenotypes according to the BSLMM model.}
\usage{
simulBslmm(Q = 3, mu = 50, mean.a = 5, sd.a = 2, X, pi = NULL,
  h = NULL, rho = NULL, tau = 1, perc.NA = 0, err.df = Inf,
  seed = NULL)
}
\arguments{
\item{Q}{number of years during which individuals are phenotyped (starting in 2010)}

\item{mu}{overall mean}

\item{mean.a}{mean of the prior on alpha[2:Q]}

\item{sd.a}{std dev of the prior on alpha[2:Q]}

\item{X}{matrix of SNP genotypes encoded as allele doses in {0,1,2}, with individuals in rows and SNPs in columns (SNPs with missing values or low MAF should be discarded beforehand)}

\item{pi}{proportion of beta-tilde values that are non-zero}

\item{h}{approximation to E[PVE] (h and rho should be NULL or not together)}

\item{rho}{approximation to E[PGE]}

\item{tau}{precision of the errors (i.e. 1 / sigma^2)}

\item{perc.NA}{percentage of missing phenotypes, at random}

\item{err.df}{degrees of freedom of errors' Student's t-distribution}

\item{seed}{seed for the pseudo-random number generator}
}
\value{
list
}
\description{
y = W alpha + Z X tilde{beta} + Z u + epsilon
where y is N x 1; W is N x Q; alpha is Q x 1; Z is N x I; X is I x P; u is I x 1 and epsilon is N x 1
for all p, tilde{beta}_p ~ pi Norm_1(0, sigma_betat^2/sigma^2) + (1 - pi) delta_0
u ~ Norm_I(0, (sigma_u^2/sigma^2) A)
epsilon ~ Norm_N(0, sigma^2 I)
See Zhou, Carbonetto & Stephens (2013).
}
\author{
Timothee Flutre
}

