% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{simulCoalescent}
\alias{simulCoalescent}
\title{Coalescent with recombination}
\usage{
simulCoalescent(nb.inds = 100, ind.ids = NULL, nb.reps = 20,
  pop.mut.rate = 50, pop.recomb.rate = 5, chrom.len = 10^3, nb.pops = 1,
  mig.rate = 5, get.trees = FALSE, get.tmrca = FALSE, verbose = 0)
}
\arguments{
\item{nb.inds}{diploids (thus nb of haplotypes is 2 * nb.inds)}

\item{ind.ids}{vector of identifiers (one per individual)}

\item{nb.reps}{number of independent loci that will be produced (could be seen as distinct chromosomes)}

\item{pop.mut.rate}{theta = 4 N0 mu}

\item{pop.recomb.rate}{rho = 4 N0 r}

\item{chrom.len}{in bp}

\item{nb.pops}{number of populations}

\item{mig.rate}{migration rate = 4 N0 m (symmetric)}

\item{get.trees}{get gene genealogies in the Newick format}

\item{get.tmrca}{get time to most recent common ancestor and local tree lengths}

\item{verbose}{verbosity level (default=0=nothing, 1=few, 2=more)}
}
\value{
list with haplotypes (list), genotypes as allele doses (matrix) and SNP coordinates (data.frame)
}
\description{
Simulate haplotypes according to an approximation to the coalescent with recombination named the Sequential Coalescent with Recombination Model. Requires the scrm package (Staab et al, 2014).
}
\author{
Timothee Flutre
}

