% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_game.R
\name{simulTraits12}
\alias{simulTraits12}
\title{Simul breeding game}
\usage{
simulTraits12(
  dat,
  mu = c(trait1 = 100, trait2 = 15),
  sigma.alpha2 = c(trait1 = 230, trait2 = 10),
  Alpha = NULL,
  X,
  afs,
  Beta,
  h2 = NULL,
  sigma2 = c(trait1 = 330, trait2 = 0.6),
  cor.E.inter.trait = 0,
  set.neg.to.zero = TRUE,
  verbose = 1
)
}
\arguments{
\item{dat}{data.frame specifying the structure of the initial data set of phenotypes, with columns "ind", "year", etc}

\item{mu}{vector of global means, for each trait}

\item{sigma.alpha2}{vector of variance for the "year" effects, for each trait (ignored if \code{Alpha} is not NULL)}

\item{Alpha}{matrix of "year" effects, for each trait, the years corresponding to those indicated in \code{dat} (if NULL, will be simulated using \code{sigma.alpha2})}

\item{X}{matrix of bi-allelic SNP genotypes encoded in allele dose in \{0,1,2\}, with individuals in rows in the same order as the levels of \code{dat$ind}}

\item{afs}{vector of allele frequencies (to center X so that the genotypic values are also centered) which names are the SNPs}

\item{Beta}{matrix of additive SNP effects, for each trait}

\item{h2}{vector of heritabilities, with the name of each trait, for instance \code{c(trait1=0.3, trait2=0.4)} (if NULL, \code{sigma2} will be used, but do not specify both)}

\item{sigma2}{vector of error variances, with the name of each trait, for instance \code{c(trait1=0.467, trait2=0.210)} (if NULL, \code{h2} will be used, but do not specify both)}

\item{cor.E.inter.trait}{correlation of errors between both traits}

\item{set.neg.to.zero}{if TRUE, the negative phenotypic values, if any, will be set to zero}

\item{verbose}{verbosity level (0/1)}
}
\value{
list
}
\description{
Simulate phenotypes of traits 1 and 2 jointly: Y = Z.J * Alpha + Z.I * G.A + E.
}
\examples{
\dontrun{set.seed(1859)
X <- simulGenosDose(nb.genos=825, nb.snps=2000)
snp.effects12 <- simulSnpEffectsTraits12(snp.ids=colnames(X))
dat <- makeDfInitPhenos(line.ids=rownames(X))
phenos <- simulTraits12(dat, X=X, Beta=snp.effects12$Beta)
regplot(phenos$G.A[,1], phenos$G.A[,2])
}
}
\seealso{
\code{\link{simulSnpEffectsTraits12}}, \code{\link{makeDfInitPhenos}}
}
\author{
Timothee Flutre
}
