% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantgen.R
\name{splitGenomesTrainTest}
\alias{splitGenomesTrainTest}
\title{Genomes}
\usage{
splitGenomesTrainTest(genomes, nb.inds.pred)
}
\arguments{
\item{genomes}{list, e.g. returned by \code{\link{simulCoalescent}}}

\item{nb.inds.pred}{number of individuals for whom prediction will be performed}
}
\value{
list composed of two lists, genomes and genomes.pred
}
\description{
From a set of individual genomes simulated together, split them into a training (for estimation) and testing (for prediction) sets.
}
\author{
Timothee Flutre
}

