% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{stanAMwriteModel}
\alias{stanAMwriteModel}
\title{Animal model}
\usage{
stanAMwriteModel(stan.file = NULL, errors.Student = FALSE,
  missing.phenos = FALSE)
}
\arguments{
\item{stan.file}{path to a file; if NULL, a temporary one will be created}

\item{errors.Student}{use a Student's t distribution for the errors (useful to handle outliers)}

\item{missing.phenos}{indicating if there is any missing phenotypes to impute jointly with the other unknown variables}
}
\value{
invisible path to stan.file
}
\description{
Animal model
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{stanAM}}
}

