% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_sequences.R
\name{vcf2dosage}
\alias{vcf2dosage}
\title{Convert VCF to dose}
\usage{
vcf2dosage(vcf.file, genome, gdose.file, ca.file, yieldSize = NA_integer_,
  dict.file = NULL, seq.id = NULL, seq.start = NULL, seq.end = NULL,
  uncertain = FALSE, verbose = 1)
}
\arguments{
\item{vcf.file}{path to the VCF file (bgzip index should exist in same directory; should only contain SNPs and be already filtered for QD, FS, MQ, etc)}

\item{genome}{genome identifier (e.g. "VITVI_12x2")}

\item{gdose.file}{path to the output file to record genotypes as allele doses (will be gzipped)}

\item{ca.file}{path to the output file to record SNP 1-based coordinates and alleles (will be gzipped)}

\item{yieldSize}{number of records to yield each time the file is read from (see ?TabixFile) if seq.id is NULL}

\item{dict.file}{path to the SAM dict file (see \url{https://broadinstitute.github.io/picard/command-line-overview.html#CreateSequenceDictionary}) if seq.id is specified with no start/end}

\item{seq.id}{see \code{\link{seqIdStartEnd2GRanges}}}

\item{seq.start}{see \code{\link{seqIdStartEnd2GRanges}}}

\item{seq.end}{see \code{\link{seqIdStartEnd2GRanges}}}

\item{uncertain}{logical indicating whether the genotypes to convert should come from the "GT" field (uncertain=FALSE) or the "GP" or "GL" field (uncertain=TRUE)}

\item{verbose}{verbosity level (0/1)}
}
\value{
an invisible list with both output file paths
}
\description{
Convert genotypes at bi-allelic variants from a VCF file into allele doses.
}
\author{
Timothee Flutre
}
\seealso{
\code{\link{gtVcf2dose}}
}

