% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen.R
\name{writeInputsForFastPhase}
\alias{writeInputsForFastPhase}
\title{Convert genotypes}
\usage{
writeInputsForFastPhase(X, snp.coords, alleles, file, verbose = 0)
}
\arguments{
\item{X}{matrix of SNP genotypes encoded in number of copies of the 2nd allele, i.e. as allele doses in {0,1,2}, with individuals in rows and SNPs in columns}

\item{snp.coords}{data.frame with SNP identifiers as row names, and two columns, "chr" and "coord" or "pos"}

\item{alleles}{data.frame with SNPs in rows (names as row names) and alleles in columns (named "minor" and "major")}

\item{file}{write the genotype data to this file (for instance 'genotypes_fastphase.txt' or 'genotypes_fastphase.txt.gz', but don't use \code{\link[base]{gzfile}})}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
nothing
}
\description{
Use the external software fcGENE from \href{dx.plos.org/10.1371/journal.pone.0097589}{Roshyara and Scholz (2014)}.
}
\author{
Timothee Flutre
}

